### 2023-11-21 #6076 @editorialbot
[PRE REVIEW]: proteusPy: A Python Package for Protein Structure and Disulfide Bond Modeling and Analysis
**Submitting author:** @suchanek (<a href="http://orcid.org/0009-0009-0891-1507">Eric Suchanek</a>)
**Repository:** https://github.com/suchanek/proteusPy
**Branch with paper.md** (empty if default branch):
**Version:** v0.81
