### 2023-11-24 #6076 @arfon
> @openjournals/dev our bot is picking up this paper file: https://github.com/suchanek/proteusPy/blob/master/documents/DisulfideBondAnalysis/SS_paper.md

Please rename this paper to something else. Essentially the regex is picking up anything with `paper.md` in the path. @xuanxu – we should probably make this regex more sensitive (i.e., only pick up files that are exactly `paper.md`)?
