### 2023-11-26 #6076 @xuanxu
The code for the bot has been updated, this should work now with the correct paper file
