### 2023-11-26 #6076 @xuanxu
The problem now is caused by the presence of another `paper.md` file, here:
https://github.com/suchanek/proteusPy/blob/master/documents/gaia/paper.md
