### 2023-12-20 #6076 @richardjgowers
@AnjaConev would you be interested in reviewing this for JOSS?
