### 2023-12-20 #6076 @richardjgowers
@JoaoRodrigues would you be interested in reviewing this for JOSS?
