### 2023-12-20 #6076 @JoaoRodrigues
Hi Richard,

Sure, happy to.

Cheers,

João

A quarta, 20/12/2023, 16:25, Richard Gowers ***@***.***>
escreveu:

> @JoaoRodrigues <https://github.com/JoaoRodrigues> would you be interested
> in reviewing this for JOSS?
>
