+++ _pyproject.toml
@@ -1,26 +1,0 @@
-[tool.poetry]
-name = "proteusPy"
-version = "0.8"
-description = "[]"
-authors = ["Eric Suchanek <suchanek@mac.com>"]
-license = "mit"
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "3.11"
-panel = "0.14.2"
-proteusPy = "0.8"
-pyvista = "0.42.3"
-biopython = "1.81"
-plotly = "5.18.0"
-plotly-express = "^0.4.1"
-tqdm = "4.66.1"
-jupyter = "^1.0.0"
-ipykernel = "^6.26.0"
-gdown = "^4.7.1"
-nodejs = "^0.1.1"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
