### 2023-12-24 suchanek/proteusPy#3 @AnjaConev
JOSS review - License file
Please make sure that your license file is named correctly.
You have a valid BSD2 license in the markdown file https://github.com/suchanek/proteusPy/blob/master/License.md

It seems that for the JOSS purposes it should the file should be https://github.com/suchanek/proteusPy/blob/master/LICENSE
You just have to rename the file.
Thanks! 😄
