### 2023-12-24 suchanek/proteusPy#4 @AnjaConev
JOSS review - Installation comments
Here are some of my comments on how you can improve the installation instructions:

## Things that failed
#### 1. Git-lfs fails

Description:
` git-lfs track "*.csv" "*.pkl" "*.mp4"` fails for me with a message: `Not in a Git repository.`

Solution:
   - I had to clone the repository first and cd to it
   - It would be good to change the instructions to
   ```bash
      $ git clone https://github.com/suchanek/proteusPy
      $ cd proteusPy
      $ git-lfs track "*.csv" "*.pkl" "*.mp4"
      ....
   ```
   - I am still not sure what `git-lfs track` will change and if it is necessary if we don't plan to commit large files, but it works this way

#### 2. Conda env create fails

Description:
`conda env create --name proteusPy --file=proteusPy.yml` fails with the `SpecNotFound: Invalid name 'proteusPy', try the format: user/package`

Solution:
   - the yml file on the master branch is currently named `ppy.yml` - either change the file name or the instruction
   - I was able to proceed by running conda command with `--file=ppy.yml`

#### 3. Jupyter command fails

Description:
`jupyter nbextension enable --py --sys-prefix widgetsnbextension` fails with `Permission denied`

Solution:
   - I got it to work by running `sudo jupyter ...`

## Recommendations to improve the experience:
- The instructions seem to be for linux/mac OS (it would be helpful to specify this)
- There are a couple of `yml` files in the root directory - are all those needed? If not, it would help to keep only the yml file needed for conda environment.
- Conda install is very slow at resolving the environment - recommend using mamba or micromamba https://mamba.readthedocs.io/en/latest/
- Maybe you can make your package available on pyPI or a conda channel for easier installation
- After the installation steps, it would help to have a test or an example case that users can run to verify that the installation was successful
