### 2023-12-24 suchanek/proteusPy#5 @AnjaConev
JOSS review - widgets not rendering 
When I try to render any of the widgets in the Jupyter Notebooks I get errors.

For example, when I run the second cell of `notebooks/Display_SS.ipynb`, I get:

![image](https://github.com/suchanek/proteusPy/assets/7367905/307c1a7b-3105-4854-95ca-bdab094022c3)
