### 2023-12-24 suchanek/proteusPy#4 @AnjaConev
P.S. For the yml file - it would be nice to include the versions of the dependencies (to avoid the possibility of your tool not being compatible with some future versions of the dependencies).
