### 2023-12-24 suchanek/proteusPy#6 @AnjaConev
JOSS review - tests
If possible, please provide some tests for different modules of your package!

Thanks! 🎄 😄
