### 2023-12-24 suchanek/proteusPy#5 @AnjaConev
This error might be related to the platform on which I am running this right now.
I was testing the tool on Windows 10 with the WSL2 (Linux subsystem).
I do not have access to a different platform right now.
I will try again later on a Linux desktop and let you know if I have any luck.
