### 2023-12-24 suchanek/proteusPy#7 @AnjaConev
JOSS review - software paper - current state of the field
Could you give a more extensive overview of the current state of the field?
You mention that your database is the first database of disulfide bonds to your knowledge.
I found a part of biotite package that can visualize disulfide bonds to some (limited) extent https://www.biotite-python.org/examples/gallery/structure/disulfide_bonds.html
Is there any other software that calculates/visualizes the disulfide bonds of the PDB files?

Thanks! 😸
