### 2023-12-24 #6169 @AnjaConev
Here are some comments (I made related issues in the original repo):

- Please make sure that the LICENSE file is named correctly - https://github.com/suchanek/proteusPy/issues/3
- I detected a couple of minor failures in installation - https://github.com/suchanek/proteusPy/issues/4
- As I was going through the notebooks I was not able to visualize the widgets, I made an issue - https://github.com/suchanek/proteusPy/issues/5
- Examples - you have jupyter notebooks with very nice examples, make sure to reference these notebooks in the README so that the users know where to find them.
- Automated tests - https://github.com/suchanek/proteusPy/issues/6
- Adding more info on related work - https://github.com/suchanek/proteusPy/issues/7
