+++ proteusPy.egg-info/requires.txt
@@ -1,0 +1,20 @@
+numpy
+matplotlib
+pandas
+pyvista
+jupyter
+seaborn
+pillow
+tqdm
+plotly
+datetime
+jupyter_bokeh
+openai
+plotly
+panel
+ipywidgets
+scikit-learn
+gdown
+plotly_express
+trame
+biopython
