+++ proteusPy.yml
@@ -2,7 +2,7 @@ name: proteusPy
 channels:
   - conda-forge
 dependencies:
-  - python=3.9
+  - python=3.10
   - biopython
   - ca-certificates
   - openssl
