+++ proteusPy1.yml
@@ -1,42 +1,0 @@
-name: proteusPy1
-channels:
-  - conda-forge
-dependencies:
-  - python=3.9
-  - Bio
-  - ca-certificates
-  - gdown
-  - openssl
-  - pyvista
-  - jupyter
-  - pandas<2.0.0
-  - matplotlib
-  - notebook
-  - certifi
-  - seaborn
-  - plotly_express
-  - ipygany
-  - qtpy
-  - pyvistaqt
-  - pythreejs
-  - panel
-  - jupyterlab
-  - ipyparallel
-  - pycryptodome
-  - nodejs
-  - pdoc
-  - pillow
-  - pytube
-  - openai
-  - rdkit
-  - yarn
-  - grpcio
-  - conda
-  - pip
-  - scikit-learn
-  - trame
-  - pip:
-    - wget
-    - tqdm
-    - kaleido
-prefix: /Users/suchanek/miniforge3/envs/proteusPy1
