+++ t.yml
@@ -1,397 +1,0 @@
-# packages in environment at /Users/egs/miniforge3/envs/proteus:
-#
-# Name                    Version                   Build  Channel
-_ipython_minor_entry_point 8.7.0                h8cf3c4a_0    conda-forge
-aiohttp                   3.8.3            py39h02fc5c5_1    conda-forge
-aiosignal                 1.3.1              pyhd8ed1ab_0    conda-forge
-anyio                     3.6.2              pyhd8ed1ab_0    conda-forge
-aom                       3.5.0                h7ea286d_0    conda-forge
-appdirs                   1.4.4              pyh9f0ad1d_0    conda-forge
-appnope                   0.1.3              pyhd8ed1ab_0    conda-forge
-argon2-cffi               21.3.0             pyhd8ed1ab_0    conda-forge
-argon2-cffi-bindings      21.2.0           py39h02fc5c5_3    conda-forge
-arrow                     1.2.3                    pypi_0    pypi
-asttokens                 2.2.1              pyhd8ed1ab_0    conda-forge
-astunparse                1.6.3              pyhd8ed1ab_0    conda-forge
-async-timeout             4.0.2              pyhd8ed1ab_0    conda-forge
-attrs                     22.2.0             pyh71513ae_0    conda-forge
-av                        10.0.0                   pypi_0    pypi
-babel                     2.11.0             pyhd8ed1ab_0    conda-forge
-backcall                  0.2.0              pyh9f0ad1d_0    conda-forge
-backports                 1.0                pyhd8ed1ab_3    conda-forge
-backports.functools_lru_cache 1.6.4              pyhd8ed1ab_0    conda-forge
-beautifulsoup4            4.11.1             pyha770c72_0    conda-forge
-biopython                 1.80             py39h02fc5c5_0    conda-forge
-bleach                    5.0.1              pyhd8ed1ab_0    conda-forge
-bokeh                     2.4.3              pyhd8ed1ab_3    conda-forge
-boltons                   23.0.0             pyhd8ed1ab_0    conda-forge
-boost                     1.78.0           py39h99de9ae_4    conda-forge
-boost-cpp                 1.78.0               h1cb353e_1    conda-forge
-brotli                    1.0.9                h1a8c8d9_8    conda-forge
-brotli-bin                1.0.9                h1a8c8d9_8    conda-forge
-brotlipy                  0.7.0           py39h02fc5c5_1005    conda-forge
-bzip2                     1.0.8                h3422bc3_4    conda-forge
-c-ares                    1.18.1               h3422bc3_0    conda-forge
-ca-certificates           2023.5.7             hf0a4a13_0    conda-forge
-cairo                     1.16.0            h73a0509_1014    conda-forge
-certifi                   2023.5.7                 pypi_0    pypi
-cffi                      1.15.1           py39h7e6b969_3    conda-forge
-charset-normalizer        2.1.1              pyhd8ed1ab_0    conda-forge
-click                     8.1.3           unix_pyhd8ed1ab_2    conda-forge
-colorama                  0.4.6              pyhd8ed1ab_0    conda-forge
-colored                   1.4.4                    pypi_0    pypi
-comm                      0.1.2              pyhd8ed1ab_0    conda-forge
-conda                     23.5.0           py39h2804cbe_1    conda-forge
-conda-package-handling    2.0.2              pyh38be061_0    conda-forge
-conda-package-streaming   0.8.0              pyhd8ed1ab_0    conda-forge
-contourpy                 1.0.6            py39haaf3ac1_0    conda-forge
-cryptography              38.0.4           py39he2a39a8_0    conda-forge
-curl                      7.87.0               h9049daf_0    conda-forge
-cycler                    0.11.0             pyhd8ed1ab_0    conda-forge
[…]
