### 2024-01-12 #6169 @Kevin-Mattheus-Moerman
@suchanek it looks like a reviewer provided some feedback for you to work on. Can you provide an update on related changes you will/have made?

@richardjgowers will you spur on this review again to make sure it keeps going? Thanks!
