### 2024-01-14 suchanek/proteusPy#4 @suchanek
Thank you Anja for the comments. I do have a pyPi distro, will try to get it working completely. I will try to include versions too. I've had trouble getting biopython to install properly from setup.py.
