### 2024-01-14 suchanek/proteusPy#7 @suchanek
I'll have a look. There are a number of programs that calculate overall structural parameters, RCSB uses a few to calculate overall structural validity. I haven't seen an aggregated database yet. thanks for the comments.
