### 2024-01-14 suchanek/proteusPy#5 @suchanek
thank you. this is definitely an OpenGL/gpu issue, probably related to WSL. I will try to write up some Windows specific docs. It does run fine under Windows 10.
