### 2024-01-14 suchanek/proteusPy#6 @suchanek
I have a few display tests already and will call those out. Also, you can test each python source program of the package itself by running them, since I've used docstrings throughout. The best overall notebook to run is DisulfideAnalysis.ipynb, since it thoroughly exercises the package. Cheers!
