### 2024-01-15 5c2322b
joss issue, added loader .pkl to repo for download
