+++ setup.py
@@ -75,8 +75,8 @@ setup(name='proteusPy',
       install_requires=['numpy', 'matplotlib', 'pandas', 'pyvista', 
                         'jupyter', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
-                        'plotly', 'panel', 'ipywidgets', 'scikit-learn', 'gdown',
-                        'plotly_express', 'trame', 'biopython'],
+                        'panel', 'ipywidgets', 'scikit-learn', 'gdown',
+                        'plotly_express', 'trame', ],
       source='https://github.com/suchanek/proteusPy/',
       project_urls={
         "Documentation": "https://suchanek.github.io/proteusPy/",
@@ -85,7 +85,7 @@ setup(name='proteusPy',
       },
       classifiers=[
         'Development Status :: 4 - Beta',
-        'License :: OSI Approved :: MIT License',
+        'License :: OSI Approved :: BSD License',
         'Intended Audience :: Developers',
         'Intended Audience :: Science/Research',
         'Operating System :: OS Independent',
@@ -96,6 +96,6 @@ setup(name='proteusPy',
         'Programming Language :: Python :: 3.10'],
       include_package_data=True,
       package_data={'proteusPy': ['data/*.txt', 'data/*.py', 'data/*.json', 'data/*.csv']},
-      exclude_package_data={'proteusPy': ['data/PDB_all_ss.pkl', 'data/PDB_SS_ALL_LOADER.pkl', 'PDB_all_ss_dict.pkl']},
+      exclude_package_data={'proteusPy': ['data/PDB_all_ss.pkl', 'data/PDB_SS_ALL_LOADER.pkl', 'data/PDB_all_ss_dict.pkl']},
       zip_safe=False)
       
