### 2024-01-15 56d87fc
joss issue, added loader .pkl to repo for download
