+++ proteusPy/DisulfideLoader.py
@@ -543,40 +543,145 @@ class DisulfideLoader:
    
 # class ends
 
-def Download_PDB_SS(loadpath=DATA_DIR, verbose=False, subset=False):
+def oDownload_PDB_SS(loadpath=DATA_DIR, verbose=False, subset=False):
     '''
-    Download the databases from Google Drive.
+    Download the databases from Github.
 
     :param loadpath: Path from which to load, defaults to DATA_DIR
     :param verbose: Verbosity, defaults to False
     '''
     # normally the .pkl files are local, EXCEPT for the first run from a newly-installed proteusPy 
     # distribution. In that case we need to download the files for all disulfides and the subset
-    # from the Google Drive storage.
+    # from GitHub. Complicated by the fact that they are git-lfs storage.
     
-    import gdown
-    
+    import urllib.request
+    from proteusPy.data import LOADER_GITHUB_All_URL, LOADER_GITHUB_SUBSET_URL
+
     _good1 = 0 # all data
     _good2 = 0 # subset data
     
     _fname_sub = f'{loadpath}{LOADER_SUBSET_FNAME}'
     _fname_all = f'{loadpath}{LOADER_FNAME}'
     
-    
     if verbose:
-        print(f'-> Download_PDB_SS(): Reading disulfides from Google Drive... ')
+        print(f'-> Download_PDB_SS(): Reading disulfides from GitHub ... ')
     
     if subset:
-        if gdown.download(LOADER_SUBSET_URL, _fname_sub, quiet=False) is not None:
+        if urllib.request.urlretrieve(LOADER_GITHUB_SUBSET_URL, _fname_sub) is not None:
             os.sync()
             if os.path.exists(_fname_sub):
                 _good2 = 2
     else:
-        if gdown.download(LOADER_ALL_URL, _fname_all, quiet=False) is not None:
+        if urllib.request.urlretrieve(LOADER_GITHUB_All_URL, _fname_all) is not None:
             os.sync()
             if os.path.exists(_fname_all):
                 _good1 = 1
 
+    return _good1 + _good2
+
+import requests
+def retrieve_git_lfs_files(repo_url, objects):
+        """
+        Retrieves a git-lfs json object from a specified repo.
+        It does NOT download the file.
+        """
+        batch_url = f"{repo_url.rstrip('/')}/info/lfs/objects/batch"
+        headers = {
+            "Accept": "application/vnd.git-lfs+json",
+            "Content-type": "application/json"
+        }
+        data = {
+            "operation": "download",
+            "transfer": ["basic"],
+            "objects": objects
+        }
+
+        response = requests.post(batch_url, headers=headers, json=data)
+        if response.status_code == 200:
+            # Process the response or save the files
+            # For example, you can access the file contents using response.json()
+            # and save them to the desired location on your system.
+            return response.json()
+        else:
+            # Handle error case
+            print(f"Error: {response.status_code} - {response.text}")
+            return None
+
+def Download_PDB_SS(loadpath=DATA_DIR, verbose=True, subset=False):
+    '''
+    Download the databases from Github.
+
+    :param loadpath: Path from which to load, defaults to DATA_DIR
+    :param verbose: Verbosity, defaults to False
+    '''
+
+    import urllib
+    _good1 = 0 # all data
+    _good2 = 0 # subset data
+    
+    _fname_sub = f'{loadpath}{LOADER_SUBSET_FNAME}'
+    _fname_all = f'{loadpath}{LOADER_FNAME}'
+    
+    def retrieve_git_lfs_files(repo_url, objects):
+        """
+        Retrieves a git-lfs json object from a specified repo.
+        It does NOT download the file.
+        """
+        batch_url = f"{repo_url.rstrip('/')}/info/lfs/objects/batch"
+        headers = {
[…]
