+++ proteusPy/data/__init__.py
@@ -37,7 +37,6 @@ LOADER_SUBSET_FNAME_URL = "https://raw.githubusercontent.com/suchanek/proteusPy/
 LOADER_ALL_URL = "https://drive.google.com/uc?id=1igF-sppLPaNsBaUS7nkb13vtOGZZmsFp"
 LOADER_SUBSET_URL = "https://drive.google.com/uc?id=1puy9pxrClFks0KN9q5PPV_ONKvL-hg33"
 
-
 SS_CLASS_DEFINITIONS = '''
 Idx,chi1_s,chi2_s,chi3_s,chi4_s,chi5_s,class_id,SS_Classname,FXN
 0,-1,-1,-1,-1,-1,00000,-LHSpiral,UNK
