+++ proteusPy/__init__.py
@@ -1,12 +1,12 @@
 # Initialization for the proteusPy package
-# Copyright (c) 2023 Eric G. Suchanek, PhD., all rights reserved
+# Copyright (c) 2024 Eric G. Suchanek, PhD., all rights reserved
 # Subject to the MIT public license.
 
 """
 .. include:: ../README.md
 """
 
-__version__ = "0.82"
+__version__ = "0.83"
 
 import sys
 import os
@@ -24,14 +24,11 @@ import plotly
 import pyvista as pv
 import panel as pn
 
-
-
 from Bio.PDB.vectors import calc_dihedral, calc_angle
 from Bio.PDB import Select, Vector
 
 from .ProteusPyWarning import ProteusPyWarning
 from .ProteusGlobals import PDB_DIR, MODEL_DIR
-
 from .DisulfideExceptions import DisulfideIOException, DisulfideConstructionWarning, DisulfideConstructionException
 
 from .turtle3D import Turtle3D
@@ -47,7 +44,7 @@ from .data import *
 from .angle_annotation import *
 
 from .utility import distance_squared, distance3d, get_jet_colormap, Check_chains, print_memory_used
-from .utility import image_to_ascii_art, generate_vector_dataframe
+from .utility import image_to_ascii_art, generate_vector_dataframe, retrieve_git_lfs_files
 
 from .DisulfideClasses import create_classes, create_quat_classes, plot_count_vs_classid, \
             plot_count_vs_class_df, get_half_quadrant, get_quadrant, get_sixth_quadrant, \
