+++ README.md
@@ -4,6 +4,8 @@
 
 ## Virtual Environment Creation
 
+These instructions are Unix/Mac specific. The should also work under WSL.
+
 1. *Install Anaconda (<http://anaconda.org>)*
 2. *Build the environment.*
    It's simplest to clone the repo via github since it contains all of the notebooks, test programs and raw Disulfide databases. The source code distribution can be also be used from pyPi as a normal package, within your own environment.
@@ -13,13 +15,13 @@
        - From a shell prompt:
 
          ```console
-          $ git-lfs track "*.csv" "*.pkl" "*.mp4"
           $ git clone https://github.com/suchanek/proteusPy
           $ cd proteusPy
+          $ git-lfs track "*.csv" "*.pkl" "*.mp4"
           $ conda env create --name proteusPy --file=proteusPy.yml
           $ conda activate proteusPy
           $ pip install .
-          $ jupyter nbextension enable --py --sys-prefix widgetsnbextension
+          $ sudo jupyter nbextension enable --py --sys-prefix widgetsnbextension
 
 I hope you enjoy using proteusPy and would love to hear any success/insights gleaned from it. The Disulfide database is unique as far as I know, and is ripe for mining.
 
