### 2024-01-15 45b4bdc
update docs, JOSS changes
