### 2024-01-15 6dfeefa
cleanup for JOSS
