+++ MANIFEST.in
@@ -1,1 +1,0 @@
-exclude data/PDB_SS_ALL_LOADER.pkl, data/PDB_SS_SUBSET_LOADER.pkl
