### 2024-01-15 e1c349d
cleanup for JOSS
