+++ programs/DBViewer_Working4.py
@@ -1,0 +1,334 @@
+'''
+RCSB Disulfide Bond Database Browser
+Author: Eric G. Suchanek, PhD
+Last revision: 1/19/2024
+'''
+
+import sys
+import time
+import pyvista as pv
+import panel as pn
+
+import proteusPy
+
+from proteusPy.Disulfide import Disulfide
+from proteusPy.DisulfideLoader import Load_PDB_SS
+from proteusPy.DisulfideList import DisulfideList
+
+pn.extension('vtk', sizing_mode='stretch_width', template='fast')
+
+_vers = 0.5
+
+_default_rcsid = '2q7q'
+_default_ss = '2q7q_75D_140D'
+_default_style = "Ball and Stick"
+_default_single = True
+
+_ssidlist = [
+    '2q7q_75D_140D',
+    '2q7q_81D_113D',
+    '2q7q_88D_171D',
+    '2q7q_90D_138D',
+    '2q7q_91D_135D',
+    '2q7q_98D_129D',
+    '2q7q_130D_161D']
+
+# globals
+ss_state = {}
+vers = tot = pdbs = 0
+render_style = _default_style
+single_style = _default_single
+rcsid = _default_rcsid
+
+RCSB_list = []
+
+orientation_widget = True
+enable_keybindings = True
+
+'''
+from bokeh.plotting import figure
+
+def app():
+    p = figure()
+    p.line([1, 2, 3], [1, 2, 3])
+    return p
+
+pn.serve(app, threaded=True)
+
+pn.state.execute(lambda: p.y_range.update(start=0, end=4))
+'''
+
+def get_theme() -> str:
+    """Return the current theme: 'default' or 'dark'
+
+    Returns:
+        str: The current theme
+    """
+    args = pn.state.session_args
+    if "theme" in args and args["theme"][0] == b"dark":
+        return "dark"
+    return "default"
+
+
+def click_plot(event):
+    """Force a re-render of the currently selected disulfide. Removes the pane
+    and re-adds it to the panel.
+
+    Returns:
+        None
+    """
+    global render_win, vtkpan
+    plotter = render_ss()
+
+    ss = get_ss()
+    if ss is not None:
+        update_title(ss)
+        update_info(ss)
+    else:
+        update_output("# Error: \nClick_plot can't find a disulfide for the current UI?")
+    
+    vtkpan = pn.pane.VTK(plotter.ren_win, margin=0, sizing_mode='stretch_both', 
+                     orientation_widget=orientation_widget,
+                     enable_keybindings=enable_keybindings, min_height=500)
+    
+    vtkpan.param.trigger('object')
+    render_win[0] = vtkpan
+   
+    # this position is dependent on the vtk panel position in the render_win pane!
+    # print(f'RenderWin: {render_win}')
+
+# Widgets
+
+rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=_default_ss, options=_ssidlist)
+
+button = pn.widgets.Button(name='Refresh', button_type='primary')
+button.on_click(click_plot)
+
+styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
+                                        options=['Split Bonds', 'CPK', 'Ball and Stick'], 
+                                        inline=False)
+
+single_checkbox = pn.widgets.Checkbox(name='Single View', value=True)
+
+def update_info_cb(event):
+    ss_id = event.new
+    update_output(f'Got id: {ss_id}')
+    ss = get_ss()
+    update_info(ss)
+
+def update_info(ss: Disulfide):
+    enrg = ss.energy
+    name = ss.name
[…]
