+++ programs/DBViewer_Working3.py
@@ -1,0 +1,403 @@
+'''
+RCSB Disulfide Bond Database Browser
+Author: Eric G. Suchanek, PhD
+Last revision: 1/14/2024
+'''
+
+import sys
+import time
+import pyvista as pv
+import panel as pn
+
+import proteusPy
+from proteusPy.Disulfide import Disulfide
+from proteusPy.DisulfideLoader import Load_PDB_SS
+from proteusPy.DisulfideList import DisulfideList
+
+
+_vers = 0.6
+
+_rcsid = '2q7q'
+_default_ss = '2q7q_75D_140D'
+_ssidlist = [
+    '2q7q_75D_140D',
+    '2q7q_81D_113D',
+    '2q7q_88D_171D',
+    '2q7q_90D_138D',
+    '2q7q_91D_135D',
+    '2q7q_98D_129D',
+    '2q7q_130D_161D']
+
+_style = 'Split Bonds'
+_single = True
+
+PDB_SS = Load_PDB_SS(verbose=True, subset=False)
+
+ss_state = {}
+vers = tot = pdbs = 0
+
+orientation_widget = True
+enable_keybindings = True
+
+RCSB_list = []
+
+app = pn.template.FastListTemplate(
+    title=f"RCSB Disulfide Browser: {tot} Disulfides, {pdbs} Structures, V{vers}",
+    )
+
+pn.extension('vtk', sizing_mode='stretch_width')
+
+# a few widgets
+styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
+                                        options=['Split Bonds', 'CPK', 'Ball and Stick'], 
+                                        inline=False)
+
+single_checkbox = pn.widgets.Checkbox(name='Single View', value=True)
+rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=_default_ss, options=_ssidlist)
+# not used atm    
+shadows_checkbox = pn.widgets.Checkbox(name='Shadows', value=False)
+
+rcsb_selector_widget = pn.widgets.Select(name="RCSB ID", value=_rcsid, options=RCSB_list)
+
+'''
+rcsb_selector_widget = pn.widgets.AutocompleteInput(name="RCSB ID (start typing)", 
+													value=_rcsid, restrict=True,
+                                                    placeholder="Search Here", 
+                                                    options=RCSB_list)
+'''
+
+button = pn.widgets.Button(name='Refresh', button_type='primary')
+
+# controls on sidebar
+ss_props = pn.WidgetBox('# Disulfide Selection',
+                        rcsb_selector_widget, rcsb_ss_widget
+                        ).servable(target='sidebar')
+
+ss_styles = pn.WidgetBox('# Rendering Styles',
+                         styles_group, single_checkbox, button
+                        ).servable(target='sidebar')
+
+# markdown panels for various text outputs
+title_md = pn.pane.Markdown("Title")
+output_md = pn.pane.Markdown("Output goes here")
+db_md = pn.pane.Markdown("Database Info goes here")
+
+info_md = pn.pane.Markdown("SS Info")
+ss_info = pn.WidgetBox('# Disulfide Info', info_md).servable(target='sidebar')
+db_info = pn.Column('### RCSB Database Info', db_md)
+
+
+RCSB_list = []
+
+# pn.state.template.param.update(title=f"RCSB Disulfide Browser: {tot:,} Disulfides, {pdbs:,} Structures, V{vers}")
+
+# default selections
+ss_state_default = {
+    'single': "True",
+    'style': "sb",
+    'rcsb_list': "['2q7q']", 
+    'rcsid': "2q7q",
+    'defaultss': "2q7q_75D_140D",
[…]
