+++ programs/DBViewer_Working3.py
@@ -14,10 +14,13 @@ from proteusPy.Disulfide import Disulfide
 from proteusPy.DisulfideLoader import Load_PDB_SS
 from proteusPy.DisulfideList import DisulfideList
 
+pn.extension('vtk', sizing_mode='stretch_width', template='fast')
 
 _vers = 0.6
 
-_rcsid = '2q7q'
+# defaults for the UI
+
+_rcsid_default = '2q7q'
 _default_ss = '2q7q_75D_140D'
 _ssidlist = [
     '2q7q_75D_140D',
@@ -28,24 +31,18 @@ _ssidlist = [
     '2q7q_98D_129D',
     '2q7q_130D_161D']
 
+_rcsid = '2q7q'
 _style = 'Split Bonds'
 _single = True
-
-PDB_SS = Load_PDB_SS(verbose=True, subset=False)
-
-ss_state = {}
-vers = tot = pdbs = 0
-
 orientation_widget = True
 enable_keybindings = True
 
+# globals
+ss_state = {}
+vers = tot = pdbs = 0
 RCSB_list = []
 
-app = pn.template.FastListTemplate(
-    title=f"RCSB Disulfide Browser: {tot} Disulfides, {pdbs} Structures, V{vers}",
-    )
-
-pn.extension('vtk', sizing_mode='stretch_width')
+#PDB_SS = Load_PDB_SS(verbose=True, subset=False)
 
 # a few widgets
 styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
@@ -57,14 +54,14 @@ rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=_default_ss, options=
 # not used atm    
 shadows_checkbox = pn.widgets.Checkbox(name='Shadows', value=False)
 
-rcsb_selector_widget = pn.widgets.Select(name="RCSB ID", value=_rcsid, options=RCSB_list)
+#rcsb_selector_widget = pn.widgets.Select(name="RCSB ID", value=_rcsid, options=RCSB_list)
 
-'''
+
 rcsb_selector_widget = pn.widgets.AutocompleteInput(name="RCSB ID (start typing)", 
-													value=_rcsid, restrict=True,
+													value=_rcsid_default, restrict=True,
                                                     placeholder="Search Here", 
                                                     options=RCSB_list)
-'''
+
 
 button = pn.widgets.Button(name='Refresh', button_type='primary')
 
@@ -86,9 +83,6 @@ info_md = pn.pane.Markdown("SS Info")
 ss_info = pn.WidgetBox('# Disulfide Info', info_md).servable(target='sidebar')
 db_info = pn.Column('### RCSB Database Info', db_md)
 
-
-RCSB_list = []
-
 # pn.state.template.param.update(title=f"RCSB Disulfide Browser: {tot:,} Disulfides, {pdbs:,} Structures, V{vers}")
 
 # default selections
@@ -102,9 +96,11 @@ ss_state_default = {
         '2q7q_91D_135D','2q7q_98D_129D']",
     }
 
-def set_default_state():
+def set_widgets_defaults():
+    global RCSB_list
     styles_group.value = "Split Bonds"
     single_checkbox.value = True
+    rcsb_selector_widget.options = RCSB_list
     rcsb_selector_widget.value = '2qyq'
     rcsb_ss_widget.value = '2q7q_75D_140D'
     return(ss_state_default)
@@ -113,10 +109,10 @@ def set_state(event):
     """
     Set the ss_state dict to the state variables and UI interaface. Push to cache.
     """
-    global ss_state, _rcsid, _ssidlist, _default_ss, single_checkbox, styles_group
[…]
