+++ programs/DBViewer_Working4.py
@@ -39,24 +39,14 @@ vers = tot = pdbs = 0
 render_style = _default_style
 single_style = _default_single
 rcsid = _default_rcsid
+ssidlist = _ssidlist
+default_ss = _default_ss
 
 RCSB_list = []
 
 orientation_widget = True
 enable_keybindings = True
 
-'''
-from bokeh.plotting import figure
-
-def app():
-    p = figure()
-    p.line([1, 2, 3], [1, 2, 3])
-    return p
-
-pn.serve(app, threaded=True)
-
-pn.state.execute(lambda: p.y_range.update(start=0, end=4))
-'''
 
 def get_theme() -> str:
     """Return the current theme: 'default' or 'dark'
@@ -91,22 +81,10 @@ def click_plot(event):
                      orientation_widget=orientation_widget,
                      enable_keybindings=enable_keybindings, min_height=500)
     
-    vtkpan.param.trigger('object')
-    render_win[0] = vtkpan
+    #vtkpan.param.trigger('object')
     update_output(f'RenderWin: {render_win}')
 
-# Widgets
-
-rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=_default_ss, options=_ssidlist)
-
-button = pn.widgets.Button(name='Refresh', button_type='primary')
-button.on_click(click_plot)
-
-styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
-                                        options=['Split Bonds', 'CPK', 'Ball and Stick'], 
-                                        inline=False)
-
-single_checkbox = pn.widgets.Checkbox(name='Single View', value=True)
+    render_win[0] = vtkpan
 
 def update_info_cb(event):
     ss_id = event.new
@@ -122,6 +100,12 @@ def update_info(ss: Disulfide):
     info_string = f'### {name}  \n**Resolution:** {resolution:.2f} Å  \n**Energy:** {enrg:.2f} kcal/mol  \n**Cα distance:** {ss.ca_distance:.2f} Å  \n**Cβ distance:** {ss.cb_distance:.2f} Å  \n**Torsion Length:** {ss.torsion_length:.2f}°'
     info_md.object = info_string
 
+# Widgets
+styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
+                                        options=['Split Bonds', 'CPK', 'Ball and Stick'], 
+                                        inline=False)
+
+single_checkbox = pn.widgets.Checkbox(name='Single View', value=True)
 
 def update_single(click):
     """Toggle the rendering style radio box depending on the state of the Single View checkbox.
@@ -135,9 +119,13 @@ def update_single(click):
     else:
         styles_group.disabled = False
 
+button = pn.widgets.Button(name='Refresh', button_type='primary')
+button.on_click(click_plot)
+
 # not used atm    
 shadows_checkbox = pn.widgets.Checkbox(name='Shadows', value=False)
 
+rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=default_ss, options=ssidlist)
 rcsb_selector_widget = pn.widgets.AutocompleteInput(name="RCSB ID", value=rcsid, restrict=True,
                                                     placeholder="Search Here", options=RCSB_list)
 
@@ -179,35 +167,9 @@ def load_data():
     rcsb_selector_widget.options = RCSB_list
     return PDB_SS    
 
[…]
