### 2024-01-21 suchanek/proteusPy#4 @suchanek
I've reduced to 1 .yml file, and am clarifying instructions. Mamba is a good suggestion. Still poking at the pypi issues.
