### 2024-01-22 fa20331
linux installation work, viewer
