+++ setup.py
@@ -75,8 +75,8 @@ setup(name='proteusPy',
       install_requires=['numpy', 'matplotlib', 'pandas', 'pyvista', 
                         'jupyter', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
-                        'panel', 'ipywidgets', 'scikit-learn', 'gdown',
-                        'plotly_express', 'trame', ],
+                        'panel', 'scikit-learn', 'gdown',
+                        'plotly_express', 'trame', 'jupyter_contrib_nbextensions', 'ipywidgets'],
       source='https://github.com/suchanek/proteusPy/',
       project_urls={
         "Documentation": "https://suchanek.github.io/proteusPy/",
