### 2024-01-22 be1c7be
linux installation, viewer
