+++ setup.py
@@ -69,14 +69,15 @@ setup(name='proteusPy',
       author='Eric G. Suchanek, PhD',
       author_email='suchanek@mac.com',
       license='MIT',
-      requires_python = "=3.10",
+      requires_python = ">=3.10",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
       install_requires=['numpy', 'matplotlib', 'pandas', 'pyvista', 
-                        'jupyter', 'seaborn', 'pillow',
+                        'notebook', 'jupyter', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown',
-                        'plotly_express', 'trame', 'jupyter_contrib_nbextensions', 'ipywidgets'],
+                        'plotly_express', 'trame', 'jupyter_contrib_nbextensions',
+                        'ipywidgets'],
       source='https://github.com/suchanek/proteusPy/',
       project_urls={
         "Documentation": "https://suchanek.github.io/proteusPy/",
