### 2024-01-22 24394a0
linux installation, viewer
