+++ proteusPy.yml
@@ -2,7 +2,7 @@ name: proteusPy
 channels:
   - conda-forge
 dependencies:
-  - python=3.10
+  - python>=3.10
   - biopython
   - scikit-learn
   - ca-certificates
@@ -24,13 +24,13 @@ dependencies:
   - jupyter_bokeh
   - ipyparallel
   - pycryptodome
+  - notebook
+  - jupyter_contrib_nbextensions
   - nodejs
   - pillow
   - pytube
   - openai
-  - rdkit
   - grpcio
-  - conda=23.5.0
   - pip
   - pip:
     - wget
