+++ programs/t.yml
@@ -1,399 +1,0 @@
-name: proteus
-channels:
-  - conda-forge
-dependencies:
-  - _ipython_minor_entry_point=8.7.0=h8cf3c4a_0
-  - aiohttp=3.8.3=py39h02fc5c5_1
-  - aiosignal=1.3.1=pyhd8ed1ab_0
-  - anyio=3.6.2=pyhd8ed1ab_0
-  - aom=3.5.0=h7ea286d_0
-  - appdirs=1.4.4=pyh9f0ad1d_0
-  - appnope=0.1.3=pyhd8ed1ab_0
-  - argon2-cffi=21.3.0=pyhd8ed1ab_0
-  - argon2-cffi-bindings=21.2.0=py39h02fc5c5_3
-  - asttokens=2.2.1=pyhd8ed1ab_0
-  - astunparse=1.6.3=pyhd8ed1ab_0
-  - async-timeout=4.0.2=pyhd8ed1ab_0
-  - attrs=22.2.0=pyh71513ae_0
-  - babel=2.11.0=pyhd8ed1ab_0
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
-  - beautifulsoup4=4.11.1=pyha770c72_0
-  - biopython=1.80=py39h02fc5c5_0
-  - bleach=5.0.1=pyhd8ed1ab_0
-  - bokeh=2.4.3=pyhd8ed1ab_3
-  - boltons=23.0.0=pyhd8ed1ab_0
-  - boost=1.78.0=py39h99de9ae_4
-  - boost-cpp=1.78.0=h1cb353e_1
-  - brotli=1.0.9=h1a8c8d9_8
-  - brotli-bin=1.0.9=h1a8c8d9_8
-  - brotlipy=0.7.0=py39h02fc5c5_1005
-  - bzip2=1.0.8=h3422bc3_4
-  - c-ares=1.18.1=h3422bc3_0
-  - ca-certificates=2023.5.7=hf0a4a13_0
-  - cairo=1.16.0=h73a0509_1014
-  - cffi=1.15.1=py39h7e6b969_3
-  - charset-normalizer=2.1.1=pyhd8ed1ab_0
-  - click=8.1.3=unix_pyhd8ed1ab_2
-  - colorama=0.4.6=pyhd8ed1ab_0
-  - comm=0.1.2=pyhd8ed1ab_0
-  - conda=23.5.0=py39h2804cbe_1
-  - conda-package-handling=2.0.2=pyh38be061_0
-  - conda-package-streaming=0.8.0=pyhd8ed1ab_0
-  - contourpy=1.0.6=py39haaf3ac1_0
-  - cryptography=38.0.4=py39he2a39a8_0
-  - curl=7.87.0=h9049daf_0
-  - cycler=0.11.0=pyhd8ed1ab_0
-  - debugpy=1.6.4=py39h23fbdae_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - defusedxml=0.7.1=pyhd8ed1ab_0
-  - docker-pycreds=0.4.0=py_0
-  - double-conversion=3.2.0=hb7217d7_1
-  - eigen=3.4.0=hc021e02_0
-  - entrypoints=0.4=pyhd8ed1ab_0
-  - et_xmlfile=1.1.0=pyhd8ed1ab_0
-  - executing=1.2.0=pyhd8ed1ab_0
-  - expat=2.5.0=hb7217d7_0
-  - ffmpeg=5.1.2=gpl_hf318d42_105
-  - flit-core=3.8.0=pyhd8ed1ab_0
[…]
