+++ proteusPy.egg-info/SOURCES.txt
@@ -1,4 +1,4 @@
-MANIFEST.in
+LICENSE
 README.md
 setup.cfg
 setup.py
