+++ proteusPy.yml
@@ -24,7 +24,7 @@ dependencies:
   - jupyter_bokeh
   - ipyparallel
   - pycryptodome
-  - notebook
+  - notebook=6.4.12
   - nodejs
   - pillow
   - pytube
