### 2024-01-22 3c84dba
linux installation
