### 2024-01-22 8d7cb48
linux install working
