+++ proteusPy.egg-info/requires.txt
@@ -2,7 +2,7 @@ numpy
 matplotlib
 pandas
 pyvista
-notebook
+notebook==6.4.12
 jupyter
 seaborn
 pillow
