+++ linux.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 git-lfs track "*.csv" "*.pkl" "*.mp4"
 sudo apt-get install dvipng texlive-latex-extra texlive-fonts-recommended cm-super
-mamba env create --file proteusPy.yml&& \\ 
+mamba env create --file proteusPy.yml && \\ 
 conda activate proteusPy && \
-sudo jupyter contrib nbextension install --sys-prefix && \\
+jupyter contrib nbextension install --sys-prefix && \\
 jupyter nbextension enable --py --sys-prefix widgetsnbextension  && \\
-python -m ipykernel install --user--name proteusPy --display-name "Python (proteusPy)"  && \\
+python -m ipykernel install --user --name proteusPy --display-name "Python (proteusPy)"  && \\
 pip install .
 
