### 2024-01-23 ea65059
cleanup programs, installation
