+++ programs/DBViewer_Working.py
@@ -1,228 +1,0 @@
-'''
-RCSB Disulfide Bond Database Browser
-Author: Eric G. Suchanek, PhD
-Last revision: 11/2/2023
-'''
-
-import sys
-import time
-import pyvista as pv
-import panel as pn
-
-import proteusPy
-
-from proteusPy.Disulfide import Disulfide
-from proteusPy.DisulfideLoader import Load_PDB_SS
-from proteusPy.DisulfideList import DisulfideList
-
-pn.extension('vtk', sizing_mode='stretch_width', template='fast')
-
-_vers = 0.5
-
-_rcsid = '2q7q'
-_default_ss = '2q7q_75D_140D'
-_ssidlist = [
-    '2q7q_75D_140D',
-    '2q7q_81D_113D',
-    '2q7q_88D_171D',
-    '2q7q_90D_138D',
-    '2q7q_91D_135D',
-    '2q7q_98D_129D',
-    '2q7q_130D_161D']
-
-PDB_SS = Load_PDB_SS(verbose=True, subset=False)
-
-vers = PDB_SS.version
-tot = PDB_SS.TotalDisulfides
-pdbs = len(PDB_SS.SSDict)
-orientation_widget = True
-enable_keybindings = True
-
-RCSB_list = sorted(PDB_SS.IDList)
-
-pn.state.template.param.update(title=f"RCSB Disulfide Browser: {tot:,} Disulfides, {pdbs:,} Structures, V{vers}")
-
-def get_theme() -> str:
-    """Return the current theme: 'default' or 'dark'
-
-    Returns:
-        str: The current theme
-    """
-    args = pn.state.session_args
-    if "theme" in args and args["theme"][0] == b"dark":
-        return "dark"
-    return "default"
-
-
-def click_plot(event):
-    """Force a re-render of the currently selected disulfide. Removes the pane
-    and re-adds it to the panel.
-
-    Returns:
-        None
-    """
-    global render_win, vtkpan
-    plotter = render_ss()
-    vtkpan = pn.pane.VTK(plotter.ren_win, margin=0, sizing_mode='stretch_both', 
-                         orientation_widget=orientation_widget,
-                         enable_keybindings=enable_keybindings, min_height=500
-                         )
-    # this position is dependent on the vtk panel position in the render_win pane!
-    print(f'RenderWin: {render_win}')
-    render_win[0] = vtkpan
-
-# Widgets
-
-rcsb_ss_widget = pn.widgets.Select(name="Disulfide", value=_default_ss, options=_ssidlist)
-
-button = pn.widgets.Button(name='Refresh', button_type='primary')
-button.on_click(click_plot)
-
-styles_group = pn.widgets.RadioBoxGroup(name='Rending Style', 
-                                        options=['Split Bonds', 'CPK', 'Ball and Stick'], 
-                                        inline=False)
-
-single_checkbox = pn.widgets.Checkbox(name='Single View', value=True)
-
-def update_single(click):
-    """Toggle the rendering style radio box depending on the state of the Single View checkbox.
-    
-    Returns:
-        None
-    """
-    single_checked = single_checkbox.value
-    if single_checked is not True:
-        styles_group.disabled = True
-    else:
-        styles_group.disabled = False
-    click_plot(click)
-
-# not used atm    
-shadows_checkbox = pn.widgets.Checkbox(name='Shadows', value=False)
-
-rcsb_selector_widget = pn.widgets.AutocompleteInput(name="RCSB ID", value=_rcsid, restrict=True,
-                                                    placeholder="Search Here", options=RCSB_list)
-
[…]
