### 2024-01-23 5d9e300
cleanup programs, installation
