+++ update_docs.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 conda activate proteus
+cd ~/repos/proteusPy
 pdoc proteusPy —math -o docs —logo "./logo.png"
-python setup.py sdist
+python -m build
+

[…]
