+++ documents/a1.md
@@ -1,5 +1,0 @@
-$\\sqrt(\chi_{1}^{2}$ + \\chi_{2}^{2} + \\chi_{3}^{2} + \\chi_{4}^{2} + \\chi_{5}^{2})$
-$\chi_{1}$
-
-
-$$≈ α$$
