+++ setup.py
@@ -72,7 +72,7 @@ setup(name='proteusPy',
       requires_python = ">=3.10",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
-      install_requires=['numpy', 'matplotlib', 'pandas', 'pyvista', 
+      install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 
                         'notebook==6.4.12', 'jupyter', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown',
