+++ update_docs.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-conda activate proteus
 cd ~/repos/proteusPy
-pdoc proteusPy —math -o docs —logo "./logo.png"
+conda activate proteusPy && \\ 
+pdoc -o docs --math --logo "./logo.png" ./proteusPy && \\
 python -m build
 
