### 2024-01-23 suchanek/proteusPy#5 @suchanek
Ok I confirmed it won't run under wsl. I do have it running fine under windows 11 virtual machine on my mac. lol
