### 2024-01-24 7a6e47b
small changes in rendering, typo corrections
