### 2024-01-24 d8d0533
small changes in rendering, typo corrections
