+++ update_docs.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
 cd ~/repos/proteusPy
-conda activate proteusPy && \\ 
-pdoc -o docs --math --logo "./logo.png" ./proteusPy && \\
-python -m build
+conda activate proteusPy && pdoc -o docs --math --logo "./logo.png" ./proteusPy
 
