+++ proteusPy.yml
@@ -3,15 +3,13 @@ channels:
   - conda-forge
 dependencies:
   - python<3.12
+  - pandas<2.0
+  - matplotlib
   - biopython
   - scikit-learn
-  - ca-certificates
-  - openssl
   - pyvista
+  - notebook=6.4.12
   - jupyter
-  - pandas
-  - matplotlib
-  - certifi
   - seaborn
   - plotly_express
   - tqdm
@@ -20,8 +18,6 @@ dependencies:
   - pyvistaqt
   - pythreejs
   - panel
-  - pandas<2.0
-  - notebook=6.4.12
   - jupyterlab
   - jupyter_bokeh
   - ipyparallel
