### 2024-01-29 17eae79
adjust setup.py for pandas
