+++ setup.py
@@ -69,7 +69,7 @@ setup(name='proteusPy',
       author='Eric G. Suchanek, PhD',
       author_email='suchanek@mac.com',
       license='MIT',
-      requires_python = ">=3.10",
+      requires_python = "<3.12",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
       install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 
