+++ proteusPy.egg-info/requires.txt
@@ -4,6 +4,7 @@ pandas<2.0
 pyvista
 notebook==6.4.12
 jupyter
+jupyterlab
 seaborn
 pillow
 tqdm
