+++ setup.py
@@ -73,7 +73,7 @@ setup(name='proteusPy',
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
       install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 
-                        'notebook==6.4.12', 'jupyter', 'seaborn', 'pillow',
+                        'notebook==6.4.12', 'jupyter', 'jupyterlab', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown',
                         'plotly_express', 'trame', 'jupyter_contrib_nbextensions',
