### 2024-01-29 bec9a66
tweak for yml, removing extra jupyter entries
