### 2024-01-30 4051cd8
proteuspy.yml changes, windows works
