+++ proteusPy.yml
@@ -3,6 +3,7 @@ channels:
   - conda-forge
 dependencies:
   - python<3.12
+  - notebook=6.4.12
   - pandas<2.0
   - matplotlib
   - biopython
@@ -29,9 +30,9 @@ dependencies:
     - gdown
     - tqdm
     - ipykernel
-    - jupyterlab
     - jupyter_bokeh
     - jupyter_contrib_nbextensions
   - trame
+  - ipywidgets
 prefix: /Users/egs/miniforge3/envs/proteusPy
 
