+++ linux.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
-git-lfs track "*.csv" "*.pkl" "*.mp4"
-sudo apt-get install dvipng texlive-latex-extra texlive-fonts-recommended cm-super
-mamba env create --file proteusPy.yml && \\ 
-conda activate proteusPy && \
+#git-lfs track "*.csv" "*.pkl" "*.mp4"
+#sudo apt-get install dvipng texlive-latex-extra texlive-fonts-recommended cm-super
+#mamba env create --file proteusPy.yml && \\ 
+conda activate ppy && pip install . && \\
 jupyter contrib nbextension install --sys-prefix && \\
 jupyter nbextension enable --py --sys-prefix widgetsnbextension  && \\
-python -m ipykernel install --user --name proteusPy --display-name "Python (proteusPy)"  && \\
-pip install .
+python -m ipykernel install --user --name ppy --display-name "Python (pPy)"
 
