+++ freeze.yml
@@ -1,177 +1,0 @@
-aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1701099584873/work
-aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
-annotated-types==0.6.0
-anyio==4.2.0
-appnope==0.1.3
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
-async-lru==2.0.4
-attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
-Babel==2.14.0
-beautifulsoup4==4.12.2
-biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1704895510696/work
-bleach==6.1.0
-bokeh==3.3.3
-certifi==2023.11.17
-cffi==1.16.0
-charset-normalizer==3.3.2
-colorcet==3.0.1
-comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
-contourpy==1.2.0
-cycler==0.12.1
-DateTime==5.4
-debugpy==1.8.0
-decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
-defusedxml==0.7.1
-distro==1.9.0
-exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
-executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
-fastjsonschema==2.19.1
-filelock==3.13.1
-fonttools==4.47.2
-fqdn==1.5.1
-frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1702645558715/work
-gdown==4.7.1
-h11==0.14.0
-holoviews==1.18.1
-httpcore==1.0.2
-httpx==0.26.0
-hvplot==0.9.1
-idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
-ipykernel==6.28.0
-ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1704718870316/work
-ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1694607144474/work
-isoduration==20.11.0
-jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
-Jinja2==3.1.3
-joblib==1.3.2
-json5==0.9.14
-jsonpointer==2.4
-jsonschema==4.20.0
-jsonschema-specifications==2023.12.1
-jupyter==1.0.0
-jupyter-bokeh==3.0.7
-jupyter-console==6.6.3
-jupyter-events==0.9.0
-jupyter-lsp==2.2.1
-jupyter_client==8.6.0
-jupyter_core==5.7.1
-jupyter_server==2.12.4
-jupyter_server_terminals==0.5.1
-jupyterlab==4.0.10
-jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1694598704522/work
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.25.2
-kiwisolver==1.4.5
-linkify-it-py==2.0.2
-Markdown==3.5.2
-markdown-it-py==3.0.0
-MarkupSafe==2.1.3
-matplotlib==3.8.2
-matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
-mdit-py-plugins==0.4.0
-mdurl==0.1.2
-mistune==3.0.2
[…]
