### 2024-01-30 8f62a01
fixed error in building pyPi dist, version bump
