+++ linux.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #git-lfs track "*.csv" "*.pkl" "*.mp4"
-#sudo apt-get install dvipng texlive-latex-extra texlive-fonts-recommended cm-super
-conda env create --file proteusPy.yml && \\ 
+sudo apt-get install dvipng texlive-latex-extra texlive-fonts-recommended cm-super
+mamba env create --file proteusPy.yml && \\ 
 conda activate proteusPy && pip install . && \\
 jupyter contrib nbextension install --sys-prefix && \\
 jupyter nbextension enable --py --sys-prefix widgetsnbextension  && \\
