### 2024-01-31 3b8429d
installation instructions, added jupyter shell script
