### 2024-02-02 e898068
version 0.88, pip installation working
