+++ setup.py
@@ -72,10 +72,12 @@ setup(name='proteusPy',
       requires_python = "<3.12",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
-      install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 
+      install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 'traitlets==5.9.0',
                         'notebook==6.4.12', 'jupyter', 'jupyterlab', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
-                        'panel', 'scikit-learn', 'gdown',
+                        'panel', 'scikit-learn', 'gdown', 'ipykernel', 'seaborn', 'ipygany',
+                        'qtpy', 'pyvistaqt', 'pythreejs', 'nodejs', 'pytube', 'openai',
+                        'grpcio', 'pip', 'wget',
                         'plotly_express', 'trame', 'jupyter_contrib_nbextensions',
                         'ipywidgets'],
       source='https://github.com/suchanek/proteusPy/',
@@ -90,11 +92,11 @@ setup(name='proteusPy',
         'Intended Audience :: Developers',
         'Intended Audience :: Science/Research',
         'Operating System :: OS Independent',
-        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
         'Topic :: Scientific/Engineering',
         'Topic :: Scientific/Engineering :: Bio-Informatics',
         'Topic :: Scientific/Engineering :: Chemistry',
-        'Programming Language :: Python :: 3.10'],
+        'Programming Language :: Python :: 3.11'],
       include_package_data=True,
       package_data={'proteusPy': ['data/*.txt', 'data/*.py', 'data/*.json', 'data/*.csv']},
       exclude_package_data={'proteusPy': ['data/PDB_all_ss.pkl', 'data/PDB_SS_ALL_LOADER.pkl', 'data/PDB_all_ss_dict.pkl']},
