### 2024-02-02 fee103d
version 0.88, pip installation working
