### 2024-02-02 suchanek/proteusPy#4 @suchanek
I've updated the docs and have a script to install the Jupyter extensions. Installation is much cleaner.
