### 2024-02-02 suchanek/proteusPy#6 @suchanek
I am writing a program to run through the modules. In progress.
