### 2024-02-02 suchanek/proteusPy#7 @suchanek
There is a multitude of programs to visualize protein structures. The unique aspect of the database is that it's a coherent extract of all disulfide bond containing structures in the RCSB. I haven't seen any other database like it. Thoughts?
