### 2024-02-02 6ebaf2c
installation tweaks
