### 2024-02-02 891f7a4
installation tweaks
