### 2024-02-03 7bb8118
installation tweaks
