### 2024-02-03 d0193e3
installation tweaks
