+++ ppy.yml
@@ -14,6 +14,7 @@ dependencies:
     - gdown
     - tqdm
     - ipykernel
-  - trame
-  - ipywidgets
+    - jupyterlab>=3 
+    - ipywidgets 
+    - pyvista[all,trame]
 
