+++ setup.py
@@ -72,11 +72,11 @@ setup(name='proteusPy',
       requires_python = "<3.12",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
-      install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista', 'traitlets==5.9.0',
+      install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista[all,trame]', 'traitlets==5.9.0',
                         'notebook==6.4.12', 'jupyter', 'jupyterlab', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown', 'ipykernel', 'seaborn', 'ipygany',
-                        'qtpy', 'pyvistaqt', 'pythreejs', 'nodejs', 'pytube', 'openai',
+                        'nodejs', 'pytube', 'openai',
                         'grpcio', 'pip', 'wget',
                         'plotly_express', 'trame-jupyter-extension', 'jupyter_contrib_nbextensions',
                         'ipywidgets'],
