+++ ppy.yml
@@ -7,6 +7,8 @@ dependencies:
   - notebook=6.4.12
   - pandas<2.0
   - matplotlib
+  - vtk
+  - nodejs
   - biopython
   - pip
   - pip:
@@ -17,4 +19,5 @@ dependencies:
     - jupyterlab>=3 
     - ipywidgets 
     - pyvista[all,trame]
+    - jupyter_contrib_nbextensions
 
