### 2024-02-04 8e86b7a
cleanup, using makefile
