+++ setup.py
@@ -68,7 +68,7 @@ setup(name='proteusPy',
       url='https://github.com/suchanek/proteusPy/',
       author='Eric G. Suchanek, PhD',
       author_email='suchanek@mac.com',
-      license='MIT',
+      license='BSD',
       requires_python = "<3.12",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
