+++ README.md
@@ -2,10 +2,21 @@
 
 **proteusPy** is a Python package specializing in the modeling and analysis of proteins of known structure with an emphasis on Disulfide Bonds. This package reprises my molecular modeling program [Proteus](https://doi.org/10.1021/bi00368a023), and relies on the [Turtle3D](https://suchanek.github.io/proteusPy/proteusPy/turtle3D.html) class. The turtle implements the functions ``Move``, ``Roll``, ``Yaw``, ``Pitch`` and ``Turn`` for movement in a three-dimensional space. The [Disulfide](https://suchanek.github.io/proteusPy/proteusPy/Disulfide.html) class implements methods to analyze the protein structure stabilizing element known as a *Disulfide Bond*. This class and its underlying methods are being used to perform a structural analysis of over 35,800 disulfide-bond containing proteins in the RCSB protein data bank.
 
-# Installation (MacOS/Linux)
+# General Capabilities
 
-It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installing the Biopython fork is optional unless you want to actually rebuild the database. I highly recommend using Miniforge, especially in MacOS.
+1) Display Disulfides contained in the RCSB interactively in multiple display styles
+2) Calculate geometric and energetic properties about these disulfides
+3) Create structural classes by characterizing the disulfide torsional angles into *n* classes
+4) Overlap disulfides onto a common frame of reference for display
+5) Build protein backbones from a backbone dihedral angle template
 
+*See https://suchanek.github.io/proteusPy/proteusPy.html for the API documentation with examples*
+
+# Installation 
+
+It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installation includes installing my fork of Biopython fork. This is required to rebuild the database. I highly recommend using Miniforge, especially in MacOS.
+
+## MacOS/Linux
 - Install Miniforge: <https://github.com/conda-forge/miniforge> (existing Anaconda installations are fine but please install mamba)
 - Install git-lfs:
   - <https://help.github.com/en/github/managing-large-files/installing-git-large-file-storage>
@@ -14,8 +25,25 @@ It's simplest to clone the repo via github since it contains all of the notebook
   ```console
   $ git clone https://github.com/suchanek/proteusPy.git
   $ git clone https://github.com/suchanek/biopython.git
-  $ git-lfs track "*.csv" "*.mp4"
   $ cd proteusPy
+  $ git-lfs track "*.csv" "*.mp4"
+  $ make pkg
+  $ mamba activate proteusPy
+  $ make install
+  ```
+## Windows
+- Install Miniforge: <https://github.com/conda-forge/miniforge> (existing Anaconda installations are fine but please install mamba)
+- Install git for Windows and configure for Bash:
+  - https://git-scm.com/download/win
+- Install git-lfs:
+  - https://git-lfs.github.com/
+- Install Visual Studio C++ development environment: https://learn.microsoft.com/en-us/cpp/build/vscpp-step-0-installation?view=msvc-170
+- From an Anaconda Powershell prompt while sitting in your repo dir:
+  ```console
+  $ git clone https://github.com/suchanek/proteusPy.git
+  $ git clone https://github.com/suchanek/biopython.git
+  $ cd proteusPy
+  $ git-lfs track "*.csv" "*.mp4"
   $ make pkg
   $ mamba activate proteusPy
   $ make install
@@ -23,9 +51,9 @@ It's simplest to clone the repo via github since it contains all of the notebook
 
 ## General Usage
 
[…]
