+++ paper/paper.md
@@ -32,18 +32,17 @@ The work has resulted in a freely-accessible database of over 120,494 disulfide 
 
 # Statement of Need
 
-Disulfide bonds are the only naturally occuring covalent bond in proteins. They provide important structural stabilization both within and between protein subunits. They are also involved in enzymatic catalysis, regulate protein activities and protect against oxidative stress. 
-I initially implemented ``proteusPy`` to explore the [RCSB Protein Databank](https://www.rcsb.org) and perform a structural analysis of the disulfide bonds contained therein.
+Disulfide bonds represent the sole naturally occurring covalent bond in proteins, playing a pivotal role in structural stabilization within and between protein subunits. Moreover, they participate in enzymatic catalysis, regulate protein activities, and offer protection against oxidative stress. Establishing an accessible structural database of these disulfides would serve as an invaluable resource for exploring these critical structural elements.
 
-This necessitated the implementation of a python-based package of data structures and algorithms capable of manipulation protein structures and rending them. This has led to the creation an object-oriented database capable of introspection, analysis and display. The API [@API] for the package is available online at: <https://suchanek.github.io/proteusPy/proteusPy.html> and provides more details and numerous examples.
+Accordingly, I have developed the `proteusPy` package to delve into the RCSB Protein Data Bank, furnishing tools for visualizing and analyzing the contained disulfide bonds. This endeavor necessitated the creation of a python-based package encompassing data structures and algorithms capable of manipulating protein structures and rendering them. Consequently, an object-oriented database has been crafted, facilitating introspection, analysis, and display. The package's API [@API] is accessible online at: [proteusPy API](https://suchanek.github.io/proteusPy/proteusPy.html), offering comprehensive details and numerous illustrative examples.
 
 # Requirements
 
-1. PC running MacOS, Linux, Windows
+1. PC running MacOS, Linux, Windows with git, git-lfs and compiler installed.
 2. 16 GB RAM
-3. 2 GB disk space
+3. 3 GB disk space
 
-# Installation (MacOS/Linux)
+# Installation
 
 It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installation includes installing my fork of Biopython fork. This is required to rebuild the database. I highly recommend using Miniforge, especially in MacOS.
 
@@ -82,12 +81,12 @@ It's simplest to clone the repo via github since it contains all of the notebook
 
 # General Usage
 
-Once the package is installed one can use the existing notebooks for analysis of the RCSB Disulfide database. The [notebooks](https://github.com/suchanek/proteusPy/blob/master/notebooks/) directory contains all of my Jupyter notebooks and is a good place to start. The [DisulfideAnalysis.ipynb](https://github.com/suchanek/proteusPy/blob/master/notebooks/DisulfideAnalysis.ipynb) notebook contains the first analysis paper. The [programs](https://github.com/suchanek/proteusPy/tree/master/programs) subdirectory contains the primary programs for downloading the RCSB disulfide-containing structure files:
[…]
