+++ proteusPy/__init__.py
@@ -22,7 +22,7 @@ import numpy
 import Bio
 import plotly
 import pyvista as pv
-import panel as pn
+# import panel as pn
 
 from Bio.PDB.vectors import calc_dihedral, calc_angle
 from Bio.PDB import Select, Vector
