+++ proteusPy/DisulfideLoader.py
@@ -4,21 +4,19 @@ the analysis and modeling of protein structures, with an emphasis on disulfide b
 This work is based on the original C/C++ implementation by Eric G. Suchanek. \n
 
 Author: Eric G. Suchanek, PhD
-Last revision: 7/2/2023
+Last revision: 2/9/2024
 '''
 
 # Author: Eric G. Suchanek, PhD.
-# Last modification: 2/20/23
+# Last modification: 2/10/2024
 # Cα N, Cα, Cβ, C', Sγ Å °
 
 import sys
 import copy
 import time
-import matplotlib.pyplot as plt
 
 import pandas as pd
 import pickle
-import numpy as np
 
 import proteusPy
 from proteusPy.ProteusGlobals import PDB_DIR, MODEL_DIR,  REPO_DATA_DIR
@@ -73,10 +71,8 @@ class DisulfideLoader:
     >>> from proteusPy.Disulfide import Disulfide
     >>> from proteusPy.DisulfideLoader import DisulfideLoader
     >>> from proteusPy.DisulfideList import DisulfideList
-    >>> SS1 = DisulfideList([],'tmp1')
-    >>> SS2 = DisulfideList([],'tmp2')
     
-    >>> PDB_SS = DisulfideLoader(verbose=False, subset=True)
+    >>> PDB_SS = DisulfideLoader(verbose=True, subset=True)
 
     Accessing by index value:
     >>> SS1 = PDB_SS[0]
@@ -97,11 +93,13 @@ class DisulfideLoader:
     >>> SSlist = PDB_SS[:4]
     '''
 
-    def __init__(self, verbose=True, datadir=REPO_DATA_DIR, picklefile=SS_PICKLE_FILE, 
-                pickle_dict_file=SS_DICT_PICKLE_FILE,
-                torsion_file=SS_TORSIONS_FILE, quiet=True, 
-                subset=False,
-                cutoff=-1.0):
+    def __init__(self, verbose: bool = True, 
+                 datadir: str = REPO_DATA_DIR, picklefile: str = SS_PICKLE_FILE, 
+                pickle_dict_file: str = SS_DICT_PICKLE_FILE,
+                torsion_file: str = SS_TORSIONS_FILE, 
+                quiet: bool = True, 
+                subset: bool = False,
+                cutoff: float = -1.0) -> None:
         '''
         Initializing the class initiates loading either the entire Disulfide dataset,
         or the 'subset', which consists of the first 1000 PDB structures. The subset
@@ -182,7 +180,20 @@ class DisulfideLoader:
 
     # 
     # overload __getitem__ to handle slicing and indexing, and access by name
+    
     def __getitem__(self, item):
+        """
+        Implements indexing and slicing to retrieve DisulfideList objects from the 
+        DisulfideLoader. Supports:
+        
+        - Integer indexing to retrieve a single DisulfideList 
+        - Slicing to retrieve a subset as a DisulfideList
+        - Lookup by PDB ID to retrieve all Disulfides for that structure
+        - Lookup by full disulfide name
+        
+        Raises DisulfideException on invalid indices or names.
+        """
+
         res = DisulfideList([], 'none')
 
         if isinstance(item, slice):
@@ -266,7 +277,7 @@ class DisulfideLoader:
         '''
         return copy.deepcopy(self)
 
-    def extract_class(self, clsid):
+    def extract_class(self, clsid) -> DisulfideList:
         """
         Return the list of disulfides corresponding to the input `clsid`.
     
@@ -310,7 +321,7 @@ class DisulfideLoader:
             if ss.name == name:
                 return ss  # or ss.copy() !!!
         return None
-    def describe(self):
+    def describe(self) -> None:
         '''
         Provides information about the Disulfide database contained in ```self```.
 
[…]
