+++ setup.py
@@ -69,15 +69,15 @@ setup(name='proteusPy',
       author='Eric G. Suchanek, PhD',
       author_email='suchanek@mac.com',
       license='BSD',
-      requires_python = "<3.12",
+      requires_python = "3.11.7",
       packages=['proteusPy'],
       keywords='proteusPy suchanek disulfide',
       install_requires=['numpy', 'matplotlib', 'pandas<2.0', 'pyvista[all,trame]', 'traitlets==5.9.0',
-                        'notebook==6.4.12', 'jupyter', 'jupyterlab', 'seaborn', 'pillow',
+                        'notebook==6.4.12', 'jupyter', 'jupyterlab>=3', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown', 'ipykernel', 'seaborn', 'ipygany',
                         'nodejs', 'pytube', 'openai',
-                        'grpcio', 'pip', 'wget',
+                        'grpcio', 'pip', 'wget', 'vtk',
                         'plotly_express', 'trame-jupyter-extension', 'jupyter_contrib_nbextensions',
                         'ipywidgets'],
       source='https://github.com/suchanek/proteusPy/',
