### 2024-02-10 a6478b6
setup.py cleanup
