### 2024-02-12 da03bbc
disulfideloader changed to dataclass
