+++ proteusPy/DisulfideLoader.py
@@ -44,9 +44,94 @@ except NameError:
 
 # Now use tqdm as normal, depending on your environment
 
+from dataclasses import dataclass
+
+@dataclass
+class DisulfideLoader:
+    
+    verbose: bool = True
+    datadir: str = REPO_DATA_DIR
+    picklefile: str = SS_PICKLE_FILE
+    pickle_dict_file: str = SS_DICT_PICKLE_FILE
+    torsion_file: str = SS_TORSIONS_FILE  
+    quiet: bool = True
+    subset: bool = False
+    cutoff: float = -1.0
+
+    def __post_init__(self):
+        self.ModelDir = self.datadir
+        self.PickleFile = f'{self.datadir}{self.picklefile}'
+        self.PickleDictFile = f'{self.datadir}{self.pickle_dict_file}'
+        self.PickleClassFile = f'{self.datadir}{SS_CLASS_DICT_FILE}'
+        self.TorsionFile = f'{self.datadir}{self.torsion_file}'
+        self.SSList = DisulfideList([], 'ALL_PDB_SS')
+        self.SSDict = {}
+        self.TorsionDF = pd.DataFrame()
+        self.TotalDisulfides = 0
+        self.IDList = []
+        
+        self.tclass = None        # disulfideClass_constructor to manage classes
+        self.cutoff = self.cutoff      # distance cutoff used to bulid the database
+        self.verbose = self.verbose
+        self.timestamp = time.time()
+        self.version = proteusPy.__version__
+
+        idlist = []
+
+        if self.subset:
+            self.PickleFile = f'{self.datadir}{SS_SUBSET_PICKLE_FILE}'
+            self.PickleDictFile = f'{self.datadir}{SS_SUBSET_DICT_PICKLE_FILE}'
+            self.TorsionFile = f'{self.datadir}{SS_SUBSET_TORSIONS_FILE}'
+        
+        if self.verbose:
+            print(f'-> DisulfideLoader(): Reading disulfides from: {self.PickleFile}... ', end='')
+    
+        with open(self.PickleFile, 'rb') as f:
+            sslist = pickle.load(f)
+            self.SSList = sslist
+            self.TotalDisulfides = len(self.SSList)
+
+        if self.verbose:
+            print(f'done.',)
+
+        if self.verbose:
+            print(f'-> DisulfideLoader(): Reading disulfide dict from: {self.PickleDictFile}...', end='')
+    
+        with open(self.PickleDictFile, 'rb') as f:
+            self.SSDict = pickle.load(f)
+            for key in self.SSDict:
+                idlist.append(key)
+            self.IDList = idlist.copy()
+            totalSS_dict = len(self.IDList)
+    
+        if self.verbose:
+            print(f'done.')
+
+        if self.verbose:
+            print(f'-> DisulfideLoader(): Reading Torsion DF from: {self.TorsionFile}...', end='')
+
+        tmpDF  = pd.read_csv(self.TorsionFile)
+        tmpDF.drop(tmpDF.columns[[0]], axis=1, inplace=True)
+
+        self.TorsionDF = tmpDF.copy()
+        self.TotalDisulfides = len(self.SSList)
+
+        if self.verbose:
+            print(f' done.')
+
+        self.tclass = DisulfideClass_Constructor(self, self.verbose)
+
+        if self.verbose:    
+            print(f'-> DisulfideLoader(): Loading complete.')
+            self.describe()
+        return
+
+from dataclasses import dataclass
+
+@dataclass
 class DisulfideLoader:
     '''
-    This class represents the disulfide database itself and is its primary means of accession. 
+    This dataclass represents the disulfide database itself and is its primary means of accession. 
[…]
