+++ test.py
@@ -1,0 +1,125 @@
+'''
+    This function creates .pkl files needed for the 
+    proteusPy.DisulfideLoader.DisulfideLoader class. 
+    The ```Disulfide``` objects are contained in a ```DisulfideList``` object and 
+    ```Dict``` within these files. In addition, .csv files containing all of 
+    the torsions for the disulfides and problem IDs are written. The optional
+    ```dist_cutoff``` allows for removal of Disufides whose Cα-Cα distance is >
+    than the cutoff value. If it's -1.0 then the function keeps all Disulfides.
+
+    :param numb:           number of entries to process, defaults to all
+    :param verbose:        more messages
+    :param quiet:          turns off DisulfideConstruction warnings
+    :param pdbdir:         path to PDB files
+    :param datadir:        path to resulting .pkl files
+    :param picklefile:     name of the disulfide .pkl file
+    :param torsionfile:    name of the disulfide torsion file .csv created
+    :param problemfile:    name of the .csv file containing problem ids
+    :param dictfile:       name of the .pkl file
+    :param dist_cutoff:    Ca distance cutoff to reject a Disulfide.
+    
+    The following examples illustrate some basic functions of the disulfide classes:
+
+    >>> from proteusPy.Disulfide import Disulfide 
+    >>> from proteusPy.DisulfideLoader import DisulfideLoader
+    >>> from proteusPy.DisulfideList import DisulfideList
+        
+    Instantiate some variables. Note: the list is initialized with an iterable and a name (optional)
+
+    >>> SS = Disulfide('tmp')
+    >>> SSlist = DisulfideList([],'ss')
+    
+    Load the Disulfide subset database. This contains around 8300 disulfides and loads
+    fairly quickly.
+
+    >>> PDB_SS = DisulfideLoader(verbose=False, subset=True)
+
+    The dataset can be indexed numerically, up to index: PDB_SS.Length(). Get the first SS:
+    >>> SS = PDB_SS[0]
+    >>> SS
+    <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>
+
+    The dataset can also be indexed by PDB ID. Get the DisulfideList for ID 4yys:
+
+    >>> SS4yys = PDB_SS['4yys']
+    >>> SS4yys
+    [<Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_56A_98A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_156A_207A, Source: 4yys, Resolution: 1.35 Å>]
+
+    Make some empty disulfides:
+
+    >>> ss1 = Disulfide('ss1')
+    >>> ss2 = Disulfide('ss2')
+
+    Make a DisulfideList containing ss1, named 'tmp':
+
+    >>> sslist = DisulfideList([ss1], 'tmp')
+
+    Append ss2:
+    >>> sslist.append(ss2)
+
+    Extract the first disulfide and print it:
+
+    >>> ss1 = PDB_SS[0]
+    >>> ss1.pprint_all()
+    <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å
+     Proximal Chain fullID: <('4yys', 0, 'A', (' ', 22, ' '))> Distal Chain fullID: <('4yys', 0, 'A', (' ', 65, ' '))> 
+    Proximal Coordinates:
[…]
