### 2024-02-15 #6169 @suchanek
So far:
* Simplified installation
* Have docstring testing in place
* Writing some basic unit tests
* Made some comments on the state of the field.
*
