### 2024-02-15 2233802
cleanup and docstring tests
