### 2024-02-15 02924ad
cleanup and docstring tests
