+++ proteusPy/DisulfideLoader.py
@@ -7,15 +7,11 @@ Author: Eric G. Suchanek, PhD
 Last revision: 2/9/2024
 '''
 
-# Author: Eric G. Suchanek, PhD.
-# Last modification: 2/10/2024
-# Cα N, Cα, Cβ, C', Sγ Å °
-
 import sys
 import copy
 import time
 
-import pandas as pd
+import pandas
 import pickle
 
 import proteusPy
@@ -72,9 +68,7 @@ class DisulfideLoader:
     >>> from proteusPy.DisulfideLoader import DisulfideLoader
     >>> from proteusPy.DisulfideList import DisulfideList
     
-    >>> PDB_SS = DisulfideLoader(verbose=True, subset=True)
-
-    Accessing by index value:
+    >>> PDB_SS = DisulfideLoader(verbose=False, subset=True)
     >>> SS1 = PDB_SS[0]
     >>> SS1
     <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>
@@ -117,7 +111,7 @@ class DisulfideLoader:
         self.TorsionFile = f'{datadir}{torsion_file}'
         self.SSList = DisulfideList([], 'ALL_PDB_SS')
         self.SSDict = {}
-        self.TorsionDF = pd.DataFrame()
+        self.TorsionDF = pandas.DataFrame()
         self.TotalDisulfides = 0
         self.IDList = []
         self.QUIET = quiet
@@ -162,7 +156,7 @@ class DisulfideLoader:
         if self.verbose:
             print(f'-> DisulfideLoader(): Reading Torsion DF from: {self.TorsionFile}...', end='')
 
-        tmpDF  = pd.read_csv(self.TorsionFile)
+        tmpDF  = pandas.read_csv(self.TorsionFile)
         tmpDF.drop(tmpDF.columns[[0]], axis=1, inplace=True)
 
         self.TorsionDF = tmpDF.copy()
@@ -178,7 +172,6 @@ class DisulfideLoader:
             self.describe()
         return
 
-    # 
     # overload __getitem__ to handle slicing and indexing, and access by name
     
     def __getitem__(self, item):
@@ -285,7 +278,7 @@ class DisulfideLoader:
         :return: The list of disulfide bonds from the class.
         """
         
-        from tqdm import tqdm
+        # from tqdm import tqdm
         six = self.tclass.sixclass_df
         tot_classes = six.shape[0]
         class_disulfides = DisulfideList([], clsid, quiet=True)
@@ -326,16 +319,20 @@ class DisulfideLoader:
         Provides information about the Disulfide database contained in ```self```.
 
         Example:
-        >>> from proteusPy.DisulfideLoader import Load_PDB_SS
-        >>> PDB_SS = Load_PDB_SS(verbose=False, subset=False)
-        >>> PDB_SS.describe()
+        
+        from proteusPy.DisulfideLoader import Load_PDB_SS
+        PDB_SS = Load_PDB_SS(verbose=False, subset=False)
+        PDB_SS.describe()
             =========== RCSB Disulfide Database Summary ==============
-        PDB IDs present:                    35818
-        Disulfides loaded:                  120697
-        Average structure resolution:       2.34 Å
-        Lowest Energy Disulfide:            2q7q_75D_140D
-        Highest Energy Disulfide:           1toz_456A_467A
-        Total RAM Used:                     29.26 GB.
+               =========== Built: 2024-02-12 17:48:13 ==============
+       PDB IDs present:                    35818
+       Disulfides loaded:                  120494
+       Average structure resolution:       2.34 Å
+       Lowest Energy Disulfide:            2q7q_75D_140D
[…]
