### 2024-02-17 14aa479
version 0.92, lots of small cleanup, docs
