+++ paper/paper.md
@@ -4,7 +4,7 @@ tags:
   - Python
   - Disulfide Bonds
   - Protein Structure
-  - RCSB protein databank
+  - RCSB Protein Databank
 author: "Eric G. Suchanek, PhD."
 authors:
   - name: Eric G Suchanek
@@ -12,9 +12,9 @@ authors:
     equal-contrib: true
     affiliation: 1
 affiliations:
- - name: tbd
+ - name: Flux-Frontiers
    index: 1
-date: 31 October, 2023
+date: 16 February, 2024
 header-includes:
   - \let\oldAA\AA
   - \renewcommand{\AA}{\text{\normalfont\oldAA}}
@@ -38,13 +38,13 @@ Accordingly, I have developed the `proteusPy` package to delve into the RCSB Pro
 
 # Requirements
 
-1. PC running MacOS, Linux, Windows with git, git-lfs and compiler installed.
+1. PC running MacOS, Linux, Windows with git, git-lfs and C compiler installed.
 2. 16 GB RAM
 3. 3 GB disk space
 
 # Installation
 
-It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installation includes installing my fork of Biopython fork. This is required to rebuild the database. I highly recommend using Miniforge, especially in MacOS.
+It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installation also installs my fork of Biopython. While this is not technically required for using the analytical capabilities of the package it is required to rebuild the database. I highly recommend using Miniforge, especially in MacOS.
 
 ## MacOS/Linux
 - Install Miniforge: <https://github.com/conda-forge/miniforge> (existing Anaconda installations are fine but please install mamba)
@@ -86,7 +86,7 @@ Once the package is installed one can use the existing notebooks for analysis of
 * [DisulfideExtractor.py](https://github.com/suchanek/proteusPy/blob/master/programs/DisulfideExtractor.py): Extracts the disulfides and creating the database loaders
 * [DisulfideClass_Analysis.py](https://github.com/suchanek/proteusPy/blob/master/programs/DisulfideExtractor.py): Performs cluster analysis on the disulfide database.
 
-The first time one loads the database via [Load_PDB_SS()](https://suchanek.github.io/proteusPy/proteusPy/DisulfideLoader.html#Load_PDB_SS) the system will attempt to download the full and subset database from the GitHub repository. If this fails the system will attempt to rebuild the database from the repo's ``data`` subdirectory (not the package's). If you've downloaded from github this will work correctly. If you've installed from pyPi via ``pip`` it will fail.
+The first time one loads the database via [Load_PDB_SS()](https://suchanek.github.io/proteusPy/proteusPy/DisulfideLoader.html#Load_PDB_SS) the system will attempt to download the full and subset database from Google Drive. If this fails the system will attempt to rebuild the database from the repo's ``data`` subdirectory (not the package's). If you've downloaded from github this will work correctly. If you've installed from pyPi via ``pip`` it will fail.
 
 # Class Details
 
@@ -197,7 +197,7 @@ object's orientation in a *local* coordinate frame. The Turtle developed in `pro
 
 By implementing the functions ``Move``, ``Roll``, ``Yaw``, ``Pitch`` and ``Turn`` the turtle is capable of movement in a three-dimensional space. See [@Pabo_1986] for more details.
 
[…]
