### 2024-02-17 e5631be
version 0.92, lots of small cleanup, docs
