+++ update_docs.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-cd ~/repos/proteusPy
-conda activate proteusPy && pdoc -o docs --math --logo "./logo.png" ./proteusPy
-
