+++ README.md
@@ -1,21 +1,28 @@
 # Summary
 
-**proteusPy** is a Python package specializing in the modeling and analysis of proteins of known structure with an emphasis on Disulfide Bonds. This package reprises my molecular modeling program [Proteus](https://doi.org/10.1021/bi00368a023), and relies on the [Turtle3D](https://suchanek.github.io/proteusPy/proteusPy/turtle3D.html) class. The turtle implements the functions ``Move``, ``Roll``, ``Yaw``, ``Pitch`` and ``Turn`` for movement in a three-dimensional space. The [Disulfide](https://suchanek.github.io/proteusPy/proteusPy/Disulfide.html) class implements methods to analyze the protein structure stabilizing element known as a *Disulfide Bond*. This class and its underlying methods are being used to perform a structural analysis of over 35,800 disulfide-bond containing proteins in the RCSB protein data bank.
+**proteusPy** is a Python package specializing in the modeling and analysis of proteins of known structure with an emphasis on Disulfide bonds. This package reprises my molecular modeling program [Proteus](https://doi.org/10.1021/bi00368a023), a structure-based program developed as part of my graduate thesis. The package relies on the [Turtle3D](https://suchanek.github.io/proteusPy/proteusPy/turtle3D.html) class to create and manipulate local coordinate systems. The turtle implements the functions ``Move``, ``Roll``, ``Yaw``, ``Pitch`` and ``Turn`` for movement in a three-dimensional space. This initial implementation focuses on the [Disulfide](https://suchanek.github.io/proteusPy/proteusPy/Disulfide.html) class. The class implements methods to analyze the protein structure stabilizing element known as a *Disulfide Bond*. This class and its underlying methods are being used to perform a structural analysis of over 35,800 disulfide-bond containing proteins in the RCSB protein data bank (https://www.rcsb.org).
 
 # General Capabilities
-
 - Interactively display disulfides contained in the RCSB in a variety of display styles
 - Calculate geometric and energetic properties about these disulfides
-- Create structural classes by characterizing the disulfide torsional angles into *n* classes
+- Create binary and sextant structural classes by characterizing the disulfide torsional angles into *n* classes
+- Build idealized disulfide bonds from dihedral angle input
+- Find disulfide neighbors based on dihedral angle input
 - Overlap disulfides onto a common frame of reference for display
 - Build protein backbones from backbone phi, psi dihedral angle templates
-- more in development
+- More in development
 
 *See https://suchanek.github.io/proteusPy/proteusPy.html for the API documentation with examples*
 
+# Requirements
+
+1. PC running MacOS, Linux, Windows with git, git-lfs and C compiler installed.
+2. 16 GB RAM
+3. 3 GB disk space
+
 # Installation 
 
-It's simplest to clone the repo via github since it contains all of the notebooks, data and test programs. Installation includes installing my Biopython fork. This is required to rebuild the database. I highly recommend using Miniforge since it includes mamba. The installation instructions below assume a clean install with no package manager or compiler installed.
+It's simplest to clone the repo via GitHub since it contains all of the notebooks, data and test programs. Installation includes installing my Biopython fork. This is required to rebuild the database. I highly recommend using Miniforge since it includes mamba. The installation instructions below assume a clean install with no package manager or compiler installed.
 
 ## MacOS/Linux
 - Install Miniforge: <https://github.com/conda-forge/miniforge> (existing Anaconda installations are fine but please install mamba)
@@ -54,11 +61,36 @@ It's simplest to clone the repo via github since it contains all of the notebook
 
[…]
