### 2024-02-18 5ef86b7
corrected error in window title for ss list display
