### 2024-02-18 f97f655
unittests, reorg, version bump 0.92.2
