### 2024-02-18 6aaf8ee
unittests, reorg, version bump 0.92.2
