+++ programs/DisulfideClass_Analysis.py
@@ -1,8 +1,7 @@
 '''
-Disulfide class analysis using `proteusPy.Disulfide`. Disulfide families are defined
-using the +/- formalism of Hogg et al. (Biochem, 2006, 45, 7429-7433), across
-all 32 possible classes ($$2^5$$). Classes are named per Hogg's convention.
-
+Disulfide class analysis using `proteusPy.Disulfide` package. Disulfide binary families are defined
+using the +/- formalism of Schmidt et al. (Biochem, 2006, 45, 7429-7433), across
+all 32 possible classes ($$2^5$$). Classes are named per the paper's convention.
 
 +----+----------+----------+----------+----------+----------+------------+----------------+------------+
 | IDX|   chi1_s |   chi2_s |   chi3_s |   chi4_s |   chi5_s |   class_id | SS_Classname   | FXN        |
@@ -71,88 +70,34 @@ all 32 possible classes ($$2^5$$). Classes are named per Hogg's convention.
 +----+----------+----------+----------+----------+----------+------------+----------------+------------+
 | 31 |        1 |        1 |        1 |        1 |        1 |      22222 | +RHSpiral      | UNK        |
 +----+----------+----------+----------+----------+----------+------------+----------------+------------+
+
+The sextant class approach is unique to proteusPy, wherein the dihedral circle is divided into 6 quadrants,
+and a dihedral angle vector defined by characterizing each dihedral angle into one of these six quadrants.
+This yields $6^{5}$ or 7776 possible classes. This program analyzes the RCSB database and creates graphs illustrating
+the membership across the binary and sextant classes. The graphs are stored in the global SAVE_DIR location.
+Binary analysis takes approximately 28 minutes with Sextant analysis taking about
+75 minutes on a 2023 M3 Max Macbook Pro.
+
+Author: Eric G. Suchanek, PhD.
 '''
 
-
-# DisulfideBond Class Exploration
-# Author: Eric G. Suchanek, PhD.
-# (c) 2023 Eric G. Suchanek, PhD., All Rights Reserved
-# License: MIT
-
-# Last Modification: 2/24/2023
+# Last Modification: 2/19/2024
 # Cα Cβ Sγ
-
-import pandas as pd
 import time
+import numpy as np
+import pickle
+import argparse
 
-import pyvista as pv
-from pyvista import set_plot_theme
+from datetime import timedelta
+from tqdm import tqdm
 
-from Bio.PDB import *
-
-# for using from the repo we 
 import proteusPy
-
-from proteusPy.data import DATA_DIR
-from proteusPy.Disulfide import *
+from proteusPy import Load_PDB_SS, DATA_DIR
 from proteusPy.DisulfideList import DisulfideList
-from proteusPy.DisulfideLoader import Load_PDB_SS, DisulfideLoader
+from proteusPy.Disulfide import Disulfide
+SAVE_DIR = '/Users/egs/Documents/proteusPy/'
 
-SAVE_DIR = '/Users/egs/Documents/proteusPy/ClassAnalysis'
-
-start = time.time()
-
-def analyze_binary_classes(loader: DisulfideLoader, do_graph=True, do_consensus=True) -> DisulfideList:
-    """
-    Analyze the binary disulfide bond classes.
-
-    :param loader: The disulfide loader object.
-    :param do_graph: Whether or not to display torsion statistics graphs. Default is True.
-    :param do_consensus: Whether or not to compute average conformations for each class. Default is True.
-    :param cutoff: The cutoff percentage for each class. If the percentage of disulfides for a class is below
-                   this value, the class will be skipped. Default is 0.1.
-
-    :return: A list of disulfide bonds, where each disulfide bond represents the average conformation for a class.
[…]
