+++ setup.py
@@ -76,8 +76,7 @@ setup(name='proteusPy',
                         'notebook==6.4.12', 'jupyter', 'jupyterlab>=3', 'seaborn', 'pillow',
                         'tqdm', 'plotly', 'datetime', 'jupyter_bokeh', 'openai',
                         'panel', 'scikit-learn', 'gdown', 'ipykernel', 'seaborn', 'ipygany',
-                        'nodejs', 'pytube', 'openai',
-                        'grpcio', 'pip', 'wget', 'vtk',
+                        'nodejs', 'pytube', 'openai','grpcio', 'pip', 'wget', 'vtk', 'kaleido',
                         'plotly_express', 'trame-jupyter-extension', 'jupyter_contrib_nbextensions',
                         'ipywidgets'],
       source='https://github.com/suchanek/proteusPy/',
