### 2024-02-20 suchanek/proteusPy#7 @suchanek
Very good.
