### 2024-02-20 suchanek/proteusPy#6 @suchanek
I've decided to use make to ease install and testing. I have the doctest testing integrated now.
