### 2024-02-20 suchanek/proteusPy#5 @AnjaConev
Yes, it is an issue with wsl - checked on a regular linux and everything worked smoothly!
