### 2024-02-20 suchanek/proteusPy#6 @AnjaConev
Perfect!! One more thing - would it be possible to add the instructions for resting in the readme too - just in case someone else wants to run the tests too. Maybe they are already somewhere and I missed something?
