### 2024-02-20 #6169 @AnjaConev
Great work - I think this package is a great contribution to the field.
You have addressed all my concerns.
There is just one more thing left from my list:
- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Can you add these guidelines to the README?
In the meantime, I will also check the format of references in the text
