+++ paper/joss.bib
@@ -5,6 +5,7 @@
 @article{Hogg_2020,	doi = {10.1021/bi00368a023},	url = {https://doi.org/10.3390/molecules25235729},	year = 2020,	month = {dec},	publisher = {Molecules},	volume = {25},	number = {23},	pages = {5729--5734},	author = {Philip J Hogg},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Molecules}}
 @article{Schmidt_2006,	doi = {10.1021/bi0603064},	url = {https://doi.org/10.1021/bi0603064},	year = 2006,	month = {jun},	publisher = {American Chemical Society ({ACS})},	volume = {45},	number = {24},	pages = {7429–74334},	author = {Brian Schmidt},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Biochemistry}}
 @article{Hamelyrck_2003,	doi = {10.1093/bioinformatics/btg2994},	url = {https://doi.org/10.1093/bioinformatics/btg299},	year = 2003,	month = {nov},	publisher = {Bioinformatics},	volume = {19},	number = {17},	pages = {2308-2310},	author = {Thomas Hamelyrck and Bernard Manderick},	title = {PDB file parser and structure class implemented in Python},	journal = {Bioinformatics}}
+@article{Wong_2010, doi = {10.1111/j.1538-7836.2010.03894.x}, url = {http://dx.doi.org/10.1111/j.1538-7836.2010.03894.x}
 @misc{proteusPy,
   author = {Eric G. Suchanek},
   title = {proteusPy: A Package for Modeling and Analyzing Proteins of Known Structure},
