### 2024-02-21 6b05776
joss work, bib correction
