+++ ppy.yml
@@ -4,7 +4,6 @@ channels:
 dependencies:
   - python==3.11.7
   - traitlets=5.9.0
-  - notebook=6.4.12
   - pandas<2.0
   - matplotlib
   - nodejs
