+++ proteusPy/__init__.py
@@ -6,7 +6,7 @@
 .. include:: ../README.md
 """
 
-__version__ = "0.92.3"
+__version__ = "0.92.5"
 
 import sys
 import os
@@ -21,10 +21,6 @@ import matplotlib
 import Bio
 import plotly
 
-# import pyvista as pv
-# import pandas as pd
-# import panel as pn
-
 from Bio.PDB.vectors import calc_dihedral, calc_angle
 from Bio.PDB import Select, Vector
 
@@ -33,20 +29,22 @@ from .ProteusGlobals import PDB_DIR, MODEL_DIR, WINFRAME, WINSIZE, CAMERA_POS, _
 
 from .DisulfideExceptions import DisulfideIOException, DisulfideConstructionWarning, DisulfideConstructionException
 
-from .turtle3D import Turtle3D
+from .turtle3D import Turtle3D, test_turtle
 from .turtle3D import ORIENT_BACKBONE, ORIENT_SIDECHAIN
 from .Residue import build_residue, get_backbone_from_chain, to_alpha, to_carbonyl, to_nitrogen, to_oxygen
 
-from .Disulfide import Download_Disulfides, Extract_Disulfides
+from .Disulfide import Download_Disulfides, Extract_Disulfides, check_header_from_file, check_header_from_id, Disulfide_Energy_Function
 from .DisulfideList import load_disulfides_from_id
-from .DisulfideLoader import DisulfideLoader, Load_PDB_SS
+from .DisulfideLoader import Load_PDB_SS, Download_PDB_SS, Download_PDB_SS_GitHub
 
 from .atoms import *
 from .data import *
 from .angle_annotation import *
 
-from .utility import distance_squared, distance3d, get_jet_colormap, Check_chains, print_memory_used
-from .utility import image_to_ascii_art, generate_vector_dataframe, retrieve_git_lfs_files
+from .utility import ( 
+    distance_squared, distance3d, get_jet_colormap, Check_chains, print_memory_used,
+    image_to_ascii_art, generate_vector_dataframe, retrieve_git_lfs_files
+)
 
 from .DisulfideClasses import (
         create_classes, create_quat_classes, plot_count_vs_classid,
