### 2024-02-24 42291d1
JOSS work, docs, added fxns to Disulfide.py for modeling
