### 2024-02-24 76e393d
JOSS work, docs, added fxns to Disulfide.py for modeling
