### 2024-02-24 bce9371
example notebooks, makefile work
