+++ makefile
@@ -5,19 +5,28 @@
 CONDA = mamba
 
 # this MUST match the version in proteusPy/__init__.py
-VERS = 0.92.6
-MESS = "JOSS work, docs, added fxns to Disulfide.py for modeling"
+VERS = 0.92.7
+MESS = "example notebooks, makefile work"
 
 DEVNAME = ppydev
 OUTFILES = sdist.out, bdist.out, docs.out
 nuke: clean devclean
 	@rm $(OUTFILES)
 
-dev:
+Odev:
 	@echo "Building... $(DEVNAME)"
 	@$(CONDA) env create --name $(DEVNAME) --file ppy.yml -y -q
-	@$(CONDA) install --name $(DEVNAME) pdoc -y -q
+	@$(CONDA) install --name $(DEVNAME) pdoc twine -y -q
 	@echo "Step 1 done. Now activate the environment with 'conda activate $(DEVNAME)' and run 'make install'"
+
+pkg:
+	@echo "Starting installation step 1/2..."
+	@$(CONDA) create --name proteusPy -y python=3.11.7
+	@echo "Step 1 done. Now activate the environment with 'conda activate proteusPy' and run 'make install'"
+dev:
+	@echo "Building $(DEVNAME)..."
+	$(CONDA) create --name $(DEVNAME) -y python=3.11.7
+	@echo "Step 1 done. Now activate the environment with 'conda activate $(DEVNAME)' and run 'make install_dev'"
 
 clean:
 	@echo "Removing proteusPy environment..."
@@ -27,29 +36,25 @@ devclean:
 	@echo "Removing $(DEVNAME) environment..."
 	$(CONDA) env remove --name $(DEVNAME) -y
 
-pkg:
-	@echo "Starting installation step 1/2..."
-	@$(CONDA) env create --name proteusPy --file ppy.yml -y -q
-	@echo "Step 1 done. Now activate the environment with 'conda activate proteusPy' and run 'make install'"
-
-pkg2:
-	$(CONDA) create --name ppy2 -y python=3.11.7
 	
 # activate the package before running!
-
 install:
 	@echo "Starting installation step 2/2..."
 	@pip install --quiet . && cd ../biopython && pip install --quiet .
-	jupyter contrib nbextension install --sys-prefix
-	jupyter nbextension enable --py --sys-prefix widgetsnbextension
-	python -m ipykernel install --user --name proteusPy --display-name "proteusPy $(VERS)"
-	
-
-install_dev:
-	@pip install -e . && cd ../biopython && pip install .
 	@jupyter contrib nbextension install --sys-prefix
 	@jupyter nbextension enable --py --sys-prefix widgetsnbextension
-	python -m ipykernel install --user --name ppy_dev --display-name "ppy_dev $(VERS)"
+	@python -m ipykernel install --user --name proteusPy --display-name "proteusPy $(VERS)"
+	@echo "Installation finished!"
+
+install_dev:
+	@pip install . 
+	@cd ../biopython 
+	@pip install --quiet . && cd ../biopython && pip install --quiet .
+	@pip install pdoc twine
+	@jupyter contrib nbextension install --sys-prefix
+	@jupyter nbextension enable --py --sys-prefix widgetsnbextension
+	@python -m ipykernel install --user --name ppydev --display-name "ppydev $(VERS)"
+	@echo "Installation finished!"
 
 # package development targets
 build_dev: sdist docs
