### 2024-02-24 d6977f7
corrections for animation creation
