### 2024-02-24 46d3c31
corrections for animation creation
