### 2024-02-25 suchanek/proteusPy#6 @suchanek
Done. Adding the testing allowed me to catch a few problems with the animation rendering. 👍
