### 2024-02-25 49016ec
additions for windows install
