### 2024-02-25 129202d
additions for windows install
