### 2024-02-25 fb8e071
added avg resolution calcs, updated docs
