### 2024-02-25 56f0198
added avg resolution calcs, updated docs
