### 2024-02-25 aa14420
demo notebook, cleanup
