### 2024-02-26 eef7a33
moved versioning into __version__.txt, docs
