### 2024-02-26 e5fc5b6
moved versioning into __version__.txt, docs
