+++ __version__.txt
@@ -1,0 +1,1 @@
+__version__ = "0.92.10"
