+++ __version__.txt
@@ -1,1 +1,0 @@
-__version__ = "0.92.10"
