+++ .pdocconfig.py
@@ -1,0 +1,3 @@
+pdoc_config = {
+    "exclude": ["proteusPy.__version__.py"] 
+}
