+++ proteusPy/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.92.11"
+__version__ = "0.92.12"
