+++ proteusPy/DisulfideLoader.py
@@ -1,11 +1,11 @@
-'''
+"""
 This module is part of the proteusPy package, a Python package for 
 the analysis and modeling of protein structures, with an emphasis on disulfide bonds.
 This work is based on the original C/C++ implementation by Eric G. Suchanek. \n
 
 Author: Eric G. Suchanek, PhD
 Last revision: 2/9/2024
-'''
+"""
 
 import sys
 import copy
@@ -15,10 +15,15 @@ import pandas
 import pickle
 
 import proteusPy
-from proteusPy.ProteusGlobals import PDB_DIR, MODEL_DIR,  REPO_DATA_DIR
+from proteusPy.ProteusGlobals import PDB_DIR, MODEL_DIR, REPO_DATA_DIR
 from proteusPy.atoms import *
 
-from proteusPy.data import SS_PICKLE_FILE, SS_TORSIONS_FILE, SS_DICT_PICKLE_FILE, DATA_DIR
+from proteusPy.data import (
+    SS_PICKLE_FILE,
+    SS_TORSIONS_FILE,
+    SS_DICT_PICKLE_FILE,
+    DATA_DIR,
+)
 from proteusPy.data import LOADER_FNAME, LOADER_SUBSET_FNAME
 
 from proteusPy.DisulfideList import DisulfideList
@@ -31,7 +36,7 @@ from proteusPy.DisulfideClass_Constructor import DisulfideClass_Constructor
 try:
     # Check if running in Jupyter
     shell = get_ipython().__class__.__name__
-    if shell == 'ZMQInteractiveShell':
+    if shell == "ZMQInteractiveShell":
         from tqdm.notebook import tqdm
     else:
         from tqdm import tqdm
@@ -40,84 +45,89 @@ except NameError:
 
 # Now use tqdm as normal, depending on your environment
 
+
 class DisulfideLoader:
-    '''
-    This class represents the disulfide database itself and is its primary means of accession. 
+    """
+    This class represents the disulfide database itself and is its primary means of accession.
     The entirety of the RCSB disulfide database is stored within the class via a
     proteusPy.DisulfideList.DisulfideList, a ```Pandas``` .csv file, and a ```dict``` of
     indices mapping the PDB IDs into their respective list of disulfides. The datastructures allow
-    simple, direct and flexible access to the disulfide structures contained herein. 
+    simple, direct and flexible access to the disulfide structures contained herein.
     This makes it possible to access the disulfides by array index, PDB structure ID or disulfide name.
-    
-    The class can also render Disulfides overlaid on a common coordinate system to a pyVista window using the 
+
+    The class can also render Disulfides overlaid on a common coordinate system to a pyVista window using the
     [display_overlay()](#DisulfideLoader.display_overlay) method. See below for examples.\n
 
     Important note: For typical usage one will access the database via the `Load_PDB_SS()` function.
     The difference is that the latter function loads the compressed database from its single
     source. the `Load_PDB_SS()` function will load the individual torsions and disulfide .pkl,
     builds the classlist structures.
-    
+
     *Developer's Notes:*
     The .pkl files needed to instantiate this class and save it into its final .pkl file are
     defined in the proteusPy.data class and should not be changed. Upon initialization the class
-    will load them and initialize itself. 
+    will load them and initialize itself.
 
     Example:
     >>> import proteusPy
     >>> from proteusPy.Disulfide import Disulfide
     >>> from proteusPy.DisulfideLoader import DisulfideLoader
     >>> from proteusPy.DisulfideList import DisulfideList
-    
+
     >>> PDB_SS = DisulfideLoader(verbose=False, subset=True)
     >>> SS1 = PDB_SS[0]
     >>> SS1
     <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>
-    
+
[…]
