+++ proteusPy/DisulfideClass_Constructor.py
@@ -1,4 +1,4 @@
-'''
+"""
 DisulfideBond Class Analysis Dictionary creation
 Author: Eric G. Suchanek, PhD.
 License: BSD
@@ -6,13 +6,13 @@ Last Modification: 2/19/24 -egs-
 
 Disulfide Class creation and manipulation using the +/- formalism of Hogg et al. (Biochem, 2006, 45, 7429-7433), 
 across all 32 possible classes. Classes are named per Hogg's convention.
-'''
+"""
 
-# this workflow reads in the torsion database, groups it by torsions 
-# to create the classes merges with the master class spreadsheet, and saves the 
+# this workflow reads in the torsion database, groups it by torsions
+# to create the classes merges with the master class spreadsheet, and saves the
 # resulting dict to {DATA_DIR}PDB_SS_merged.csv
 
-__pdoc__ = {'__all__': True}
+__pdoc__ = {"__all__": True}
 
 import pandas as pd
 from io import StringIO
@@ -22,23 +22,48 @@ import tqdm
 from Bio.PDB import *
 
 import proteusPy
-from proteusPy.data import SS_CLASS_DICT_FILE, SS_CLASS_DEFINITIONS, DATA_DIR, CLASSOBJ_FNAME, SS_CONSENSUS_FILE
+from proteusPy.data import (
+    SS_CLASS_DICT_FILE,
+    SS_CLASS_DEFINITIONS,
+    DATA_DIR,
+    CLASSOBJ_FNAME,
+    SS_CONSENSUS_FILE,
+)
 from proteusPy.DisulfideList import DisulfideList
 from proteusPy.Disulfide import Disulfide
 from proteusPy.ProteusGlobals import DPI
 
-merge_cols = ['chi1_s','chi2_s','chi3_s','chi4_s','chi5_s','class_id','SS_Classname','FXN','count','incidence','percentage','ca_distance_mean',
-'ca_distance_std','torsion_length_mean','torsion_length_std','energy_mean','energy_std', 'ss_id']
+merge_cols = [
+    "chi1_s",
+    "chi2_s",
+    "chi3_s",
+    "chi4_s",
+    "chi5_s",
+    "class_id",
+    "SS_Classname",
+    "FXN",
+    "count",
+    "incidence",
+    "percentage",
+    "ca_distance_mean",
+    "ca_distance_std",
+    "torsion_length_mean",
+    "torsion_length_std",
+    "energy_mean",
+    "energy_std",
+    "ss_id",
+]
 
-class DisulfideClass_Constructor():
-    '''
+
+class DisulfideClass_Constructor:
+    """
     This Class manages structural classes for the disulfide bonds contained
     in the proteusPy disulfide database.
 
     Build the internal dictionary mapping disulfides to class names.
 
-    Disulfide binary classes are defined using the ± formalism described by 
-    Schmidt et al. (Biochem, 2006, 45, 7429-7433), across all 32 (2^5), possible 
+    Disulfide binary classes are defined using the ± formalism described by
+    Schmidt et al. (Biochem, 2006, 45, 7429-7433), across all 32 (2^5), possible
     binary sidechain torsional combinations. Classes are named per Schmidt's convention.
     The ``class_id`` represents the sign of each dihedral angle $\chi_{1} - \chi_{1'}$
     where *0* repreents *negative* dihedral angle and *2* a *positive* angle.
@@ -77,8 +102,8 @@ class DisulfideClass_Constructor():
     |      22202 | +RHHook        | UNK        |     593 |  0.00491313 |
     |      22220 | ±RHSpiral      | UNK        |    2544 |  0.0210776  |
     |      22222 | +RHSpiral      | UNK        |    3665 |  0.0303653  |
-    
-    '''
+
+    """
 
[…]
