+++ .pdocconfig.py
@@ -1,3 +1,1 @@
-pdoc_config = {
-    "exclude": ["proteusPy/__version__.py"] 
-}
+pdoc_config = {"exclude": ["proteusPy/__version__.py"]}
