+++ proteusPy/DisulfideClasses.py
@@ -1,4 +1,4 @@
-'''
+"""
 Functions to create Disulfide Bond structural classes based on
 dihedral angle rules. This module is part of the proteusPy package.
 Many of the plotting functions have been folded into the DisulfideClassConstructor
@@ -10,7 +10,7 @@ Author: Eric G. Suchanek, PhD.
 License: BSD
 Last Modification: 2/19/24 -egs-
 
-'''
+"""
 
 from io import StringIO
 import pandas
@@ -24,13 +24,14 @@ from proteusPy.data import SS_CLASS_DEFINITIONS
 from proteusPy.angle_annotation import AngleAnnotation
 from proteusPy.ProteusGlobals import DPI
 
+
 def create_classes(df):
     """
     Group the DataFrame by the sign of the chi columns and create a new class ID column for each unique grouping.
 
     :param df: A pandas DataFrame containing columns 'ss_id', 'chi1', 'chi2', 'chi3', 'chi4', 'chi5', 'ca_distance', 'cb_distance', 'torsion_length', and 'energy'.
     :return: A pandas DataFrame containing columns 'class_id', 'ss_id', and 'count', where 'class_id' is a unique identifier for each grouping of chi signs, 'ss_id' is a list of all 'ss_id' values in that grouping, and 'count' is the number of rows in that grouping.
-    
+
     Example:
     >>> import pandas as pd
     >>> df = pandas.DataFrame({
@@ -55,21 +56,24 @@ def create_classes(df):
 
     """
     # Create new columns with the sign of each chi column
-    chi_columns = ['chi1', 'chi2', 'chi3', 'chi4', 'chi5']
-    sign_columns = [col + '_s' for col in chi_columns]
+    chi_columns = ["chi1", "chi2", "chi3", "chi4", "chi5"]
+    sign_columns = [col + "_s" for col in chi_columns]
     df[sign_columns] = df[chi_columns].applymap(lambda x: 1 if x >= 0 else -1)
-    
+
     # Create a new column with the class ID for each row
-    class_id_column = 'class_id'
-    df[class_id_column] = (df[sign_columns] + 1).apply(lambda x: ''.join(x.astype(str)), axis=1)
+    class_id_column = "class_id"
+    df[class_id_column] = (df[sign_columns] + 1).apply(
+        lambda x: "".join(x.astype(str)), axis=1
+    )
 
     # Group the DataFrame by the class ID and return the grouped data
-    grouped = df.groupby(class_id_column)['ss_id'].unique().reset_index()
-    grouped['count'] = grouped['ss_id'].apply(lambda x: len(x))
-    grouped['incidence'] = grouped['ss_id'].apply(lambda x: len(x)/len(df))
-    grouped['percentage'] = grouped['incidence'].apply(lambda x: 100 * x)
+    grouped = df.groupby(class_id_column)["ss_id"].unique().reset_index()
+    grouped["count"] = grouped["ss_id"].apply(lambda x: len(x))
+    grouped["incidence"] = grouped["ss_id"].apply(lambda x: len(x) / len(df))
+    grouped["percentage"] = grouped["incidence"].apply(lambda x: 100 * x)
 
     return grouped
+
 
 def angle_within_range(angle, min_angle, max_angle):
     """
@@ -84,6 +88,7 @@ def angle_within_range(angle, min_angle, max_angle):
         bool: True if the angle is within the range, False otherwise.
     """
     import math
+
     # Convert angles to radians
     angle_rad = math.radians(angle)
     min_angle_rad = math.radians(min_angle)
@@ -94,6 +99,7 @@ def angle_within_range(angle, min_angle, max_angle):
         return True
     else:
         return False
+
 
 def get_quadrant(angle_deg):
     """
[…]
