+++ proteusPy/turtle3D.py
@@ -1,15 +1,15 @@
-'''
+"""
 Implementation of a 3D 'Turtle' in Python.
 
 Part of the program proteusPy, https://github.com/suchanek/proteusPy, 
 a Python packages for the manipulation and analysis of macromolecules. 
 Based on the C implementation originally authored by Eric G. Suchanek PhD, 1990.
 
-'''
+"""
 
 # Last modification 7/15/23 -egs-
 
-__pdoc__ = {'__all__': True}
+__pdoc__ = {"__all__": True}
 
 import numpy
 import math
@@ -25,6 +25,7 @@ _ORIENTATION_INIT = -1
 
 ORIENT_BACKBONE = 2
 ORIENT_SIDECHAIN = 1
+
 
 # Class Definition begins
 class Turtle3D:
@@ -45,26 +46,31 @@ class Turtle3D:
 
         self._name = name
         self._pen = _UP_
-        self._orientation = _ORIENTATION_INIT # will be set to 1 or 2 later when used for residue building
+        self._orientation = _ORIENTATION_INIT  # will be set to 1 or 2 later when used for residue building
         self._recording = False
         self._tape = []
-    
-    def new(self, name: str, pos: Vector, 
-            head=Vector(1,0,0), 
-            left=Vector(0,1,0), 
-            up=Vector(0,0,1), 
-            pen='up', recording=False):
-        '''
+
+    def new(
+        self,
+        name: str,
+        pos: Vector,
+        head=Vector(1, 0, 0),
+        left=Vector(0, 1, 0),
+        up=Vector(0, 0, 1),
+        pen="up",
+        recording=False,
+    ):
+        """
         Initialize a Turtle with a name and initial position
 
         :param name: Turtle's Name
-        :param pos: Turtle's Position 
+        :param pos: Turtle's Position
         :param head: Turtle's Heading vector, defaults to Vector(1,0,0)
         :param left: Turtle's Left vector, defaults to Vector(0,1,0)
         :param up: Turtle's Up vector, defaults to Vector(0,0,1)
         :param pen: Pen state, defaults to 'up'
         :param recording: _description_, defaults to False
-        '''
+        """
         self.Name = name
         self.Position = pos
         self.Heading = head
@@ -74,11 +80,11 @@ class Turtle3D:
         self.Recording = recording
 
     def copy_coords(self, source):
-        '''
-        Copy the Position, Heading, Left and Up coordinate system from 
+        """
+        Copy the Position, Heading, Left and Up coordinate system from
         the input source into self. Argument: source: Turtle3D
         Returns: None
-        '''
+        """
 
         # copy the Arrays
         self._position = source._position.copy()
@@ -101,33 +107,34 @@ class Turtle3D:
         Returns: None
         """
         self.__init__()
-    
+
     @property
     def Orientation(self):
         return self._orientation
-    
+
     @Orientation.setter
     def Orientation(self, orientation):
-        assert orientation == ORIENT_BACKBONE or orientation == ORIENT_SIDECHAIN, \
-         f'Orientation must be {ORIENT_BACKBONE} or {ORIENT_SIDECHAIN}'
+        assert (
+            orientation == ORIENT_BACKBONE or orientation == ORIENT_SIDECHAIN
+        ), f"Orientation must be {ORIENT_BACKBONE} or {ORIENT_SIDECHAIN}"
         self._orientation = orientation
-    
+
     @property
     def Pen(self):
         if self._pen == _UP_:
-            return('up')
+            return "up"
         else:
-            return('down')
-    
+            return "down"
+
     @Pen.setter
     def Pen(self, pen):
-        if pen == 'up':
+        if pen == "up":
[…]
