+++ programs/DisulfideClass_Analysis.py
@@ -99,9 +99,7 @@ from proteusPy.Disulfide import Disulfide
 SAVE_DIR = "/Users/egs/Documents/proteusPy/"
 
 
-def analyze_six_classes(
-    loader, do_graph=True, do_consensus=True, cutoff=0.1
-) -> DisulfideList:
+def analyze_six_classes(loader, do_graph=True, do_consensus=True, cutoff=0.1) -> DisulfideList:
     """
     Analyze the six classes of disulfide bonds.
 
@@ -147,9 +145,7 @@ def analyze_six_classes(
             loader.SSList.remove(_ss)
 
         if do_graph:
-            class_disulfides.display_torsion_statistics(
-                display=False, save=True, fname=fname, light=True, stats=False
-            )
+            class_disulfides.display_torsion_statistics(display=False, save=True, fname=fname, light=True, stats=False)
 
         if do_consensus:
             # get the average conformation - array of dihedrals
@@ -171,18 +167,14 @@ def analyze_six_classes(
             res_list.append(exemplar)
 
     if do_consensus:
-        print(
-            f"--> analyze_six_classes(): Writing consensus structures to: {class_filename}"
-        )
+        print(f"--> analyze_six_classes(): Writing consensus structures to: {class_filename}")
         with open(class_filename, "wb+") as f:
             pickle.dump(res_list, f)
 
     return res_list
 
 
-def analyze_binary_classes(
-    loader, do_graph=True, do_consensus=True, cutoff=0.1
-) -> DisulfideList:
+def analyze_binary_classes(loader, do_graph=True, do_consensus=True, cutoff=0.1) -> DisulfideList:
     """
     Analyze the binary classes of disulfide bonds.
 
@@ -228,17 +220,13 @@ def analyze_binary_classes(
             # loader.SSList.remove(_ss)
 
         if do_graph:
-            class_disulfides.display_torsion_statistics(
-                display=False, save=True, fname=fname, light=True, stats=False
-            )
+            class_disulfides.display_torsion_statistics(display=False, save=True, fname=fname, light=True, stats=False)
 
         if do_consensus:
             # get the average conformation - array of dihedrals
             avg_conformation = np.zeros(5)
 
-            print(
-                f"--> analyze_binary_classes(): Computing avg conformation for: {cls}"
-            )
+            print(f"--> analyze_binary_classes(): Computing avg conformation for: {cls}")
             avg_conformation = class_disulfides.Average_Conformation
 
             # build the average disulfide for the class
@@ -254,9 +242,7 @@ def analyze_binary_classes(
             res_list.append(exemplar)
 
     if do_consensus:
-        print(
-            f"--> analyze_binary_classes(): Writing consensus structures to: {class_filename}"
-        )
+        print(f"--> analyze_binary_classes(): Writing consensus structures to: {class_filename}")
         with open(class_filename, "wb+") as f:
             pickle.dump(res_list, f)
 
@@ -281,9 +267,7 @@ def plot_classes_vs_cutoff(cutoff, steps):
         tot = class_df["percentage"].sum()
         tot_list.append(tot)
         members_list.append(class_df.shape[0])
-        print(
-            f"Cutoff: {c:5.3} accounts for {tot:7.2f}% and is {class_df.shape[0]:5} members long."
-        )
+        print(f"Cutoff: {c:5.3} accounts for {tot:7.2f}% and is {class_df.shape[0]:5} members long.")
 
     fig, ax1 = plt.subplots()
 
[…]
