+++ .github/workflows/publish-to-pypi.yml
@@ -1,34 +1,0 @@
-name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
-permissions:
-  id-token: write
-  
-on: push
-jobs:
-  build-n-publish:
-    name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@master
-      - name: Set up Python 3.10
-        uses: actions/setup-python@v3
-        with:
-          python-version: "3.10"
-      - name: Install pypa/build
-        run: >-
-          python -m
-          pip install
-          build
-          --user
-      - name: Install proteusPy
-        run: >-
-          python -m
-          pip install .
-      - name: Build a binary wheel and a source tarball
-        run: >-
-          python -m
-          setup sdist
-      - name: Publish package to PyPI
-        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-        uses: pypa/gh-action-pypi-publish@release/v1
-
-
