+++ README.md
@@ -32,7 +32,6 @@ It's simplest to clone the repo via GitHub since it contains all of the notebook
 - From a shell prompt while sitting in your repo dir:
   ```console
   $ git clone https://github.com/suchanek/proteusPy.git
-  $ git clone https://github.com/suchanek/biopython.git
   $ cd proteusPy
   $ git-lfs track "*.csv" "*.mp4" "*.pkl"
   $ make pkg
@@ -57,7 +56,6 @@ Installing in Windows is a bit more involved than Linux or Macos. Sadly, the pac
 - Open a Miniforge prompt and cd into your repo dir:
   ```console
   (base) C:\Users\egs\repos> git clone https://github.com/suchanek/proteusPy.git
-  (base) C:\Users\egs\repos> git clone https://github.com/suchanek/biopython.git
   (base) C:\Users\egs\repos> cd proteusPy
   (base) C:\Users\egs\repos\proteuspy> git-lfs track "*.csv" "*.mp4" "*.pkl"
   (base) C:\Users\egs\repos\proteuspy> make pkg
