### 2024-03-05 #6169 @JoaoRodrigues
Will get to it this week - was waiting for the previous rounds to finish not to overlap/overwhelm @suchanek. Incidentally, I'm also reviewing his PR in Biopython, which I'm pretty sure relates to this work :)
