### 2024-03-05 #6169 @suchanek
Thank you João, you are correct. My PR is for this. Cheers!Eric G. Suchanek, PhD Astronomer at largeMonterey Institute for Research in Astronomy200 Eighth Street, Marina CA 93933(831-883-1000) Fax: 883-1031Dedicated to Research and Education in Astronomy for 50 Years.On Mar 5, 2024, at 11:54 AM, João Rodrigues ***@***.***> wrote:﻿
Will get to it this week - was waiting for the previous rounds to finish not to overlap/overwhelm @suchanek. Incidentally, I'm also reviewing his PR in Biopython, which I'm pretty sure relates to this work :)
