+++ proteusPy/Residue.py
@@ -24,7 +24,9 @@ def build_residue(turtle: Turtle3D):
     Returns: <Vector> n, ca, cb, c
     """
 
-    assert turtle._orientation == 2, f"build_residue() requires Turtle3D to be in orientation #2"
+    assert (
+        turtle._orientation == 2
+    ), f"build_residue() requires Turtle3D to be in orientation #2"
 
     # canonical internal coordinates for backbone atoms with Turtle3D at Ca, heading towards C with N on the left
     # AKA Orientation #2
@@ -52,7 +54,9 @@ def get_backbone_from_chain(chain, resnumb):
     """
     residue = chain[resnumb]
 
-    assert residue is not None, f"get_backbone_from_sidechain() requires valid residue number"
+    assert (
+        residue is not None
+    ), f"get_backbone_from_sidechain() requires valid residue number"
 
     # proximal residue
     n = residue["N"].get_vector()
