+++ Oppy.yml
@@ -1,20 +1,0 @@
-name: proteusPy
-channels:
-  - conda-forge
-dependencies:
-  - python==3.11.7
-  - traitlets=5.9.0
-  - notebook=6.4.12
-  - pandas<2.0
-  - matplotlib
-  - nodejs
-  - pip
-  - pip:
-    - tqdm
-    - ipykernel
-    - jupyterlab>=3 
-    - ipywidgets 
-    - pyvista[all,trame]
-    - jupyter_contrib_nbextensions
-    - kaleido
-
