+++ pyproject.toml
@@ -1,14 +1,13 @@
-[tool.poetry]
+[project]
 name = "proteuspy"
 version = "0.92.14"
 description = "proteusPy - Protein Structure Analysis and Modeling Tools"
-authors = ["Eric Suchanek <suchanek@mac.com>"]
-license = "BSD"
+authors = [{name = "Eric Suchanek", email = "suchanek@mac.com"}]
+license = {file = "LICENSE"}
 readme = "README.md"
-repository = "https://github.com/suchanek/proteusPy/"
-homepage = "https://github.com/suchanek/proteusPy/"
-documentation = "https://suchanek.github.io/proteusPy/"
 keywords = ["structural biophysics", "proteus", "disulfide bonds", "RCSB"]
+requires-python = "<3.12"
+
 classifiers=[
         "Development Status :: 4 - Beta",
         "License :: OSI Approved :: BSD License",
@@ -22,49 +21,56 @@ classifiers=[
         "Programming Language :: Python :: 3.11",
         ]
 
+dependencies = [
+"numpy",
+"matplotlib",
+"pandas<2.0",
+"pyvista[all,trame]",
+"traitlets==5.9.0",
+"notebook==6.4.12",
+"jupyter",
+"jupyterlab>=3",
+"seaborn",
+"pillow",
+"tqdm",
+"plotly",
+"datetime",
+"jupyter_bokeh",
+"openai",
+"panel",
+"scikit-learn",
+"gdown",
+"ipykernel",
+"ipygany",
+"nodejs",
+"pytube",
+"grpcio",
+"pip",
+"wget",
+"vtk",
+"kaleido",
+"plotly_express",
+"trame-jupyter-extension",
+"jupyter_contrib_nbextensions",
+"ipywidgets",
+"imageio[ffmpeg]",
+"biopython @ git+https://github.com/suchanek/biopython.git@egs_ssbond_240305",
+]
 
-packages = [{include = "proteusPy"}]
+[tool.setuptools.packages.find]
+where = ["proteusPy"]
 
-[tool.poetry.dependencies]
-python = "^3.11"
-numpy = "^1.26.4"
-kaleido = "0.2.1"
-pandas = "^1.0.0"
-notebook = "6.4.12"
-jupyter = "^1.0.0"
-jupyterlab = ">=3"
-seaborn = "^0.13.2"
-pillow = "^10.2.0"
-tqdm = "^4.66.2"
-plotly = "^5.19.0"
-DateTime = "^5.4"
-jupyter_bokeh = "^4.0.0"
-openai = "^1.13.3"
-panel = "^1.3.8"
-scikit-learn = "^1.4.1.post1"
-gdown = "^5.1.0"
-ipykernel = "^6.29.3"
-ipygany = "^0.5.0"
-nodejs = "^0.1.1"
-pytube = "^15.0.0"
-grpcio = "^1.62.0"
-pip = "^24.0"
-wget = "^3.2"
-vtk = "^9.3.0"
-plotly-express = "^0.4.1"
-trame-jupyter-extension = "^2.0.1"
-jupyter-contrib-nbextensions = "^0.7.0"
-ipywidgets = "^8.1.2"
-imageio = {extras = ["ffmpeg"], version = "^2.34.0"}
-poethepoet = "^0.25.0"
-pyvista = {extras = ["trame"], version = "^0.43.3"}
-biopython = {git = "https://github.com/suchanek/biopython.git"}
+[tool.setuptools.package-data]
+proteusPy = ["data/*.txt", "data/*.py", "data/*.json", "data/*.csv", "data/SS_consensus_class_sext.pkl"]
 
-[tool.poetry.dev-dependencies]
-pytest = "^6.2.1"
[…]
