+++ makefile
@@ -2,15 +2,15 @@
 # Author: Eric G. Suchanek, PhD
 # Last revision: 3/4/24 -egs-
 
-VERS := $(shell grep ^__version__ proteusPy/__version__.py | cut -d= -f2 | tr -d \" | sed 's/^[[:space:]]*//')
+VERS := $(shell grep ^0 VERSION | cut -d= -f2 | tr -d \" | sed 's/^[[:space:]]*//')
 
 PYPI_PASSWORD := $(shell echo $$PYPI_PASSWORD)
 CONDA = mamba
 
-MESS = "disulfide module work, docs"
+MESS = "cleanup"
 
 DEVNAME = ppydev
-OUTFILES = sdist.out, bdist.out, docs.out
+OUTFILES = sdist.out, bdist.out, docs.out tag.out
 
 PHONY = .
 
@@ -39,7 +39,6 @@ devclean:
 install:
 	@echo "Starting installation step 2/2 for $(VERS)..."
 	pip install . 
-	# && cd ../biopython && pip install .
 	jupyter contrib nbextension install --sys-prefix
 	jupyter nbextension enable --py --sys-prefix widgetsnbextension
 	python -m ipykernel install --user --name proteusPy --display-name "proteusPy $(VERS)"
@@ -80,24 +79,19 @@ format: sdist
 
 build: sdist docs format
 
-sdist: .
-	python setup.py sdist
+sdist: tag
+	python -m build
 	@echo $(VERS) > sdist.out
-
-bdist: sdist
-	@python setup.py bdist
-	@echo $(VERS) > bdist.out
 
 docs: sdist
 	@pdoc -o docs --math --logo "./logo.png" ./proteusPy
 	@echo $(VERS) > docs.out
 
 # normally i push to PyPi via github action
-upload: .
-	# @poetry publish --build --username "__token__" --password $(PYPI_PASSWORD)
+upload: sdist
 	twine upload dist/*.gz
-tag: sdist
-	@git tag -a $(VERS) -m $(VERS)
+tag: .
+	@git tag -a $(VERS) -m $(MESS)
 	@echo $(VERS) > tag.out
 
 commit:
