+++ proteusPy/DisulfideClasses.py
@@ -124,7 +124,9 @@ def get_quadrant(angle_deg):
     elif angle_deg >= -90 and angle_deg < 0:
         return str(4)
     else:
-        raise ValueError("Invalid angle value: angle must be in the range [-180, 180).")
+        raise ValueError(
+            "Invalid angle value: angle must be in the range [-180, 180)."
+        )
 
 
 def torsion_to_sixclass(tors):
@@ -167,7 +169,9 @@ def get_sixth_quadrant(angle_deg):
     elif angle_deg >= 300 and angle_deg < 360:
         return str(1)
     else:
-        raise ValueError("Invalid angle value: angle must be in the range [-360, 360).")
+        raise ValueError(
+            "Invalid angle value: angle must be in the range [-360, 360)."
+        )
 
 
 def get_half_quadrant(angle_deg):
@@ -198,7 +202,9 @@ def get_half_quadrant(angle_deg):
     elif angle_deg >= -180 and angle_deg < -135:
         return str(8)
     else:
-        raise ValueError("Invalid angle value: angle must be in the range [-180, 180).")
+        raise ValueError(
+            "Invalid angle value: angle must be in the range [-180, 180)."
+        )
 
 
 def create_quat_classes(df):
@@ -412,7 +418,9 @@ def plot_class_chart(classes: int) -> None:
     # Show the chart
 
 
-def plot_count_vs_class_df(df, title="title", theme="light", save=False, savedir="."):
+def plot_count_vs_class_df(
+    df, title="title", theme="light", save=False, savedir="."
+):
     """
     Plots a line graph of count vs class ID using Plotly.
 
@@ -472,7 +480,10 @@ def plot_count_vs_classid(df, cls=None, title="title", theme="light"):
         fig = px.line(subset, x="class_id", y="count", title=f"{title}")
 
     fig.update_layout(
-        xaxis_title="Class ID", yaxis_title="Count", showlegend=True, title_x=0.5
+        xaxis_title="Class ID",
+        yaxis_title="Count",
+        showlegend=True,
+        title_x=0.5,
     )
     fig.layout.autosize = True
 
