+++ programs/DisulfideLogo.py
@@ -33,8 +33,14 @@ if __name__ == "__main__":
 
     best = PDB_SS.get_by_name("2q7q_75D_140D")
 
-    best.screenshot(single=True, style="cpk", shadows=False, fname="img/logo_cpk.png")
-    best.screenshot(single=True, style="bs", shadows=False, fname="img/logo_bs.png")
-    best.screenshot(single=True, style="sb", shadows=False, fname="img/logo_sb.png")
+    best.screenshot(
+        single=True, style="cpk", shadows=False, fname="img/logo_cpk.png"
+    )
+    best.screenshot(
+        single=True, style="bs", shadows=False, fname="img/logo_bs.png"
+    )
+    best.screenshot(
+        single=True, style="sb", shadows=False, fname="img/logo_sb.png"
+    )
 
     exit()

[…]
