+++ programs/DBViewer_Working2.py
@@ -94,7 +94,9 @@ button = pn.widgets.Button(name="Refresh", button_type="primary")
 button.on_click(click_plot)
 
 styles_group = pn.widgets.RadioBoxGroup(
-    name="Rending Style", options=["Split Bonds", "CPK", "Ball and Stick"], inline=False
+    name="Rending Style",
+    options=["Split Bonds", "CPK", "Ball and Stick"],
+    inline=False,
 )
 
 single_checkbox = pn.widgets.Checkbox(name="Single View", value=True)
@@ -241,7 +243,9 @@ def render_ss(clk=True):
     update_info(ss)
     update_output(ss)
 
-    plotter = ss.plot(plotter, single=single, style=style, shadows=False, light=light)
+    plotter = ss.plot(
+        plotter, single=single, style=style, shadows=False, light=light
+    )
     vtkpan = pn.pane.VTK(
         plotter.ren_win,
         margin=0,
