+++ tests/test_turtle3d.py
@@ -1,6 +1,8 @@
 import unittest
 
+import numpy
 from Bio.PDB.vectors import Vector
+from numpy.testing import assert_array_equal
 
 from proteusPy.turtle3D import Turtle3D
 
@@ -9,52 +11,53 @@ class TestTurtle3D(unittest.TestCase):
 
     def test_init(self):
         turtle = Turtle3D()
-        self.assertEqual(turtle.position, Vector(0.0, 0.0, 0.0))
-        self.assertEqual(turtle.heading, Vector(1.0, 0.0, 0.0))
-        self.assertEqual(turtle.left, Vector(0.0, 1.0, 0.0))
-        self.assertEqual(turtle.up, Vector(0.0, 0.0, 1.0))
+        assert_array_equal(turtle._position, numpy.array((0.0, 0.0, 0.0)))
+        assert_array_equal(turtle._heading, numpy.array((1.0, 0.0, 0.0)))
+        assert_array_equal(turtle._left, numpy.array((0.0, 1.0, 0.0)))
+        assert_array_equal(turtle._up, numpy.array((0.0, 0.0, 1.0)))
 
     def test_new(self):
-        turtle = Turtle3D.new("test", Vector(1, 2, 3), Vector(4, 5, 6))
-        self.assertEqual(turtle.name, "test")
-        self.assertEqual(turtle.position, Vector(1, 2, 3))
-        self.assertEqual(turtle.heading, Vector(4, 5, 6))
+        turtle = Turtle3D()
+        turtle.new("test", Vector((1.0, 2.0, 3.0)), Vector((4.0, 5.0, 6.0)))
+        assert turtle._name == "test"
+        assert_array_equal(turtle._position, numpy.array((1.0, 2.0, 3.0)))
+        assert_array_equal(turtle._heading, numpy.array((4.0, 5.0, 6.0)))
 
     def test_forward(self):
         turtle = Turtle3D()
         turtle.move(10)
-        self.assertEqual(turtle.position, Vector(10, 0, 0))
+        assert_array_equal(turtle._position, numpy.array((10.0, 0.0, 0.0)))
 
     def test_back(self):
         turtle = Turtle3D()
         turtle.move(10)
         turtle.move(-5)
-        self.assertEqual(turtle.position, Vector(5, 0, 0))
+        assert_array_equal(turtle._position, numpy.array((5.0, 0.0, 0.0)))
 
     def test_right(self):
         turtle = Turtle3D()
         turtle.turn(90)
-        self.assertEqual(turtle.heading, Vector(0, 1, 0))
+        assert_array_equal(turtle._heading, numpy.array((0.0, 1.0, 0.0)))
 
     def test_left(self):
         turtle = Turtle3D()
         turtle.turn(-90)
-        self.assertEqual(turtle.heading, Vector(0, -1, 0))
+        assert_array_equal(turtle._heading, numpy.array((0.0, -1.0, 0.0)))
 
     def test_roll(self):
         turtle = Turtle3D()
         turtle.roll(90)
-        self.assertEqual(turtle.left, Vector(0, 0, 1))
-        self.assertEqual(turtle.up, Vector(-1, 0, 0))
+        assert_array_equal(turtle._left, numpy.array((0.0, 0.0, 1.0)))
+        assert_array_equal(turtle._up, numpy.array((-1.0, 0.0, 0.0)))
 
     def test_pitch(self):
         turtle = Turtle3D()
         turtle.pitch(90)
-        self.assertEqual(turtle.Heading, Vector(0, 0, -1))
-        self.assertEqual(turtle.Up, Vector(0, 1, 0))
[…]
