+++ makefile
@@ -48,7 +48,7 @@ install:
 install_dev:
 	@echo "Starting installation step 2/2 for $(VERS)..."
 	pip install .
-	pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=proteusPy
+	pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
 	pip install pdoc twine black
 	jupyter contrib nbextension install --sys-prefix
 	jupyter nbextension enable --py --sys-prefix widgetsnbextension
