+++ tests/test_turtle3d.py
@@ -1,63 +1,85 @@
 import unittest
 
-import numpy
+import numpy as np
 from Bio.PDB.vectors import Vector
 from numpy.testing import assert_array_equal
 
 from proteusPy.turtle3D import Turtle3D
+
+# Define a _tolerance threshold
+_tolerance = 1e-8
+
+
+def cmp_vec(v1: Vector, v2: Vector, tol: float) -> bool:
+    "Return true if the length of the difference between the two vectors is less than a tolerance."
+    _diff = v2 - v1
+    _len = _diff.norm()
+    return _len < tol
 
 
 class TestTurtle3D(unittest.TestCase):
 
     def test_init(self):
         turtle = Turtle3D()
-        assert_array_equal(turtle._position, numpy.array((0.0, 0.0, 0.0)))
-        assert_array_equal(turtle._heading, numpy.array((1.0, 0.0, 0.0)))
-        assert_array_equal(turtle._left, numpy.array((0.0, 1.0, 0.0)))
-        assert_array_equal(turtle._up, numpy.array((0.0, 0.0, 1.0)))
+        assert_array_equal(turtle._position, np.array((0.0, 0.0, 0.0)))
+        assert_array_equal(turtle._heading, np.array((1.0, 0.0, 0.0)))
+        assert_array_equal(turtle._left, np.array((0.0, 1.0, 0.0)))
+        assert_array_equal(turtle._up, np.array((0.0, 0.0, 1.0)))
 
     def test_new(self):
+        _pos = Vector(1.0, 2.0, 3.0)
+        _head = Vector((4.0, 5.0, 6.0))
         turtle = Turtle3D()
-        turtle.new("test", Vector((1.0, 2.0, 3.0)), Vector((4.0, 5.0, 6.0)))
+        turtle.new("test", _pos, _head)
         assert turtle._name == "test"
-        assert_array_equal(turtle._position, numpy.array((1.0, 2.0, 3.0)))
-        assert_array_equal(turtle._heading, numpy.array((4.0, 5.0, 6.0)))
+        assert cmp_vec(turtle.heading, _head, _tolerance)
+        assert cmp_vec(turtle.position, _pos, _tolerance)
 
     def test_forward(self):
+        _pos = Vector(10.0, 0.0, 0.0)
         turtle = Turtle3D()
-        turtle.move(10)
-        assert_array_equal(turtle._position, numpy.array((10.0, 0.0, 0.0)))
+        turtle.move(10.0)
+        assert cmp_vec(turtle.position, _pos, _tolerance)
 
     def test_back(self):
+        _pos = Vector(5.0, 0.0, 0.0)
         turtle = Turtle3D()
-        turtle.move(10)
-        turtle.move(-5)
-        assert_array_equal(turtle._position, numpy.array((5.0, 0.0, 0.0)))
+        turtle.move(10.0)
+        turtle.move(-5.0)
+        assert cmp_vec(turtle.position, _pos, _tolerance)
 
     def test_right(self):
+        _head = Vector(0.0, 1.0, 0.0)
         turtle = Turtle3D()
-        turtle.turn(90)
-        assert_array_equal(turtle._heading, numpy.array((0.0, 1.0, 0.0)))
+        turtle.turn(90.0)
+        _h2 = turtle.heading
+        assert cmp_vec(_h2, _head, _tolerance)
 
     def test_left(self):
+        _head = Vector(0.0, -1.0, 0.0)
         turtle = Turtle3D()
-        turtle.turn(-90)
-        assert_array_equal(turtle._heading, numpy.array((0.0, -1.0, 0.0)))
+        turtle.turn(-90.0)
[…]
