+++ pyproject.toml
@@ -69,7 +69,7 @@ enabled = true
 version_file = "VERSION"
 count_commits_from_version_file = true
 template = "{tag}" # default setting
-dev_template = "{tag}"
+dev_template = "{tag}+{branch}.{ccount}"
 dirty_template = "{tag}+{branch}.{ccount}"
 
 [project.urls]
