+++ proteusPy/turtle3D.py
@@ -58,7 +58,7 @@ class Turtle3D:
         left: Vector = (0.0, 1.0, 0.0),
         up: Vector = (0.0, 0.0, 1.0),
         recording=False,
-    ):
+    ) -> None:
         """
         Initialize a Turtle with a name and initial position
 
@@ -77,7 +77,7 @@ class Turtle3D:
         self.up = up
         self._recording = recording
 
-    def copy_coords(self, source):
+    def copy_coords(self, source) -> None:
         """
         Copy the Position, Heading, Left and Up coordinate system from
         the input source into self. Argument: source: Turtle3D
@@ -98,7 +98,7 @@ class Turtle3D:
 
         self._orientation = source._orientation
 
-    def reset(self):
+    def reset(self) -> None:
         """
         Reset the Turtle to be at the Origin, with correct Heading, Left and Up vectors.
         Arguments: None
@@ -107,7 +107,7 @@ class Turtle3D:
         self.__init__()
 
     @property
-    def Orientation(self):
+    def Orientation(self) -> numpy.array:
         return self._orientation
 
     @Orientation.setter
@@ -118,14 +118,14 @@ class Turtle3D:
         self._orientation = orientation
 
     @property
-    def Pen(self):
+    def Pen(self) -> str:
         if self._pen == _UP_:
             return "up"
         else:
             return "down"
 
     @Pen.setter
-    def Pen(self, pen):
+    def Pen(self, pen) -> None:
         if pen == "up":
             self._pen = _UP_
         elif pen == "down":
@@ -134,14 +134,14 @@ class Turtle3D:
             self._pen = _DOWN_
 
     @property
-    def Recording(self):
+    def Recording(self) -> bool:
         return self._recording
 
     @Recording.setter
     def Recording(self, recording: bool):
         self._recording = recording
 
-    def ResetTape(self):
+    def ResetTape(self) -> None:
         self.Recording(False)
         self._tape = []
 
@@ -156,7 +156,7 @@ class Turtle3D:
         return self.position
 
     @Position.setter
-    def Position(self, x, y=None, z=None):
+    def Position(self, x, y=None, z=None) -> None:
         """
         Set's the Turtle's Position
 
@@ -183,25 +183,8 @@ class Turtle3D:
         self.position = Vector(self._position)
         return
 
-    @Position.setter
-    def Position(self, pos: Vector):
-        """
-        Set's the Turtle's Position with Vector input
-
-        :param pos: Position vector
-        :type x: Vector
-        :raises ValueError: Type error
-        """
-        if not isinstance(pos, Vector):
-            raise ValueError("Turtle3d: pos is not a Vector")
-
-        self._position = pos._ar
-        self.position = pos
-
-        return
-
     @property
-    def Heading(self):
+    def Heading(self) -> Vector:
         """
         Get the Turtle's Heading
 
@@ -211,7 +194,7 @@ class Turtle3D:
         return self.heading
 
     @Heading.setter
-    def Heading(self, x, y=None, z=None):
+    def Heading(self, x, y=None, z=None) -> None:
         """
         Set the turtle's Heading direction vector
 
@@ -238,7 +221,7 @@ class Turtle3D:
         return
 
     @property
-    def Left(self):
+    def Left(self) -> Vector:
         """
         Get the Turtle's Left direction vector
 
@@ -275,7 +258,7 @@ class Turtle3D:
         return
 
     @property
-    def Up(self):
+    def Up(self) -> Vector:
         """
         The Turtle's Up direction vector
 
[…]
