+++ pyproject.toml
@@ -55,6 +55,18 @@ dependencies = [
 "imageio[ffmpeg]",
 ]
 
+[project.optional-dependencies]
+dev = [
+    "pdoc",
+    "black",
+]
+test = [
+    "pytest-cov ~=3.0.0",
+]
+
+[tool.coverage.run]
+source = ["proteusPy"]
+
 [tool.setuptools.packages.find]
 where = ["proteusPy"]
 
@@ -115,5 +127,5 @@ install = ["_install", "_jup_install"]
   deps = ["_clean_docs"]
 
 [build-system]
-requires = ["setuptools >= 61.0", "setuptools-git-versioning<2"]
+requires = ["setuptools >= 61.0", "setuptools-git-versioning<2", "poethepoet"]
 build-backend = "setuptools.build_meta"
