### 2024-03-10 f9ed298
cleanup makefile
