+++ pyproject.toml
@@ -77,7 +77,7 @@ proteusPy = ["data/*.txt", "data/*.py", "data/*.json", "data/*.csv", "data/SS_co
 proteusPy = ["data/PDB_all_ss.pkl", "data/PDB_SS_ALL_LOADER.pkl", "data/PDB_all_ss_dict.pkl"]
 
 [tool.setuptools-git-versioning]
-enabled = true
+enabled = false
 version_file = "VERSION"
 count_commits_from_version_file = true
 template = "{tag}" # default setting
