+++ pyproject.toml
@@ -6,8 +6,8 @@ license = {file = "LICENSE"}
 readme = "README.md"
 keywords = ["structural biophysics", "proteus", "disulfide bonds", "RCSB"]
 requires-python = "<3.12"
-dynamic = ["version"]
-
+dynamic = []
+version = "0.92.20"
 classifiers=[
         "Development Status :: 4 - Beta",
         "License :: OSI Approved :: BSD License",
