+++ proteusPy/turtle3D.py
@@ -53,14 +53,14 @@ class Turtle3D:
     def new(
         self,
         name: str,
-        pos: Vector = (0.0, 0.0, 0.0),
-        head: Vector = (1.0, 0.0, 0.0),
-        left: Vector = (0.0, 1.0, 0.0),
-        up: Vector = (0.0, 0.0, 1.0),
+        pos: Vector,
+        head: Vector = Vector(1.0, 0.0, 0.0),
+        left: Vector = Vector(0.0, 1.0, 0.0),
+        up: Vector = Vector(0.0, 0.0, 1.0),
         recording=False,
     ) -> None:
         """
-        Initialize a Turtle with a name and initial position
+        Initialize a Turtle with a name, position, optionally heading and left.
 
         :param name: Turtle's Name
         :param pos: Turtle's Position
@@ -72,9 +72,15 @@ class Turtle3D:
         """
         self._name = name
         self.pos = pos
-        self.h = head
-        self.l = left
-        self.u = up
+        self.h = head.norm()
+        self.l = left.norm()
+        self.u = up.norm()
+
+        self._position = pos._ar
+        self._left = self.l._ar
+        self._heading = self.h._ar
+        self._up = self.u._ar
+
         self._recording = recording
 
     def copy_coords(self, source) -> None:
@@ -91,10 +97,10 @@ class Turtle3D:
         self._up = source._up.copy()
 
         # copy the Vectors - create new ones from the source arrays
-        self.position = Vector(source._position)
-        self.heading = Vector(source._heading)
-        self.left = Vector(source._left)
-        self.up = Vector(source._up)
+        self.pos = Vector(source._position)
+        self.h = Vector(source._heading)
+        self.l = Vector(source._left)
+        self.u = Vector(source._up)
 
         self._orientation = source._orientation
 
@@ -212,17 +218,19 @@ class Turtle3D:
         :raises ValueError: illegal value
         """
 
-        if y is None and z is None:
+        if isinstance(x, Vector):
+            self.h = x
+            self._heading = x.get_array()
+
+        elif y is None and z is None:
             # Array, list, tuple...
             if len(x) != 3:
-                raise ValueError(
-                    "Turtle3D: x is not a list/tuple/array of 3 numbers"
-                )
+                raise ValueError("Turtle3D: x is not a list/tuple/array of 3 numbers")
             self._heading = numpy.array(x, "d")
         else:
             # Three numbers
             self._heading = numpy.array((x, y, z), "d")
-        self.heading = Vector(self._heading)
+        self.h = Vector(self._heading)
         return
 
     @property
@@ -234,7 +242,7 @@ class Turtle3D:
         :rtype: Vector
         """
 
-        return self.left
+        return self.l
 
     @Left.setter
     def Left(self, x, y=None, z=None):
@@ -249,17 +257,19 @@ class Turtle3D:
         :type z: float, optional
         :raises ValueError: illegal value
         """
-        if y is None and z is None:
+        if isinstance(x, Vector):
+            self.l = x
+            self._left = x.get_array()
+
+        elif y is None and z is None:
             # Array, list, tuple...
             if len(x) != 3:
-                raise ValueError(
-                    "Turtle3D: x is not a list/tuple/array of 3 numbers"
-                )
+                raise ValueError("Turtle3D: x is not a list/tuple/array of 3 numbers")
             self._left = numpy.array(x, "d")
         else:
             # Three numbers
             self._left = numpy.array((x, y, z), "d")
-        self.left = Vector(self._left)
+        self.l = Vector(self._left)
         return
 
     @property
@@ -271,7 +281,7 @@ class Turtle3D:
[…]

[…]
