+++ proteusPy/Disulfide.py
@@ -238,9 +238,7 @@ class Disulfide:
         self.chi3 = _ANG_INIT
         self.chi4 = _ANG_INIT
         self.chi5 = _ANG_INIT
-        self.rho = (
-            _ANG_INIT  # new dihedral angle: Nprox - Ca_prox - Ca_dist - N_dist
-        )
+        self.rho = _ANG_INIT  # new dihedral angle: Nprox - Ca_prox - Ca_dist - N_dist
 
         self.torsion_length = _FLOAT_INIT
 
@@ -1246,9 +1244,7 @@ class Disulfide:
 
         energy = 2.0 * (cos(torad(3.0 * chi1)) + cos(torad(3.0 * chi5)))
         energy += cos(torad(3.0 * chi2)) + cos(torad(3.0 * chi4))
-        energy += (
-            3.5 * cos(torad(2.0 * chi3)) + 0.6 * cos(torad(3.0 * chi3)) + 10.1
-        )
+        energy += 3.5 * cos(torad(2.0 * chi3)) + 0.6 * cos(torad(3.0 * chi3)) + 10.1
 
         self.energy = energy
         return energy
@@ -1550,10 +1546,7 @@ class Disulfide:
         prox_residue = chain1[prox]
         dist_residue = chain2[dist]
 
-        if (
-            prox_residue.get_resname() != "CYS"
-            or dist_residue.get_resname() != "CYS"
-        ):
+        if prox_residue.get_resname() != "CYS" or dist_residue.get_resname() != "CYS":
             print(
                 f"build_disulfide() requires CYS at both residues: {prox} {prox_residue.get_resname()} {dist} {dist_residue.get_resname()} Chain: {prox_residue.get_segid()}"
             )
@@ -1571,9 +1564,7 @@ class Disulfide:
         self.distal_residue_fullid = dist_residue.get_full_id()
 
         if quiet:
-            warnings.filterwarnings(
-                "ignore", category=DisulfideConstructionWarning
-            )
+            warnings.filterwarnings("ignore", category=DisulfideConstructionWarning)
         else:
             warnings.simplefilter("always")
 
@@ -1630,9 +1621,7 @@ class Disulfide:
 
         except Exception:
             mess = f"Missing coords for: {id} {prox-1} or {dist+1} for SS {proximal}-{distal}"
-            cprev_prox = nnext_prox = cprev_dist = nnext_dist = Vector(
-                -1.0, -1.0, -1.0
-            )
+            cprev_prox = nnext_prox = cprev_dist = nnext_dist = Vector(-1.0, -1.0, -1.0)
             self.missing_atoms = True
             warnings.warn(mess, DisulfideConstructionWarning)
 
@@ -1774,9 +1763,7 @@ class Disulfide:
              Unable to find residue: {resnumb} "
             raise DisulfideConstructionWarning(mess)
 
-    def make_movie(
-        self, style="sb", fname="ssbond.mp4", verbose=False, steps=360
-    ):
+    def make_movie(self, style="sb", fname="ssbond.mp4", verbose=False, steps=360):
         """
         Create an animation for ```self``` rotating one revolution about the Y axis,
         in the given ```style```, saving to ```filename```.
@@ -1966,9 +1953,7 @@ class Disulfide:
         return f"{self.repr_ss_chain_ids()}"
 
     def compute_rho(self):
-        self.rho = calc_dihedral(
-            self.n_prox, self.ca_prox, self.ca_dist, self.n_dist
-        )
+        self.rho = calc_dihedral(self.n_prox, self.ca_prox, self.ca_dist, self.n_dist)
         return self.rho
 
     def reset(self) -> None:
@@ -2473,9 +2458,7 @@ def Download_Disulfides(
     for entry in pbar:
[…]
