+++ proteusPy/DisulfideClass_Constructor.py
@@ -215,9 +215,7 @@ class DisulfideClass_Constructor:
         angle_maps = {"0": ["4", "5", "6"], "2": ["1", "2", "3"]}
         class_lists = [angle_maps[char] for char in class_str]
         class_combinations = itertools.product(*class_lists)
-        class_strings = [
-            "".join(combination) for combination in class_combinations
-        ]
+        class_strings = ["".join(combination) for combination in class_combinations]
         return class_strings
 
     def build_yourself(self, loader) -> None:
@@ -240,9 +238,7 @@ class DisulfideClass_Constructor:
         tors_df = loader.getTorsions()
 
         if self.verbose:
-            print(
-                f"-> DisulfideClass_Constructor(): creating binary SS classes..."
-            )
+            print(f"-> DisulfideClass_Constructor(): creating binary SS classes...")
 
         grouped = self.create_binary_classes(tors_df)
 
@@ -272,9 +268,7 @@ class DisulfideClass_Constructor:
         self.classdf = merged.copy()
 
         if self.verbose:
-            print(
-                f"-> DisulfideClass_Constructor(): creating sixfold SS classes..."
-            )
+            print(f"-> DisulfideClass_Constructor(): creating sixfold SS classes...")
 
         grouped_sixclass = self.create_six_classes(tors_df)
 
@@ -297,9 +291,7 @@ class DisulfideClass_Constructor:
         # Create new columns with the sign of each chi column
         chi_columns = ["chi1", "chi2", "chi3", "chi4", "chi5"]
         sign_columns = [col + "_s" for col in chi_columns]
-        df[sign_columns] = df[chi_columns].applymap(
-            lambda x: 1 if x >= 0 else -1
-        )
+        df[sign_columns] = df[chi_columns].applymap(lambda x: 1 if x >= 0 else -1)
 
         # Create a new column with the class ID for each row
         class_id_column = "class_id"
@@ -310,9 +302,7 @@ class DisulfideClass_Constructor:
         # Group the DataFrame by the class ID and return the grouped data
         grouped = df.groupby(class_id_column)["ss_id"].unique().reset_index()
         grouped["count"] = grouped["ss_id"].apply(lambda x: len(x))
-        grouped["incidence"] = grouped["ss_id"].apply(
-            lambda x: len(x) / len(df)
-        )
+        grouped["incidence"] = grouped["ss_id"].apply(lambda x: len(x) / len(df))
         grouped["percentage"] = grouped["incidence"].apply(lambda x: 100 * x)
 
         return grouped
@@ -341,9 +331,9 @@ class DisulfideClass_Constructor:
             _df[col_name + "_t"] = df[col_name].apply(self.get_sixth_quadrant)
 
         # create the class_id column
-        df["class_id"] = _df[
-            ["chi1_t", "chi2_t", "chi3_t", "chi4_t", "chi5_t"]
-        ].apply(lambda x: "".join(x), axis=1)
+        df["class_id"] = _df[["chi1_t", "chi2_t", "chi3_t", "chi4_t", "chi5_t"]].apply(
+            lambda x: "".join(x), axis=1
+        )
 
         # group the DataFrame by class_id and return the grouped data
         grouped = df.groupby("class_id").agg({"ss_id": "unique"})
@@ -500,9 +490,7 @@ class DisulfideClass_Constructor:
 
         # Helper function to draw angle easily.
         def plot_angle(ax, pos, angle, length=0.95, acol="C0", **kwargs):
-            vec2 = np.array(
-                [np.cos(np.deg2rad(angle)), np.sin(np.deg2rad(angle))]
-            )
[…]
