+++ proteusPy.egg-info/requires.txt
@@ -1,32 +1,0 @@
-numpy
-matplotlib
-pandas<2.0
-pyvista[all]
-traitlets==5.9.0
-notebook==6.4.12
-jupyter
-jupyterlab>=3
-seaborn
-pillow
-tqdm
-plotly
-datetime
-jupyter_bokeh
-openai
-panel
-scikit-learn
-gdown
-ipykernel
-ipygany
-nodejs
-pytube
-grpcio
-pip
-wget
-vtk
-kaleido
-plotly_express
-trame-jupyter-extension
-jupyter_contrib_nbextensions
-ipywidgets
-imageio[ffmpeg]
