+++ proteusPy.egg-info/SOURCES.txt
@@ -1,31 +1,0 @@
-LICENSE
-README.md
-setup.py
-proteusPy/Disulfide.py
-proteusPy/DisulfideClass_Constructor.py
-proteusPy/DisulfideClasses.py
-proteusPy/DisulfideExceptions.py
-proteusPy/DisulfideList.py
-proteusPy/DisulfideLoader.py
-proteusPy/ProteusGlobals.py
-proteusPy/ProteusPyWarning.py
-proteusPy/Residue.py
-proteusPy/__init__.py
-proteusPy/angle_annotation.py
-proteusPy/atoms.py
-proteusPy/proteusPyWarning.py
-proteusPy/residue.py
-proteusPy/turtle3D.py
-proteusPy/utility.py
-proteusPy.egg-info/PKG-INFO
-proteusPy.egg-info/SOURCES.txt
-proteusPy.egg-info/dependency_links.txt
-proteusPy.egg-info/not-zip-safe
-proteusPy.egg-info/requires.txt
-proteusPy.egg-info/top_level.txt
-proteusPy/data/SS_consensus_class_sext.pkl
-proteusPy/data/__init__.py
-proteusPy/data/ss_completed.txt
-proteusPy/data/ss_ids.txt
-proteusPy/data/ss_query.json
-tests/test_turtle3d.py
