### 2024-03-12 876960a
v0.92.27, moving back to setup.py
