+++ proteusPy/turtle3D.py
@@ -225,9 +225,7 @@ class Turtle3D:
         elif y is None and z is None:
             # Array, list, tuple...
             if len(x) != 3:
-                raise ValueError(
-                    "Turtle3D: x is not a list/tuple/array of 3 numbers"
-                )
+                raise ValueError("Turtle3D: x is not a list/tuple/array of 3 numbers")
             self._heading = numpy.array(x, "d")
         else:
             # Three numbers
@@ -266,9 +264,7 @@ class Turtle3D:
         elif y is None and z is None:
             # Array, list, tuple...
             if len(x) != 3:
-                raise ValueError(
-                    "Turtle3D: x is not a list/tuple/array of 3 numbers"
-                )
+                raise ValueError("Turtle3D: x is not a list/tuple/array of 3 numbers")
             self._left = numpy.array(x, "d")
         else:
             # Three numbers
@@ -307,9 +303,7 @@ class Turtle3D:
         elif y is None and z is None:
             # Array, list, tuple...
             if len(x) != 3:
-                raise ValueError(
-                    "Turtle3D: x is not a list/tuple/array of 3 numbers"
-                )
+                raise ValueError("Turtle3D: x is not a list/tuple/array of 3 numbers")
             self._up = numpy.array(x, "d")
         else:
             # Three numbers
@@ -455,7 +449,7 @@ class Turtle3D:
         self._up = self.unit(self._up)
         self.u = Vector(self._up)
 
-    def unit(self, v):
+    def unit(self, v) -> Vector:
         """
         Return a unit vector for the input vector.
 
