### 2024-03-12 1419a1a
v0.92.27, moving back to setup.py
