+++ proteusPy/Disulfide.py
@@ -238,9 +238,7 @@ class Disulfide:
         self.chi3 = _ANG_INIT
         self.chi4 = _ANG_INIT
         self.chi5 = _ANG_INIT
-        self.rho = (
-            _ANG_INIT  # new dihedral angle: Nprox - Ca_prox - Ca_dist - N_dist
-        )
+        self.rho = _ANG_INIT  # new dihedral angle: Nprox - Ca_prox - Ca_dist - N_dist
 
         self.torsion_length = _FLOAT_INIT
 
@@ -1006,7 +1004,7 @@ class Disulfide:
         self.chi4 = dihedrals[3]
         self.chi5 = dihedrals[4]
 
-    def bounding_box(self):
+    def bounding_box(self) -> np.array:
         """
         Return the bounding box array for the given disulfide
 
@@ -1027,7 +1025,7 @@ class Disulfide:
 
         return res
 
-    def build_yourself(self):
+    def build_yourself(self) -> None:
         """
         Build a model Disulfide based its internal dihedral state
         Routine assumes turtle is in orientation #1 (at Ca, headed toward
@@ -1044,7 +1042,7 @@ class Disulfide:
 
     def build_model(
         self, chi1: float, chi2: float, chi3: float, chi4: float, chi5: float
-    ):
+    ) -> None:
         """
         Build a model Disulfide based on the input dihedral angles.
         Routine assumes turtle is in orientation #1 (at Ca, headed toward
@@ -1172,7 +1170,7 @@ class Disulfide:
         _max = ic[:, idx].max()
         return _min, _max
 
-    def compute_local_coords(self):
+    def compute_local_coords(self) -> None:
         """
         Compute the internal coordinates for a properly initialized Disulfide Object.
 
@@ -1246,14 +1244,12 @@ class Disulfide:
 
         energy = 2.0 * (cos(torad(3.0 * chi1)) + cos(torad(3.0 * chi5)))
         energy += cos(torad(3.0 * chi2)) + cos(torad(3.0 * chi4))
-        energy += (
-            3.5 * cos(torad(2.0 * chi3)) + 0.6 * cos(torad(3.0 * chi3)) + 10.1
-        )
+        energy += 3.5 * cos(torad(2.0 * chi3)) + 0.6 * cos(torad(3.0 * chi3)) + 10.1
 
         self.energy = energy
         return energy
 
-    def display(self, single=True, style="sb", light=True, shadows=False):
+    def display(self, single=True, style="sb", light=True, shadows=False) -> None:
         """
         Display the Disulfide bond in the specific rendering style.
 
@@ -1463,7 +1459,7 @@ class Disulfide:
                 pl.enable_shadows()
         return pl
 
-    def Distance_neighbors(self, others: DisulfideList, cutoff: float):
+    def Distance_neighbors(self, others: DisulfideList, cutoff: float) -> DisulfideList:
         """
         Return list of Disulfides whose RMS atomic distance is within
         the cutoff (Å) in the others list.
@@ -1514,7 +1510,7 @@ class Disulfide:
         """
         return self.PERMISIVE
 
-    def get_full_id(self):
+    def get_full_id(self) -> tuple:
         """
         Return the Disulfide full IDs (Used with BIO.PDB)
 
@@ -1524,7 +1520,7 @@ class Disulfide:
 
     def initialize_disulfide_from_chain(
         self, chain1, chain2, proximal, distal, resolution, quiet=True
-    ):
+    ) -> None:
         """
         Initialize a new Disulfide object with atomic coordinates from
         the proximal and distal coordinates, typically taken from a PDB file.
@@ -1550,10 +1546,7 @@ class Disulfide:
         prox_residue = chain1[prox]
         dist_residue = chain2[dist]
 
-        if (
-            prox_residue.get_resname() != "CYS"
[…]
