+++ proteusPy/__init__.py
@@ -9,14 +9,14 @@
 from pathlib import Path
 
 __pdoc__ = {"__all__": False}
-__version__ = "0.92.25"
+# __version__ = "0.92.25"
 
-"""
+
 version_file = Path(__file__).parent.parent / "VERSION"
 if version_file.is_file():
     with open(version_file) as f:
         __version__ = f.read().strip()
-"""
+
 
 import copy
 import datetime
