### 2024-03-13 87d5a76
docs and updating install instructions
