### 2024-03-13 1fcc52a
docs and updating install instructions
