+++ makefile
@@ -4,10 +4,11 @@
 
 VERS := $(shell grep ^0 VERSION | cut -d= -f2 | tr -d \" | sed 's/^[[:space:]]*//')
 
-PYPI_PASSWORD := $(shell echo $$PYPI_PASSWORD)
+#PYPI_PASSWORD := $(shell echo $$PYPI_PASSWORD)
 CONDA = mamba
 
 MESS = "0.92.27"
+# VERS = "0.92.27"
 
 DEVNAME = ppydev
 OUTFILES = sdist.out, bdist.out, docs.out tag.out
@@ -84,7 +85,7 @@ sdist: .
 	@echo $(VERS) > sdist.out
 
 docs: .
-	pdoc -o docs --math --logo "./assets/logo.png" ./proteusPy
+	pdoc -o docs --math --logo "./logo.png" ./proteusPy
 	@echo $(VERS) > docs.out
 
 # normally i push to PyPi via github action
