### 2024-03-16 b72b48e
jupyter cleanup
