+++ proteusPy/__init__.py
@@ -9,7 +9,7 @@
 from pathlib import Path
 
 __pdoc__ = {"__all__": False}
-# __version__ = "0.92.28"
+# __version__ = "0.92.29"
 
 _version_file = Path(__file__).parent.parent / "VERSION"
 if _version_file.is_file():
