+++ proteusPy/DisulfideList.py
@@ -111,12 +111,10 @@ class DisulfideList(UserList):
 
     Get the first disulfide via indexing.
     >>> SS = PDB_SS[0]
-    >>> SS
-    <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>
+    >>> assert str(SS) == "<Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>"
 
     >>> SS4yys = PDB_SS['4yys']
-    >>> SS4yys
-    [<Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_56A_98A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_156A_207A, Source: 4yys, Resolution: 1.35 Å>]
+    >>> assert str(SS4yys) == "[<Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_56A_98A, Source: 4yys, Resolution: 1.35 Å>, <Disulfide 4yys_156A_207A, Source: 4yys, Resolution: 1.35 Å>]"
 
     Make some empty disulfides.
     >>> ss1 = Disulfide('ss1')
@@ -128,49 +126,8 @@ class DisulfideList(UserList):
 
     Extract the first disulfide
     >>> ss1 = PDB_SS[0]
-    >>> ss1.pprint_all()
-    <Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å
-     Proximal Chain fullID: <('4yys', 0, 'A', (' ', 22, ' '))> Distal Chain fullID: <('4yys', 0, 'A', (' ', 65, ' '))>
-    Proximal Coordinates:
-       N: <Vector -2.36, -20.48, 5.21>
-       Cα: <Vector -2.10, -19.89, 3.90>
-       C: <Vector -1.12, -18.78, 4.12>
-       O: <Vector -1.30, -17.96, 5.03>
-       Cβ: <Vector -3.38, -19.31, 3.32>
-       Sγ: <Vector -3.24, -18.40, 1.76>
-       Cprev <Vector -2.67, -21.75, 5.36>
-       Nnext: <Vector -0.02, -18.76, 3.36>
-     Distal Coordinates:
-       N: <Vector -0.60, -18.71, -1.62>
-       Cα: <Vector -0.48, -19.10, -0.22>
-       C: <Vector 0.92, -19.52, 0.18>
-       O: <Vector 1.10, -20.09, 1.25>
-       Cβ: <Vector -1.48, -20.23, 0.08>
-       Sγ: <Vector -3.22, -19.69, 0.18>
-       Cprev <Vector -0.73, -17.44, -2.01>
-       Nnext: <Vector 1.92, -19.18, -0.63>
-    <BLANKLINE>
-     Proximal Internal Coords:
-       N: <Vector -0.41, 1.40, -0.00>
-       Cα: <Vector 0.00, 0.00, 0.00>
-       C: <Vector 1.50, 0.00, 0.00>
-       O: <Vector 2.12, 0.71, -0.80>
-       Cβ: <Vector -0.50, -0.70, -1.25>
-       Sγ: <Vector 0.04, -2.41, -1.50>
-       Cprev <Vector -2.67, -21.75, 5.36>
-       Nnext: <Vector -0.02, -18.76, 3.36>
-    Distal Internal Coords:
-       N: <Vector 1.04, -5.63, 1.17>
[…]
