### 2024-03-16 e8623df
jupyter cleanup
