### 2024-03-16 1b1afb9
moved to pandas 2.2 by rebuilding .pkl files, 0.93
