### 2024-03-16 61f2600
moved to pandas 2.2 by rebuilding .pkl files, 0.93
