+++ proteusPy/DisulfideList.py
@@ -1063,7 +1063,7 @@ class DisulfideList(UserList):
 
 
 # utility functions
-
+from proteusPy.DisulfideExceptions import DisulfideConstructionWarning
 
 def load_disulfides_from_id(
     struct_name: str,
