### 2024-03-17 022ea2c
added .travis config, readme update
