+++ .travis.yml
@@ -1,0 +1,7 @@
+language: python
+python:
+    - "3.11"
+install:
+    - pip install .
+script:
+    - pytest
