+++ .travis.yml
@@ -1,5 +1,6 @@
 language: python
 python:
+    - "3.10"
     - "3.11"
 install:
     - pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
