+++ .travis.yml
@@ -1,7 +1,9 @@
-language: python
-python:
-    - "3.10"
-    - "3.11"
+language: generic
+dist: xenial
+before_install:
+  - sudo apt-get update
+  - sudo apt-get install -y python3.11.7
+
 install:
     - pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
     - pip install pytest
