+++ .travis.yml
@@ -1,8 +1,6 @@
-language: generic
-dist: xenial
-before_install:
-  - sudo apt-get update
-  - sudo apt-get install -y python3.11.7
+language: python
+python:
+  - "3.11-dev
 
 install:
     - pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
