+++ .travis.yml
@@ -1,10 +1,38 @@
-language: python
-python:
-  - "3.11-dev"
 
+
+#  - arm64  # No Python images yet on xenial
+
+matrix:
+  fast_finish: true
+  include:
+    - python:  'pypy3'
+    - python:  '3.11'
+    - python:  '3.11'
+      dist:  focal
+    - python:  '3.11'
+      dist:  bionic
+    - python:  '3.10'
+    - python:  '3.10'
+      dist:  focal
+    - python:  '3.10'
+      dist:  bionic
+    - python:  '3.9'
+    - python:  '3.9'
+      dist:  bionic
+    - python:  '3.8'
+    - python:  '3.7'
+    - python:  'nightly'
+    - python:  '3.11-dev'
+    - python:  '3.10-dev'
+    - python:  '3.9-dev'
+    - python:  '3.8-dev'
+    - python:  '3.7-dev'
+
+python:
+  - "3.11"
 install:
     - pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
-    - pip install pytest
+    - pip install pytest tqdm
     - pip install .
 script:
     - pytest
