+++ .travis.yml
@@ -1,3 +1,9 @@
+language: python
+dist: bionic
+arch:
+  - amd64
+python:
+  - "3.11"
 
 
 #  - arm64  # No Python images yet on xenial
@@ -16,23 +22,14 @@ matrix:
       dist:  focal
     - python:  '3.10'
       dist:  bionic
-    - python:  '3.9'
-    - python:  '3.9'
-      dist:  bionic
-    - python:  '3.8'
-    - python:  '3.7'
     - python:  'nightly'
     - python:  '3.11-dev'
     - python:  '3.10-dev'
-    - python:  '3.9-dev'
-    - python:  '3.8-dev'
-    - python:  '3.7-dev'
-
-python:
-  - "3.11"
+    
 install:
     - pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
     - pip install pytest tqdm
     - pip install .
+
 script:
     - pytest
