### 2024-03-17 94fe84d
added gh action for pytest
