### 2024-03-17 2d3217e
added gh action for pytest
