+++ .github/workflows/pytest.yml
@@ -1,0 +1,26 @@
+name: proteusPy
+
+on: [push, pull_request]
+
+jobs:
+    build:
+      runs-on: ubuntu-latest
+      steps:
+        - name: Checkout repository
+          uses: actions/checkout@v2
+        - name: Set up Python 3.11
+          uses: actions/setup-python@v3
+          with: 
+            python-version: 3.11
+
+        - name: Install dependencies
+          run: |
+              python -m pip install --upgrade pip
+              pip install pytest
+              pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
+              pip install .
+
+        - name: Test with pytest
+          run: |
+            pytestname: proteusPy
+            pytest
