+++ .github/workflows/pytest.yml
@@ -1,4 +1,4 @@
-name: proteusPy Testing
+name: Testing Testing
 
 on: [push, pull_request]
 
