+++ .github/workflows/pytest.yml
@@ -1,4 +1,4 @@
-name: Testing Testing
+name: Testing
 
 on: [push, pull_request]
 
