+++ .pdocconfig.py
@@ -1,1 +1,1 @@
-pdoc_config = {"exclude": ["proteusPy/__version__.py"]}
+pdoc_config = {"exclude": ["proteusPy/version.py"]}
