+++ get_version.py
@@ -1,0 +1,4 @@
+# get_version.py
+with open('proteusPy/version.py') as f:
+    exec(f.read())
+print(__version__)
