+++ proteusPy/ProteusGlobals.py
@@ -16,8 +16,7 @@ PDB_DIR = os.getenv("PDB")
 if PDB_DIR is None:
     PDB_DIR = "."
 
-REPO_DATA_DIR = _this_dir / "data"
-
+REPO_DATA_DIR = os.path.join(_this_dir, "data/")
 MODEL_DIR = os.path.join(PDB_DIR, "good/")
 
 WINFRAME = 512  # single panel width
