+++ proteusPy/data/__init__.py
@@ -15,7 +15,7 @@ DATA_DIR = f"{_abspath}/"
 SS_PICKLE_FILE = "PDB_all_ss.pkl"
 SS_DICT_PICKLE_FILE = "PDB_all_ss_dict.pkl"
 SS_DICT_PICKLE_FILE2 = "PDB_all_ss_dict_ind.pkl"
-SS_TORSIONS_FILE = "PDB_all_SS_torsions.csv"
+SS_TORSIONS_FILE = "PDB_all_ss_torsions.csv"
 PROBLEM_ID_FILE = "PDB_all_SS_problems.csv"
 SS_ID_FILE = "ss_ids.txt"
 
