+++ proteusPy/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.93.5"
+__version__ = "0.93.6"
