### 2024-03-18 e75ced4
utility.py cleanup, docs
