### 2024-03-18 b7869e6
docs, utility.py cleanup
