+++ proteusPy/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.93.7"
+__version__ = "0.93.8"
