+++ proteusPy/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.93.9"
+__version__ = "0.94.0"
