+++ tests/test_disulfide.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 from Bio.PDB import PDBList
 
-from proteusPy.Disulfide import Disulfide
+from proteusPy import Disulfide, Disulfide_Energy_Function, check_header_from_file
 
 
 class TestDisulfide(unittest.TestCase):
@@ -55,7 +55,6 @@ class TestDisulfide(unittest.TestCase):
         assert_allclose(result, expected_result, rtol=1e-05, atol=1e-08)
 
     def test_energy(self):
-        from proteusPy.Disulfide import Disulfide_Energy_Function
 
         dihedrals = [-60.0, -60.0, -90.0, -60.0, -90.0]
         result = Disulfide_Energy_Function(dihedrals)
@@ -66,7 +65,7 @@ class TestDisulfide(unittest.TestCase):
     def test_minimize(self):
         from scipy.optimize import minimize
 
-        from proteusPy.Disulfide import Disulfide_Energy_Function
+        from proteusPy import Disulfide_Energy_Function
 
         # initial guess for chi1, chi2, chi3, chi4, chi5
         initial_guess = [
@@ -87,25 +86,23 @@ class TestDisulfide(unittest.TestCase):
     def test_header(self):
         import tempfile
 
-        from proteusPy.Disulfide import check_header_from_file
-
         temp_dir = tempfile.TemporaryDirectory()
         pdb_home = f"{temp_dir.name}/"
         entry = "5rsa"
         pdblist = PDBList(pdb=pdb_home, verbose=False)
         ok = False
-        if not pdblist.retrieve_pdb_file(entry, file_format="pdb", pdir=pdb_home):
-            ok = False
-        else:
+        if pdblist.retrieve_pdb_file(entry, file_format="pdb", pdir=pdb_home):
             filename = f"{pdb_home}pdb5rsa.ent"
             ok = check_header_from_file(filename)
+        else:
+            ok = False
 
         self.assertTrue(ok)
 
     def test_load(self):
         import tempfile
 
-        from proteusPy.DisulfideList import DisulfideList, load_disulfides_from_id
+        from proteusPy import DisulfideList, load_disulfides_from_id
 
         temp_dir = tempfile.TemporaryDirectory()
         pdb_home = f"{temp_dir.name}/"
@@ -119,23 +116,10 @@ class TestDisulfide(unittest.TestCase):
             self.assertTrue(len(sslist) > 0)
 
     def test_compare(self):
-        import tempfile
-
-        from proteusPy.DisulfideList import DisulfideList, load_disulfides_from_id
-
         diff = 1.0
 
-        temp_dir = tempfile.TemporaryDirectory()
-        pdb_home = f"{temp_dir.name}/"
-        entry = "5rsa"
-        pdblist = PDBList(pdb=pdb_home, verbose=False)
-        if not pdblist.retrieve_pdb_file(entry, file_format="pdb", pdir=pdb_home):
-            return False
-        else:
-            sslist = DisulfideList([], "tst")
-            sslist = load_disulfides_from_id("5rsa", pdb_dir=pdb_home)
-            ss1 = sslist[0]
-            diff = ss1.Torsion_RMS(ss1)
+        ss1 = self.sslist[0]
+        diff = ss1.Torsion_RMS(ss1)
 
         self.assertTrue(diff == 0)
 
