+++ makefile
@@ -59,12 +59,11 @@ devclean: .
 
 	
 # activate the package before running!
-install:
+install: 
 	@echo "Starting installation step 2/2 for $(VERS)..."
 	@echo "Installing VTK..."
-	$(CONDA) install -y vtk
 	@echo "Installing proteusPy..."
-	@pip install -U -q . 
+	pip install dist/proteusPy-$(VERS).tar.gz
 	@echo "Installing Biopython..."
 	@pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
 	@echo "Installing jupyter..."
