### 2024-03-31 #6169 @suchanek
Thank you. Eric G. Suchanek, PhD Astronomer at largeMonterey Institute for Research in Astronomy200 Eighth Street, Marina CA 93933(831-883-1000) Fax: 883-1031Dedicated to Research and Education in Astronomy for 50 Years.On Mar 31, 2024, at 4:11 AM, Kevin Mattheus Moerman ***@***.***> wrote:﻿
@richardjgowers please try to check in here about once a week if you can to spur on this review, thanks!
