### 2024-04-02 suchanek/proteusPy#9 @AnjaConev
Paper pdf not rendering and bibliography empty
Hi, great work and thank you for addressing all my comments so far!

Here is my final issue for this review -  two things:

1. I am not able to open the paper pdf in `proteusPy/paper
/paper.pdf` - I get an error that the pdf is invalid (both in Firefox, Chrome, on GitHub).

2. When I checked out `proteusPy/paper/paper.md` I noticed that the bibliography section is empty - it might be just a randering issue though.
