### 2024-04-02 #6169 @AnjaConev
Thanks @Kevin-Mattheus-Moerman for the reminder!

I just opened my final issue for @richardjgowers related to the references and the paper pdf https://github.com/suchanek/proteusPy/issues/9

That should be all on my side!
