### 2024-04-02 #6169 @suchanek
All,I’ve had trouble getting the artifact (the paper) to drop correctly. Last workflow ran correctly. If you go to the actions for the repo you can grab it. Let me know if you have issues. Thank you!Eric G. Suchanek, PhD Astronomer at largeMonterey Institute for Research in Astronomy200 Eighth Street, Marina CA 93933(831-883-1000) Fax: 883-1031Dedicated to Research and Education in Astronomy for 50 Years.On Apr 2, 2024, at 12:36 PM, Anja Conev ***@***.***> wrote:﻿
Thanks @Kevin-Mattheus-Moerman for the reminder!
I just opened my final issue for @richardjgowers related to the references and the paper pdf suchanek/proteusPy#9
That should be all on my side!
