### 2024-04-04 315d14f
Update paper.md
Small formatting, cleanup
