### 2024-04-04 suchanek/proteusPy#9 @suchanek
I'm unable to directly upload the latest draft workflow run since I'm abroad, but you can get it from the latest workflow run. If you have trouble, let me know and I can email directly.

Thanks for your ongoing help and support!

Eric
