### 2024-04-07 84c2901
Update DisulfideClass_Constructor.py
Removed the pickle compat load
