+++ proteusPy/DisulfideClass_Constructor.py
@@ -118,13 +118,13 @@ class DisulfideClass_Constructor:
 
     def load_class_dict(self, fname=f"{DATA_DIR}{SS_CLASS_DICT_FILE}") -> dict:
         with open(fname, "rb") as f:
-            self.classdict = pd.compat.pickle_compat.load(f)
-            # self.classdict = pickle.load(f)
+            #self.classdict = pd.compat.pickle_compat.load(f)
+            self.classdict = pickle.load(f)
 
     def load_consensus_file(self, fname=f"{DATA_DIR}{SS_CONSENSUS_FILE}"):
         with open(fname, "rb") as f:
-            res = pd.compat.pickle_compat.load(f)
-            # res = pickle.load(f)
+            #res = pd.compat.pickle_compat.load(f)
+             res = pickle.load(f)
             return res
 
     def build_class_df(self, class_df, group_df):
