+++ proteusPy/DisulfideClass_Constructor.py
@@ -57,7 +57,7 @@ merge_cols = [
 
 # class DisulfideClass_Constructor:
 class DisulfideClass_Constructor:
-    """
+    r"""
     This Class manages structural classes for the disulfide bonds contained
     in the proteusPy disulfide database.
 
@@ -67,7 +67,7 @@ class DisulfideClass_Constructor:
     Schmidt et al. (Biochem, 2006, 45, 7429-7433), across all 32 (2^5), possible
     binary sidechain torsional combinations. Classes are named per Schmidt's convention.
     The ``class_id`` represents the sign of each dihedral angle $\chi_{1} - \chi_{1'}$
-    where *0* repreents *negative* dihedral angle and *2* a *positive* angle.
+    where *0* represents *negative* dihedral angle and *2* a *positive* angle.
 
     |   class_id | SS_Classname   | FXN        |   count |   incidence |
     |-----------:|:---------------|:-----------|--------:|------------:|
