### 2024-04-15 #6169 @suchanek
All,
I've dropped the latest draft manually into the repo under paper/paper.pdf. I am hopeful we can speed this along.

Cheers,
Eric
