+++ makefile
@@ -62,8 +62,10 @@ devclean: .
 install: 
 	@echo "Starting installation step 2/2 for $(VERS)..."
 	@echo "Installing VTK..."
+	$(CONDA) install vtk==9.2.6
 	@echo "Installing proteusPy..."
-	pip install dist/proteusPy-$(VERS).tar.gz
+	#@pip install dist/proteusPy-$(VERS).tar.gz
+	pip install .
 	@echo "Installing Biopython..."
 	@pip install git+https://github.com/suchanek/biopython.git@egs_ssbond_240305#egg=biopython
 	@echo "Installing jupyter..."
