### 2024-04-17 7a183c8
utility.py
