### 2024-04-17 9bd685e
setup cleanup for jupyterlab version
