### 2024-04-17 f95875d
setup cleanup for jupyterlab version
