+++ setup.py
@@ -81,11 +81,10 @@ setup(
         "numpy",
         "matplotlib",
         "pandas==2.2.1",
-        "pyvista[all]",
+        "pyvista[all,trame]",
         "traitlets==5.9.0",
         "jupyter",
-        "jupyter_server<3.0",
-        "jupyterlab<4.0",
+        "jupyter_server<2.0",
         "seaborn",
         "pillow",
         "tqdm",
