+++ proteusPy/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.95.0"
+__version__ = "0.95.2"
