+++ old/ppy_0.95.0.txt
@@ -1,0 +1,190 @@
+aiofiles==22.1.0
+aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1710511613725/work
+aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
+aiosqlite==0.20.0
+annotated-types==0.6.0
+anyio==3.7.1
+appnope==0.1.4
+argon2-cffi==23.1.0
+argon2-cffi-bindings==21.2.0
+arrow==1.3.0
+asttokens==2.4.1
+attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
+Babel==2.14.0
+beautifulsoup4==4.12.3
+biopython @ git+https://github.com/suchanek/biopython.git@481cd69ca7775b2598e64bd7c9ee7d3d0770ef76
+bleach==6.1.0
+bokeh==3.4.1
+certifi==2024.2.2
+cffi==1.16.0
+charset-normalizer==3.3.2
+cmocean==4.0.3
+colorcet==3.1.0
+comm==0.2.2
+contourpy==1.2.1
+cycler==0.12.1
+DateTime==5.5
+debugpy==1.8.1
+decorator==5.1.1
+defusedxml==0.7.1
+distro==1.9.0
+entrypoints==0.4
+executing==2.0.1
+fastjsonschema==2.19.1
+filelock==3.13.4
+fonttools==4.51.0
+fqdn==1.5.1
+frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1702645562307/work
+gdown==5.1.0
+grpcio==1.62.1
+h11==0.14.0
+httpcore==1.0.5
+httpx==0.27.0
+idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
+imageio==2.34.0
+imageio-ffmpeg==0.4.9
+ipygany==0.5.0
+ipykernel==6.29.4
+ipython==8.21.0
+ipython-genutils==0.2.0
+ipywidgets==8.1.2
+isoduration==20.11.0
+jedi==0.19.1
+Jinja2==3.1.3
+joblib==1.4.0
+json5==0.9.25
+jsonpointer==2.4
+jsonschema==4.21.1
+jsonschema-specifications==2023.12.1
+jupyter==1.0.0
+jupyter-console==6.6.3
+jupyter-events==0.10.0
+jupyter-highlight-selected-word==0.2.0
+jupyter-nbextensions-configurator==0.6.3
+jupyter-server==1.24.0
+jupyter-ydoc==0.2.5
+jupyter_bokeh==4.0.1
+jupyter_client==7.4.9
+jupyter_contrib_core==0.4.2
+jupyter_contrib_nbextensions==0.7.0
+jupyter_core==5.7.2
+jupyter_server_fileid==0.9.1
+jupyter_server_proxy==4.1.2
+jupyter_server_ydoc==0.8.0
+jupyterlab==3.6.7
+jupyterlab_pygments==0.3.0
+jupyterlab_server==2.26.0
+jupyterlab_widgets==3.0.10
+kaleido==0.2.1
+kiwisolver==1.4.5
+linkify-it-py==2.0.3
+loguru @ file:///Users/runner/miniforge3/conda-bld/loguru_1695547409960/work
+lxml==5.2.1
+Markdown==3.6
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.8.4
+matplotlib-inline==0.1.7
+mdit-py-plugins==0.4.0
+mdurl==0.1.2
+meshio==5.3.5
+mistune==3.0.2
[…]
