+++ paper/paper.md
@@ -119,7 +119,7 @@ The first time one loads the database via [Load_PDB_SS()](https://suchanek.githu
 After installation is complete, launch jupyter lab:
 
 ```console
-$ jupyter lab 
+$ jupyter notebook
 ```
 and open [Analysis_2q7q](https://github.com/suchanek/proteusPy/blob/master/notebooks/Analysis_2q7q.ipynb). This notebook analyzes the disulfide bond with the lowest energy in the entire database and performs some searches in dihedral angle space to find similar conformations. There are several other notebooks in this directory that illustrate using the program. Some of these reflect active development work so may not be 'fully baked'.
 
