### 2024-04-18 fdd4484
updates to quickstart
