### 2024-04-18 b7842a8
updates to quickstart
