### 2024-04-21 d4d59c6
working with versioning on jupyter now
