+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "proteusPy"
-version = "0.95.2"
+version = "0.95.3"
 description = "proteusPy - Protein Structure Analysis and Modeling Tools"
 authors = ["Eric Suchanek <suchanek@mac.com>"]
 license = "LICENSE"
@@ -67,6 +67,7 @@ pandas = "^2.2.1"
 trame = "^3.6.0"
 trame-vtk = "^2.8.5"
 trame-vuetify = "^2.4.3"
+toml = "^0.10.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^24.2.0"
@@ -87,11 +88,10 @@ _install = "poetry install"
 _jup_install = "poetry poe jupyter"
 _jup_install_dev = "poetry poe jupyter_dev"
 _clean_sdist.script = "shutil:rmtree('sdist', ignore_errors=1)"
-_clean_docs.script = "shutil:rmtree('docs', ignore_errors=1)"
 _jup1 = "jupyter contrib nbextension install --sys-prefix"
 _jup2 = "jupyter nbextension enable --py --sys-prefix widgetsnbextension"
-_jup3 = "python -m ipykernel install --user --name proteusPy --display-name 'proteusPy'"
-_jup4 = "python -m ipykernel install --user --name proteusPy --display-name 'ppydev $version'"
+_jup3 = "python programs/run_jupyter.py proteusPy"
+_jup4 = "python programs/run_jupyter.py ppydev"
 
 build = ["_build"]
 install = [ "_install", "_jup_install"]
@@ -115,7 +115,7 @@ test = ["_test"]
   [tool.poe.tasks.docs]
   help = "Build the docs with pdoc"
   cmd = "pdoc -o docs --math --logo ./logo.png ./proteusPy"
-  deps = ["_clean_docs"]
+  deps = []
 
 [build-system]
 requires = ["poetry-core"]
