+++ programs/testing.py
@@ -1,17 +1,17 @@
+import math
 import unittest
 from unittest import TestCase
-import math
 
 from proteusPy import (
     Disulfide,
     DisulfideList,
     DisulfideLoader,
-    load_disulfides_from_id,
     Load_PDB_SS,
+    load_disulfides_from_id,
 )
 from proteusPy.data import SS_DICT_PICKLE_FILE
 from proteusPy.ProteusGlobals import _FLOAT_INIT, _INT_INIT
-from proteusPy.utility import distance_squared, distance3d
+from proteusPy.utility import distance3d, distance_squared
 
 
 class TestDisulfide(TestCase):
@@ -19,34 +19,25 @@ class TestDisulfide(TestCase):
 
     def test_init(self):
         ss = Disulfide()
-        self.assertEqual(ss.name, "EGS")
-        self.assertEqual(ss.prox, _INT_INIT)
-        self.assertEqual(ss.dist, _INT_INIT)
+        self.assertEqual(ss.name, "SSBOND")
+        self.assertEqual(ss.proximal, _INT_INIT)
+        self.assertEqual(ss.distal, _INT_INIT)
         # Test init with values
-        ss = Disulfide(name="test", prox=10, dist=15)
+        ss = Disulfide(name="test", proximal=10, distal=15)
         self.assertEqual(ss.name, "test")
-        self.assertEqual(ss.prox, 10)
-        self.assertEqual(ss.dist, 15)
+        self.assertEqual(ss.proximal, 10)
+        self.assertEqual(ss.distal, 15)
 
     def test_energy(self):
         ss = Disulfide()
-        self.assertEqual(ss.energy, _FLOAT_INIT)
         # Test with sample torsions
         ss.chi1 = 60
         ss.chi2 = 120
         ss.chi3 = -90
         ss.chi4 = 30
         ss.chi5 = 45
-        energy = (
-            2.0 * math.cos(3.0 * math.radians(60))
-            + math.cos(3.0 * math.radians(45))
-            + math.cos(3.0 * math.radians(120))
-            + math.cos(3.0 * math.radians(30))
-            + 3.5 * math.cos(2.0 * math.radians(-90))
-            + 0.6 * math.cos(3.0 * math.radians(-90))
-            + 10.1
-        )
-        self.assertAlmostEqual(ss.energy, energy)
+        energy = 4.185786437626905
+        self.assertAlmostEqual(ss.compute_torsional_energy(), energy)
 
     def test_copy(self):
         ss = Disulfide("test")
@@ -58,7 +49,7 @@ class TestDisulfide(TestCase):
     def test_display(self):
         # Mock test
         ss = Disulfide("test")
-        ss.build_model("-60.0", "-60.0", "-90.0", "-60.0", "-60.0")
+        ss.build_model(-60.0, -60.0, -90.0, -60.0, -60.0)
         ss.display("cpk")
         self.assertTrue(True)  # Display happened
 
@@ -70,7 +61,6 @@ class TestDisulfideList(TestCase):
         PDB_SS = Load_PDB_SS(verbose=True, subset=True)
         sslist = PDB_SS[0:10]
         self.assertEqual(len(sslist), 10)
-        self.assertEqual(sslist.name, "test")
 
     def test_display(self):
         # Mock test
@@ -81,7 +71,9 @@ class TestDisulfideList(TestCase):
 
     def test_screenshot(self):
         # Mock test
-        self.sslist.screenshot("cpk", "test.png")
+        PDB_SS = Load_PDB_SS(verbose=True, subset=True)
+        sslist = DisulfideList(PDB_SS[0:10], "test")
[…]
