+++ .pdocconfig.py
@@ -1,1 +1,1 @@
-pdoc_config = {"exclude": ["proteusPy/version.py"]}
+pdoc_config = {"exclude": ["proteusPy/version.py", "proteusPy/run_jupyter.py"]}
