+++ pyproject.toml
@@ -54,7 +54,7 @@ pytube = "^15.0.0"
 grpcio = "^1.62.1"
 pip = "^24.0"
 wget = "^3.2"
-vtk = "9.2.6"
+vtk = "^9.2.6"
 kaleido = "0.2.1"
 plotly-express = "^0.4.1"
 trame-jupyter-extension = "^2.0.1"
