+++ proteusPy/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.95.2"
+__version__ = "0.95.5"
