### 2024-05-02 #6169 @suchanek
Is there some way we can move forward? I feel like we are stalled. I'd like to get this published so I can publish the analysis.

Thank you!
Eric Suchanek
