### 2024-05-03 #6169 @richardjgowers
@suchanek there's a few post-review tasks above, such as finalising the release and creating the (zenodo) DOI which we will need to finish this review too
