+++ paper/paper.bib
@@ -1,23 +1,0 @@
-@article{Kontoyianni_2008,	doi = {10.2174/092986708783330566},	url = {https://doi.org/10.2174%2F092986708783330566},	year = 2008,	month = {jan},	publisher = {Bentham Science Publishers Ltd.},	volume = {15},	number = {2},	pages = {107--116},	author = {Maria Kontoyianni and Prakash Madhav and Eric Suchanek and William Seibel},	title = {Theoretical and Practical Considerations in Virtual Screening: A Beaten Field?},	journal = {Current Medicinal Chemistry}},
-@article{Godley_1992,	doi = {10.1016/0092-8674(92)90140-8},	url = {https://doi.org/10.1016%2F0092-8674%2892%2990140-8},	year = 1992,	month = {feb},	publisher = {Elsevier {BV}},	volume = {68},	number = {4},	pages = {635--645},	author = {L. Godley and J. Pfeifer and D. Steinhauer and B. Ely and G. Shaw and R. Kaufmann and E. Suchanek and C. Pabo and J.J. Skehel and D.C. Wiley and S. Wharton},	title = {Introduction of intersubunit disulfide bonds in the membrane-distal region of the influenza hemagglutinin abolishes membrane fusion activity},	journal = {Cell}},
-@article{Sauer_1986,	doi = {10.1021/bi00368a024},	url = {https://doi.org/10.1021%2Fbi00368a024},	year = 1986,	month = {oct},	publisher = {American Chemical Society ({ACS})},	volume = {25},	number = {20},	pages = {5992--5998},	author = {R. T. Sauer and K. Hehir and R. S. Stearman and M. A. Weiss and A. Jeitler-Nilsson and E. G. Suchanek and C. O. Pabo},	title = {An engineered intersubunit disulfide enhances the stability and {DNA} binding of the N-terminal domain of .lambda. repressor},	journal = {Biochemistry}},
-@article{Pabo_1986,	doi = {10.1021/bi00368a023},	url = {https://doi.org/10.1021%2Fbi00368a023},	year = 1986,	month = {oct},	publisher = {American Chemical Society ({ACS})},	volume = {25},	number = {20},	pages = {5987--5991},	author = {C. O. Pabo and E. G. Suchanek},	title = {Computer-aided model-building strategies for protein design},	journal = {Biochemistry}}
-@article{Hogg_2020,	doi = {10.1021/bi00368a023},	url = {https://doi.org/10.3390/molecules25235729},	year = 2020,	month = {dec},	publisher = {Molecules},	volume = {25},	number = {23},	pages = {5729--5734},	author = {Philip J Hogg},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Molecules}}
-@article{Schmidt_2006,	doi = {10.1021/bi0603064},	url = {https://doi.org/10.1021/bi0603064},	year = 2006,	month = {jun},	publisher = {American Chemical Society ({ACS})},	volume = {45},	number = {24},	pages = {7429–74334},	author = {Brian Schmidt},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Biochemistry}}
[…]
