+++ paper/joss.bib
@@ -4,27 +4,13 @@
 @article{Pabo_1986,	doi = {10.1021/bi00368a023},	url = {https://doi.org/10.1021%2Fbi00368a023},	year = 1986,	month = {oct},	publisher = {American Chemical Society ({ACS})},	volume = {25},	number = {20},	pages = {5987--5991},	author = {C. O. Pabo and E. G. Suchanek},	title = {Computer-aided model-building strategies for protein design},	journal = {Biochemistry}}
 @article{Hogg_2020,	doi = {10.1021/bi00368a023},	url = {https://doi.org/10.3390/molecules25235729},	year = 2020,	month = {dec},	publisher = {Molecules},	volume = {25},	number = {23},	pages = {5729--5734},	author = {Philip J Hogg},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Molecules}}
 @article{Schmidt_2006,	doi = {10.1021/bi0603064},	url = {https://doi.org/10.1021/bi0603064},	year = 2006,	month = {jun},	publisher = {American Chemical Society ({ACS})},	volume = {45},	number = {24},	pages = {7429–74334},	author = {Brian Schmidt},	title = {Multiple Disulfide-Bonded States of Native Proteins: Estimate of Number Using Probabilities of Disulfide Bond Formation},	journal = {Biochemistry}}
-@article{Hamelyrck_2003,	doi = {10.1093/bioinformatics/btg2994},	url = {https://doi.org/10.1093/bioinformatics/btg299},	year = 2003,	month = {nov},	publisher = {Bioinformatics},	volume = {19},	number = {17},	pages = {2308-2310},	author = {Thomas Hamelyrck and Bernard Manderick},	title = {PDB file parser and structure class implemented in Python},	journal = {Bioinformatics}}
+@article{Hamelyrck_2003,	doi = {10.1093/bioinformatics/btg299},	url = {https://doi.org/10.1093/bioinformatics/btg299},	year = 2003,	month = {nov},	publisher = {Bioinformatics},	volume = {19},	number = {17},	pages = {2308-2310},	author = {Thomas Hamelyrck and Bernard Manderick},	title = {PDB file parser and structure class implemented in Python},	journal = {Bioinformatics}}
 @article{Wong_2010, doi = {10.1111/j.1538-7836.2010.03894.x}, url = {http://dx.doi.org/10.1111/j.1538-7836.2010.03894.x}, year = 2010, month = {oct}, publisher = {Journal of Thrombosis and Haemostasis}, volume = {8}, number = {10}, pages = {2345}, author = {J.H. Wong and P.J. Hogg}, title = {Analysis of disulfide bonds in protein structures}, journal = {Journal of Thrombosis and Haemostasis}}
-@misc{proteusPy,
-  author = {Eric G. Suchanek},
-  title = {proteusPy: A Package for Modeling and Analyzing Proteins of Known Structure},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/suchanek/proteusPy}
-}
-@misc{API,
-  author = {Eric G. Suchanek},
-  title = {proteusPy API},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub Documents},
-  url = {https://suchanek.github.io/suchanek/proteusPy/proteusPy.html}
-}
 @book{Abelson_DiSessa_1986, 
   place={Cambridge, MA}, 
   title={Turtle Geometry: The computer as a medium for exploring mathematics},
+  doi = {10.7551/mitpress/6933.001.0001},
[…]
