+++ paper/ppy.bib
@@ -1,0 +1,16 @@
+@misc{proteusPy,
+  author = {Eric G. Suchanek},
+  title = {proteusPy: A Package for Modeling and Analyzing Proteins of Known Structure},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/suchanek/proteusPy}
+}
+@misc{API,
+  author = {Eric G. Suchanek},
+  title = {proteusPy API},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub Documents},
+  url = {https://suchanek.github.io/suchanek/proteusPy/proteusPy.html}
+}
