### 2024-05-04 #6169 @suchanek
OK i've added some verbiage, re-rendered and put paper.pdf into the repo. Current version is 0.95.5. I cleaned up the doi references as well. You'll want to officially re-render. I won't touch it now unless you see any typos.
