+++ proteusPy/DisulfideList.py
@@ -1,0 +1,1205 @@
+"""
+This module is part of the proteusPy package, a Python package for 
+the analysis and modeling of protein structures, with an emphasis on disulfide bonds.
+This work is based on the original C/C++ implementation by Eric G. Suchanek. \n
+
+The module provides the implmentation and interface for the [DisulfideList](#DisulfideList)
+object, used extensively by Disulfide class.
+
+Author: Eric G. Suchanek, PhD
+Last revision: 3/18/2024 -egs-
+"""
+
+try:
+    # Check if running in Jupyter
+    shell = get_ipython().__class__.__name__
+    if shell == "ZMQInteractiveShell":
+        from tqdm.notebook import tqdm
+    else:
+        from tqdm import tqdm
+except NameError:
+    from tqdm import tqdm
+
+from collections import UserList
+from pathlib import Path
+
+import numpy as np
+import pandas
+import plotly.express as px
+import plotly.graph_objects as go
+import pyvista as pv
+from Bio.PDB import PDBParser
+from plotly.subplots import make_subplots
+
+import proteusPy
+from proteusPy import Disulfide
+from proteusPy.atoms import *
+from proteusPy.ProteusGlobals import MODEL_DIR, PBAR_COLS, WINSIZE
+from proteusPy.utility import get_jet_colormap, grid_dimensions
+
+# Set the figure sizes and axis limits.
+DPI = 220
+WIDTH = 6.0
+HEIGHT = 6.0
+TORMIN = -179.0
+TORMAX = 180.0
+
+
+Torsion_DF_Cols = [
+    "source",
+    "ss_id",
+    "proximal",
+    "distal",
+    "chi1",
+    "chi2",
+    "chi3",
+    "chi4",
+    "chi5",
+    "energy",
+    "ca_distance",
+    "cb_distance",
+    "phi_prox",
+    "psi_prox",
+    "phi_dist",
+    "psi_dist",
+    "torsion_length",
+    "rho",
+]
+
+Distance_DF_Cols = [
+    "source",
+    "ss_id",
+    "proximal",
+    "distal",
+    "energy",
+    "ca_distance",
+    "cb_distance",
+]
+
+
+class DisulfideList(UserList):
+    """
+    The class provides a sortable list for Disulfide objects.
+    Indexing and slicing are supported, as well as typical list operations like
+    ``.insert()``, ``.append()`` and ``.extend().`` The DisulfideList object must be initialized
+    with an iterable (tuple, list) and a name. Sorting is keyed by torsional energy.
+
+    The class can also render Disulfides to a pyVista window using the
+    [display()](#DisulfideList.display) and [display_overlay()](#DisulfideList.display_overlay)methods.
+    See below for examples.\n
+
+    Examples:
+    >>> from proteusPy import Disulfide, DisulfideLoader, DisulfideList, Load_PDB_SS
+
+    Instantiate some variables. Note: the list is initialized with an iterable and a name (optional)
+
+    >>> SS = Disulfide('tmp')
+
+    The list is initialized with an iterable, a name and resolution. Name and resolution
+    are optional.
+    >>> SSlist = DisulfideList([],'ss', -1.0)
+
+    Load the database.
+    >>> PDB_SS = Load_PDB_SS(verbose=False, subset=True)
+
+    Get the first disulfide via indexing.
+    >>> SS = PDB_SS[0]
+
+    # assert str(SS) == "<Disulfide 4yys_22A_65A, Source: 4yys, Resolution: 1.35 Å>"
+
+    >>> SS4yys = PDB_SS['4yys']
+
[…]
