### 2024-05-08 #6169 @suchanek
Apologies, I’m new to this lol. https://zenodo.org/doi/10.5281/zenodo.11148440

I believe this should do it. However, if i make any additional changes it will change, correct?

Thank you for your help Richard. :)

Cheers,
Eric

> On May 8, 2024, at 7:47 AM, Richard Gowers ***@***.***> wrote:
>
>
> @suchanek <https://github.com/suchanek> can you make a zenodo (or similar) archive of the 0.95.5 release and post the DOI here please?
>
