### 2024-05-08 #6169 @suchanek
Current and last:

10.5281/zenodo.11148441

> On May 8, 2024, at 9:33 AM, Dr Diogenes ***@***.***> wrote:
>
> Apologies, I’m new to this lol. https://zenodo.org/doi/10.5281/zenodo.11148440
>
> I believe this should do it. However, if i make any additional changes it will change, correct?
>
> Thank you for your help Richard. :)
>
> Cheers,
> Eric
>
>
>> On May 8, 2024, at 7:47 AM, Richard Gowers ***@***.***> wrote:
>>
>>
>> @suchanek <https://github.com/suchanek> can you make a zenodo (or similar) archive of the 0.95.5 release and post the DOI here please?
>>
>> —
>> Reply to this email directly, view it on GitHub <https://github.com/openjournals/joss-reviews/issues/6169#issuecomment-2100396762>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AA7PYGU7PE6OUTFGD2NXP4LZBIGFFAVCNFSM6AAAAABA6SHF5WVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCMBQGM4TMNZWGI>.
>> You are receiving this because you were mentioned.
>>
>
