### 2024-05-14 #6169 @suchanek
Are we good to go Richard?Thanks!Eric G. Suchanek, PhD On May 8, 2024, at 7:47 AM, Richard Gowers ***@***.***> wrote:﻿
@suchanek can you make a zenodo (or similar) archive of the 0.95.5 release and post the DOI here please?
