### 2024-06-05 #6169 @Kevin-Mattheus-Moerman
@joaomcteixeira @hmacdope @cbouy would you be interested in reviewing this submission ("_proteusPy: A Python Package for Protein Structure and Disulfide Bond Modeling and Analysis_"), for [JOSS: the journal of open source software](https://joss.theoj.org/)? JOSS reviews focus on evaluation of [the software](https://github.com/suchanek/proteusPy) as well as [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06169/joss.06169/10.21105.joss.06169.pdf).

This here is a review issue, where a set of check-boxes guide reviewers through our GitHub based review process. As you can see :point_up:, one reviewer has completed their review already. However, we require one more review to be completed so your help would be greatly appreciated!

You can let us know here if you are interested in helping. Thanks!
