### 2024-06-05 #6169 @suchanek
Thank you for moving this along!Eric G. Suchanek, PhD On Jun 5, 2024, at 5:00 AM, Kevin Mattheus Moerman ***@***.***> wrote:﻿
@joaomcteixeira @hmacdope @cbouy would you be interested in reviewing this submission ("proteusPy: A Python Package for Protein Structure and Disulfide Bond Modeling and Analysis"), for JOSS: the journal of open source software? JOSS reviews focus on evaluation of the software as well as a short paper.
This here is a review issue, where a set of check-boxes guide reviewers through our GitHub based review process. As you can see ☝️, one reviewer has completed their review already. However, we require one more review to be completed so your help would be greatly appreciated!
You can let us know here if you are interested in helping. Thanks!
