### 2024-06-25 #6169 @Kevin-Mattheus-Moerman
@hmacdope thanks for your help! You can get started by calling: `@editorialbot generate my checklist`
