### 2024-07-01 #6169 @hmacdope
Congrats on a wonderful piece of work @suchanek! The paper and software are well written and will be impactful.

Just a queries and things I would like fixed up:

### In the paper

- While I understand what you mean with this sentence "Disulfide bonds represent the sole naturally occurring covalent bond in proteins". It is not scientifically accurate. Please fix up.
- When you say "freely accessible database" could you provide a link to a zenodo archive with said database in a public format?
- You should reference the RCSB PDB publication(s) when referring to it.

### Code changes:

Address these two issues:

- [ ] https://github.com/suchanek/proteusPy/issues/11
- [ ] https://github.com/suchanek/proteusPy/issues/10
