### 2024-07-18 #6169 @suchanek
Greetings!

- I removed the indicated sections as requested. I would like to keep the remaining ones however. They provide important details and context.

- New zenodo doi: https://doi.org/10.5281/zenodo.12773838, although the original will resolve to latest verions.
- Updated readme for badges.
- Version: 0.96.3
- Spelling correction

I think that's it!

Cheers,
Eric
