### 2024-08-06 #6169 @Kevin-Mattheus-Moerman
@suchanek Below are some additional, and probably final, steps that require your attention:

- [ ] Please edit the ZENODO archive title to fully match the JOSS paper title.
- [ ] Please edit the ZENODO archive listed license to match your software license.
