### 2024-08-06 #6169 @suchanek
i've bumped the version, created a release that zenodo grabbed, updated the docs with the new zenodo ID. Not sure how to do the archive listed license...
