### 2024-08-07 #6169 @Kevin-Mattheus-Moerman
@suchanek

I have set the archive to the above link and updated the version used here. Let me know if you are happy with that.

On the archive. As the owner/creator of that archive you can manually edit it when you are logged into ZENODO. Below is a video showing me editing the title of one of mine, and showing you where the license can be edited.

- [ ] Please edit the title to match the paper title fully, i.e. it should read: `proteusPy: A Python Package for Protein Structure and Disulfide Bond Modeling and Analysis`.
- [ ] Please edit the archive listed license to match your project license. ZENODO defaults to a Creative Commons license but here it should be your `BSD 3 License`. You can search for BSD and various flavours show up. Make sure you pick the right one.

Let me know if you have any other questions or if I can help in any way.

https://github.com/user-attachments/assets/061c8811-6dee-43f0-a7cd-4e81daac17d7
