### 2024-08-07 #6169 @Kevin-Mattheus-Moerman
@suchanek I'll proceed with acceptance but please process one more aspect for the ZENODO archive, please remove the `suchanek/proteusPy: ` from the archive title so it just says `proteusPy: A Python Package for Protein Structure and Disulfide Bond Modeling and Analysis`. That way it matches the paper title exactly.
